% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automl_automl.R
\name{AutoMLInput}
\alias{AutoMLInput}
\title{Accepts parameters that specify an S3 input for an auto ml job}
\description{
Provides a method to turn those parameters into a dictionary.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{AutoMLInput$new()}}
\item \href{#method-to_request_list}{\code{AutoMLInput$to_request_list()}}
\item \href{#method-format}{\code{AutoMLInput$format()}}
\item \href{#method-clone}{\code{AutoMLInput$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Convert an S3 Uri or a list of S3 Uri to an AutoMLInput object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AutoMLInput$new(inputs, target_attribute_name, compression = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{inputs}}{(str, list[str]): a string or a list of string that points to (a)
S3 location(s) where input data is stored.}

\item{\code{target_attribute_name}}{(str): the target attribute name for regression
or classification.}

\item{\code{compression}}{(str): if training data is compressed, the compression type.
The default value is None.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-to_request_list"></a>}}
\if{latex}{\out{\hypertarget{method-to_request_list}{}}}
\subsection{Method \code{to_request_list()}}{
Generates a request dictionary using the parameters provided to the class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AutoMLInput$to_request_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-format"></a>}}
\if{latex}{\out{\hypertarget{method-format}{}}}
\subsection{Method \code{format()}}{
format class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AutoMLInput$format()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AutoMLInput$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
