% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_monitoring.R
\name{BaseliningJob}
\alias{BaseliningJob}
\title{Baselining Job Class}
\description{
Provides functionality to retrieve baseline-specific files output from baselining job.
}
\section{Super classes}{
\code{\link[R6sagemaker:.Job]{R6sagemaker::.Job}} -> \code{\link[R6sagemaker:ProcessingJob]{R6sagemaker::ProcessingJob}} -> \code{BaseliningJob}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{BaseliningJob$new()}}
\item \href{#method-from_processing_job}{\code{BaseliningJob$from_processing_job()}}
\item \href{#method-baseline_statistics}{\code{BaseliningJob$baseline_statistics()}}
\item \href{#method-suggested_constraints}{\code{BaseliningJob$suggested_constraints()}}
\item \href{#method-print}{\code{BaseliningJob$print()}}
\item \href{#method-clone}{\code{BaseliningJob$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="R6sagemaker" data-topic="ProcessingJob" data-id="describe">}\href{../../R6sagemaker/html/ProcessingJob.html#method-describe}{\code{R6sagemaker::ProcessingJob$describe()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="R6sagemaker" data-topic="ProcessingJob" data-id="from_processing_arn">}\href{../../R6sagemaker/html/ProcessingJob.html#method-from_processing_arn}{\code{R6sagemaker::ProcessingJob$from_processing_arn()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="R6sagemaker" data-topic="ProcessingJob" data-id="from_processing_name">}\href{../../R6sagemaker/html/ProcessingJob.html#method-from_processing_name}{\code{R6sagemaker::ProcessingJob$from_processing_name()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="R6sagemaker" data-topic="ProcessingJob" data-id="start_new">}\href{../../R6sagemaker/html/ProcessingJob.html#method-start_new}{\code{R6sagemaker::ProcessingJob$start_new()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="R6sagemaker" data-topic="ProcessingJob" data-id="stop">}\href{../../R6sagemaker/html/ProcessingJob.html#method-stop}{\code{R6sagemaker::ProcessingJob$stop()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="R6sagemaker" data-topic="ProcessingJob" data-id="wait">}\href{../../R6sagemaker/html/ProcessingJob.html#method-wait}{\code{R6sagemaker::ProcessingJob$wait()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initializes a Baselining job that tracks a baselining job kicked off by the suggest
             workflow.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseliningJob$new(
  sagemaker_session = NULL,
  job_name = NULL,
  inputs = NULL,
  outputs = NULL,
  output_kms_key = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sagemaker_session}}{(sagemaker.session.Session): Session object which
manages interactions with Amazon SageMaker APIs and any other
AWS services needed. If not specified, one is created using
the default AWS configuration chain.}

\item{\code{job_name}}{(str): Name of the Amazon SageMaker Model Monitoring Baselining Job.}

\item{\code{inputs}}{([sagemaker.processing.ProcessingInput]): A list of ProcessingInput objects.}

\item{\code{outputs}}{([sagemaker.processing.ProcessingOutput]): A list of ProcessingOutput objects.}

\item{\code{output_kms_key}}{(str): The output kms key associated with the job. Defaults to None
if not provided.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-from_processing_job"></a>}}
\if{latex}{\out{\hypertarget{method-from_processing_job}{}}}
\subsection{Method \code{from_processing_job()}}{
Initializes a Baselining job from a processing job.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseliningJob$from_processing_job(processing_job)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{processing_job}}{(sagemaker.processing.ProcessingJob): The ProcessingJob used for
baselining instance.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
sagemaker.processing.BaseliningJob: The instance of ProcessingJob created
             using the current job name.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-baseline_statistics"></a>}}
\if{latex}{\out{\hypertarget{method-baseline_statistics}{}}}
\subsection{Method \code{baseline_statistics()}}{
Returns a sagemaker.model_monitor.Statistics object representing the statistics
             JSON file generated by this baselining job.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseliningJob$baseline_statistics(
  file_name = STATISTICS_JSON_DEFAULT_FILE_NAME,
  kms_key = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{file_name}}{(str): The name of the json-formatted statistics file}

\item{\code{kms_key}}{(str): The kms key to use when retrieving the file.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
sagemaker.model_monitor.Statistics: The Statistics object representing the file that
             was generated by the job.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-suggested_constraints"></a>}}
\if{latex}{\out{\hypertarget{method-suggested_constraints}{}}}
\subsection{Method \code{suggested_constraints()}}{
Returns a sagemaker.model_monitor.Constraints object representing the constraints
             JSON file generated by this baselining job.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseliningJob$suggested_constraints(
  file_name = CONSTRAINTS_JSON_DEFAULT_FILE_NAME,
  kms_key = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{file_name}}{(str): The name of the json-formatted constraints file}

\item{\code{kms_key}}{(str): The kms key to use when retrieving the file.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
sagemaker.model_monitor.Constraints: The Constraints object representing the file that
             was generated by the job.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Printer.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseliningJob$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{(ignored).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseliningJob$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
