% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_monitor_clarify_model_monitoring.R
\name{BiasAnalysisConfig}
\alias{BiasAnalysisConfig}
\title{BiasAnalysisConfig class}
\description{
Analysis configuration for ModelBiasMonitor.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{analysis_config}}{Analysis config dictionary}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{BiasAnalysisConfig$new()}}
\item \href{#method-to_list}{\code{BiasAnalysisConfig$to_list()}}
\item \href{#method-format}{\code{BiasAnalysisConfig$format()}}
\item \href{#method-clone}{\code{BiasAnalysisConfig$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Creates an analysis config dictionary.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BiasAnalysisConfig$new(bias_config, headers = NULL, label = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{bias_config}}{(sagemaker.clarify.BiasConfig): Config object related to bias
configurations.}

\item{\code{headers}}{(list[str]): A list of column names in the input dataset.}

\item{\code{label}}{(str): Target attribute for the model required by bias metrics.
Specified as column name or index for CSV dataset, or as JSONPath for JSONLines.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-to_list}{}}}
\subsection{Method \code{to_list()}}{
Generates a request dictionary using the parameters provided to the class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BiasAnalysisConfig$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-format"></a>}}
\if{latex}{\out{\hypertarget{method-format}{}}}
\subsection{Method \code{format()}}{
format class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BiasAnalysisConfig$format()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BiasAnalysisConfig$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
