% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clarify.R
\name{BiasConfig}
\alias{BiasConfig}
\title{BiasConfig Class}
\description{
Config object related to bias configurations of the input dataset.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{analysis_config}}{Analysis config dictionary}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{BiasConfig$new()}}
\item \href{#method-get_config}{\code{BiasConfig$get_config()}}
\item \href{#method-format}{\code{BiasConfig$format()}}
\item \href{#method-clone}{\code{BiasConfig$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initializes a configuration of the sensitive groups in the dataset.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BiasConfig$new(
  label_values_or_threshold,
  facet_name,
  facet_values_or_threshold = NULL,
  group_name = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{label_values_or_threshold}}{(Any): List of label values or threshold to indicate positive
outcome used for bias metrics.}

\item{\code{facet_name}}{(str): Sensitive attribute in the input data for which we like to compare
metrics.}

\item{\code{facet_values_or_threshold}}{(list): Optional list of values to form a sensitive group or
threshold for a numeric facet column that defines the lower bound of a sensitive
group. Defaults to considering each possible value as sensitive group and
computing metrics vs all the other examples.}

\item{\code{group_name}}{(str): Optional column name or index to indicate a group column to be used
for the bias metric 'Conditional Demographic Disparity in Labels - CDDL' or
'Conditional Demographic Disparity in Predicted Labels - CDDPL'.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_config"></a>}}
\if{latex}{\out{\hypertarget{method-get_config}{}}}
\subsection{Method \code{get_config()}}{
Returns part of an analysis config dictionary.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BiasConfig$get_config()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-format"></a>}}
\if{latex}{\out{\hypertarget{method-format}{}}}
\subsection{Method \code{format()}}{
format class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BiasConfig$format()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BiasConfig$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
