% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deserializers.R
\name{BytesDeserializer}
\alias{BytesDeserializer}
\title{BytesDerializer Class}
\description{
Deserialize a stream of bytes into a bytes object.
}
\section{Super classes}{
\code{\link[R6sagemaker:BaseDeserializer]{R6sagemaker::BaseDeserializer}} -> \code{\link[R6sagemaker:SimpleBaseDeserializer]{R6sagemaker::SimpleBaseDeserializer}} -> \code{BytesDeserializer}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-deserializer}{\code{BytesDeserializer$deserializer()}}
\item \href{#method-clone}{\code{BytesDeserializer$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="R6sagemaker" data-topic="BaseDeserializer" data-id="deserialize">}\href{../../R6sagemaker/html/BaseDeserializer.html#method-deserialize}{\code{R6sagemaker::BaseDeserializer$deserialize()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="R6sagemaker" data-topic="BaseDeserializer" data-id="format">}\href{../../R6sagemaker/html/BaseDeserializer.html#method-format}{\code{R6sagemaker::BaseDeserializer$format()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="R6sagemaker" data-topic="SimpleBaseDeserializer" data-id="initialize">}\href{../../R6sagemaker/html/SimpleBaseDeserializer.html#method-initialize}{\code{R6sagemaker::SimpleBaseDeserializer$initialize()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-deserializer"></a>}}
\if{latex}{\out{\hypertarget{method-deserializer}{}}}
\subsection{Method \code{deserializer()}}{
Read a stream of bytes returned from an inference endpoint.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BytesDeserializer$deserializer(stream, content_type)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{stream}}{(botocore.response.StreamingBody): A stream of bytes.}

\item{\code{content_type}}{(str): The MIME type of the data.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
bytes: The bytes object read from the stream.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BytesDeserializer$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
