% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automl_candidate_estimator.R
\name{CandidateStep}
\alias{CandidateStep}
\title{CandidateStep Class}
\description{
A class that maintains an AutoML Candidate step's name, inputs, type, and description.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{Name of the candidate step -> (str)}

\item{\code{inputs}}{Inputs of the candidate step -> (dict)}

\item{\code{type}}{Type of the candidate step, Training or Transform -> (str)}

\item{\code{description}}{Description of candidate step job -> (dict)}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{CandidateStep$new()}}
\item \href{#method-format}{\code{CandidateStep$format()}}
\item \href{#method-clone}{\code{CandidateStep$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialize CandidateStep Class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CandidateStep$new(name, inputs, step_type, description)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{(str): Name of the candidate step}

\item{\code{inputs}}{(dict): Inputs of the candidate step}

\item{\code{step_type}}{(str): Type of the candidate step, Training or Transform}

\item{\code{description}}{(dict): Description of candidate step job}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-format"></a>}}
\if{latex}{\out{\hypertarget{method-format}{}}}
\subsection{Method \code{format()}}{
format class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CandidateStep$format()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CandidateStep$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
