% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter.R
\name{CategoricalParameter}
\alias{CategoricalParameter}
\title{CategoricalParameter Class}
\description{
A class for representing hyperparameters that have a discrete list of
             possible values.
}
\section{Super class}{
\code{\link[R6sagemaker:ParameterRange]{R6sagemaker::ParameterRange}} -> \code{CategoricalParameter}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{.name}}{Helps to categorise Class}

\item{\code{values}}{The possible values for the hyperparameter}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{CategoricalParameter$new()}}
\item \href{#method-as_tuning_range}{\code{CategoricalParameter$as_tuning_range()}}
\item \href{#method-as_json_range}{\code{CategoricalParameter$as_json_range()}}
\item \href{#method-is_valid}{\code{CategoricalParameter$is_valid()}}
\item \href{#method-cast_to_type}{\code{CategoricalParameter$cast_to_type()}}
\item \href{#method-clone}{\code{CategoricalParameter$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="R6sagemaker" data-topic="ParameterRange" data-id="format">}\href{../../R6sagemaker/html/ParameterRange.html#method-format}{\code{R6sagemaker::ParameterRange$format()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialize a ``CategoricalParameter``.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CategoricalParameter$new(values)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{values}}{(list or object): The possible values for the hyperparameter.
This input will be converted into a list of strings.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-as_tuning_range"></a>}}
\if{latex}{\out{\hypertarget{method-as_tuning_range}{}}}
\subsection{Method \code{as_tuning_range()}}{
Represent the parameter range as a dicionary suitable for a request
             to create an Amazon SageMaker hyperparameter tuning job.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CategoricalParameter$as_tuning_range(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{(str): The name of the hyperparameter.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
dict[str, list[str]]: A dictionary that contains the name and values
             of the hyperparameter.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-as_json_range"></a>}}
\if{latex}{\out{\hypertarget{method-as_json_range}{}}}
\subsection{Method \code{as_json_range()}}{
Represent the parameter range as a dictionary suitable for a request
             to create an Amazon SageMaker hyperparameter tuning job using one of the
             deep learning frameworks.
             The deep learning framework images require that hyperparameters be
             serialized as JSON.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CategoricalParameter$as_json_range(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{(str): The name of the hyperparameter.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
dict[str, list[str]]: A dictionary that contains the name and values of the
             hyperparameter, where the values are serialized as JSON.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-is_valid"></a>}}
\if{latex}{\out{\hypertarget{method-is_valid}{}}}
\subsection{Method \code{is_valid()}}{
Determine if a value is valid within this CategoricalParameter
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CategoricalParameter$is_valid(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{(object): Value of the hyperparameter}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
boolean: TRUE` or `FALSE`
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-cast_to_type"></a>}}
\if{latex}{\out{\hypertarget{method-cast_to_type}{}}}
\subsection{Method \code{cast_to_type()}}{
cast value to numeric
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CategoricalParameter$cast_to_type(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{The value to be verified.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CategoricalParameter$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
