% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_monitor_clarify_model_monitoring.R
\name{ClarifyBaseliningConfig}
\alias{ClarifyBaseliningConfig}
\title{ClarifyBaseliningConfig class}
\description{
Data class to hold some essential analysis configuration of ClarifyBaseliningJob
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{analysis_config}}{analysis config from configurations of the baselining job.}

\item{\code{features_attribute}}{JSONpath to locate features in predictor request payload.
Only required when predictor content type is JSONlines.}

\item{\code{inference_attribute}}{Index, header or JSONpath to locate predicted label in predictor response payload}

\item{\code{probability_attribute}}{Index or JSONpath location in the model output for
probabilities or scores to be used for explainability.}

\item{\code{probability_threshold_attribute}}{Value to indicate the threshold to select
the binary label in the case of binary classification. Default is 0.5.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ClarifyBaseliningConfig$new()}}
\item \href{#method-print}{\code{ClarifyBaseliningConfig$print()}}
\item \href{#method-clone}{\code{ClarifyBaseliningConfig$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialization.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClarifyBaseliningConfig$new(
  analysis_config,
  features_attribute = NULL,
  inference_attribute = NULL,
  probability_attribute = NULL,
  probability_threshold_attribute = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{analysis_config}}{(BiasAnalysisConfig or ExplainabilityAnalysisConfig): analysis config
from configurations of the baselining job.}

\item{\code{features_attribute}}{(str): JSONpath to locate features in predictor request payload.
Only required when predictor content type is JSONlines.}

\item{\code{inference_attribute}}{(str): Index, header or JSONpath to locate predicted label in
predictor response payload.}

\item{\code{probability_attribute}}{(str): Index or JSONpath location in the model output for
probabilities or scores to be used for explainability.}

\item{\code{probability_threshold_attribute}}{(float): Value to indicate the threshold to select
the binary label in the case of binary classification. Default is 0.5.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Printer.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClarifyBaseliningConfig$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{(ignored).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClarifyBaseliningConfig$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
