% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_monitor_clarify_model_monitoring.R
\name{ClarifyBaseliningJob}
\alias{ClarifyBaseliningJob}
\title{ClarifyBaseliningJob class}
\description{
Provides functionality to retrieve baseline-specific output from Clarify baselining job.
}
\section{Super classes}{
\code{\link[R6sagemaker:.Job]{R6sagemaker::.Job}} -> \code{\link[R6sagemaker:ProcessingJob]{R6sagemaker::ProcessingJob}} -> \code{\link[R6sagemaker:BaseliningJob]{R6sagemaker::BaseliningJob}} -> \code{ClarifyBaseliningJob}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ClarifyBaseliningJob$new()}}
\item \href{#method-baseline_statistics}{\code{ClarifyBaseliningJob$baseline_statistics()}}
\item \href{#method-suggested_constraints}{\code{ClarifyBaseliningJob$suggested_constraints()}}
\item \href{#method-clone}{\code{ClarifyBaseliningJob$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="R6sagemaker" data-topic=".Job" data-id="print">}\href{../../R6sagemaker/html/.Job.html#method-print}{\code{R6sagemaker::.Job$print()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="R6sagemaker" data-topic="ProcessingJob" data-id="describe">}\href{../../R6sagemaker/html/ProcessingJob.html#method-describe}{\code{R6sagemaker::ProcessingJob$describe()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="R6sagemaker" data-topic="ProcessingJob" data-id="from_processing_arn">}\href{../../R6sagemaker/html/ProcessingJob.html#method-from_processing_arn}{\code{R6sagemaker::ProcessingJob$from_processing_arn()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="R6sagemaker" data-topic="ProcessingJob" data-id="from_processing_name">}\href{../../R6sagemaker/html/ProcessingJob.html#method-from_processing_name}{\code{R6sagemaker::ProcessingJob$from_processing_name()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="R6sagemaker" data-topic="ProcessingJob" data-id="start_new">}\href{../../R6sagemaker/html/ProcessingJob.html#method-start_new}{\code{R6sagemaker::ProcessingJob$start_new()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="R6sagemaker" data-topic="ProcessingJob" data-id="stop">}\href{../../R6sagemaker/html/ProcessingJob.html#method-stop}{\code{R6sagemaker::ProcessingJob$stop()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="R6sagemaker" data-topic="ProcessingJob" data-id="wait">}\href{../../R6sagemaker/html/ProcessingJob.html#method-wait}{\code{R6sagemaker::ProcessingJob$wait()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="R6sagemaker" data-topic="BaseliningJob" data-id="from_processing_job">}\href{../../R6sagemaker/html/BaseliningJob.html#method-from_processing_job}{\code{R6sagemaker::BaseliningJob$from_processing_job()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initializes a ClarifyBaseliningJob that tracks a baselining job by suggest_baseline()
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClarifyBaseliningJob$new(processing_job)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{processing_job}}{(sagemaker.processing.ProcessingJob): The ProcessingJob used for
baselining instance.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-baseline_statistics"></a>}}
\if{latex}{\out{\hypertarget{method-baseline_statistics}{}}}
\subsection{Method \code{baseline_statistics()}}{
Not implemented.
             The class doesn't support statistics.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClarifyBaseliningJob$baseline_statistics()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-suggested_constraints"></a>}}
\if{latex}{\out{\hypertarget{method-suggested_constraints}{}}}
\subsection{Method \code{suggested_constraints()}}{
Returns a sagemaker.model_monitor.
             Constraints object representing the constraints JSON file generated by this baselining job.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClarifyBaseliningJob$suggested_constraints(file_name = NULL, kms_key = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{file_name}}{(str): Keep this parameter to align with method signature in super class,
but it will be ignored.}

\item{\code{kms_key}}{(str): The kms key to use when retrieving the file.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
sagemaker.model_monitor.Constraints: The Constraints object representing the file that
             was generated by the job.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClarifyBaseliningJob$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
