% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_monitor_clarify_model_monitoring.R
\name{ClarifyModelMonitor}
\alias{ClarifyModelMonitor}
\title{Base class of Amazon SageMaker Explainability API model monitors.}
\description{
This class is an ``abstract base class``, please instantiate its subclasses
             if you want to monitor bias metrics or feature attribution of an endpoint.
}
\section{Super class}{
\code{\link[R6sagemaker:ModelMonitor]{R6sagemaker::ModelMonitor}} -> \code{ClarifyModelMonitor}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ClarifyModelMonitor$new()}}
\item \href{#method-run_baseline}{\code{ClarifyModelMonitor$run_baseline()}}
\item \href{#method-latest_monitorying_statistics}{\code{ClarifyModelMonitor$latest_monitorying_statistics()}}
\item \href{#method-list_executions}{\code{ClarifyModelMonitor$list_executions()}}
\item \href{#method-clone}{\code{ClarifyModelMonitor$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="R6sagemaker" data-topic="ModelMonitor" data-id="attach">}\href{../../R6sagemaker/html/ModelMonitor.html#method-attach}{\code{R6sagemaker::ModelMonitor$attach()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="R6sagemaker" data-topic="ModelMonitor" data-id="baseline_statistics">}\href{../../R6sagemaker/html/ModelMonitor.html#method-baseline_statistics}{\code{R6sagemaker::ModelMonitor$baseline_statistics()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="R6sagemaker" data-topic="ModelMonitor" data-id="create_monitoring_schedule">}\href{../../R6sagemaker/html/ModelMonitor.html#method-create_monitoring_schedule}{\code{R6sagemaker::ModelMonitor$create_monitoring_schedule()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="R6sagemaker" data-topic="ModelMonitor" data-id="delete_monitoring_schedule">}\href{../../R6sagemaker/html/ModelMonitor.html#method-delete_monitoring_schedule}{\code{R6sagemaker::ModelMonitor$delete_monitoring_schedule()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="R6sagemaker" data-topic="ModelMonitor" data-id="describe_latest_baselining_job">}\href{../../R6sagemaker/html/ModelMonitor.html#method-describe_latest_baselining_job}{\code{R6sagemaker::ModelMonitor$describe_latest_baselining_job()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="R6sagemaker" data-topic="ModelMonitor" data-id="describe_schedule">}\href{../../R6sagemaker/html/ModelMonitor.html#method-describe_schedule}{\code{R6sagemaker::ModelMonitor$describe_schedule()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="R6sagemaker" data-topic="ModelMonitor" data-id="format">}\href{../../R6sagemaker/html/ModelMonitor.html#method-format}{\code{R6sagemaker::ModelMonitor$format()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="R6sagemaker" data-topic="ModelMonitor" data-id="latest_monitoring_constraint_violations">}\href{../../R6sagemaker/html/ModelMonitor.html#method-latest_monitoring_constraint_violations}{\code{R6sagemaker::ModelMonitor$latest_monitoring_constraint_violations()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="R6sagemaker" data-topic="ModelMonitor" data-id="latest_monitoring_statistics">}\href{../../R6sagemaker/html/ModelMonitor.html#method-latest_monitoring_statistics}{\code{R6sagemaker::ModelMonitor$latest_monitoring_statistics()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="R6sagemaker" data-topic="ModelMonitor" data-id="monitoring_type">}\href{../../R6sagemaker/html/ModelMonitor.html#method-monitoring_type}{\code{R6sagemaker::ModelMonitor$monitoring_type()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="R6sagemaker" data-topic="ModelMonitor" data-id="start_monitoring_schedule">}\href{../../R6sagemaker/html/ModelMonitor.html#method-start_monitoring_schedule}{\code{R6sagemaker::ModelMonitor$start_monitoring_schedule()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="R6sagemaker" data-topic="ModelMonitor" data-id="stop_monitoring_schedule">}\href{../../R6sagemaker/html/ModelMonitor.html#method-stop_monitoring_schedule}{\code{R6sagemaker::ModelMonitor$stop_monitoring_schedule()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="R6sagemaker" data-topic="ModelMonitor" data-id="suggested_constraints">}\href{../../R6sagemaker/html/ModelMonitor.html#method-suggested_constraints}{\code{R6sagemaker::ModelMonitor$suggested_constraints()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="R6sagemaker" data-topic="ModelMonitor" data-id="update_monitoring_schedule">}\href{../../R6sagemaker/html/ModelMonitor.html#method-update_monitoring_schedule}{\code{R6sagemaker::ModelMonitor$update_monitoring_schedule()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initializes a monitor instance.
             The monitor handles baselining datasets and creating Amazon SageMaker
             Monitoring Schedules to monitor SageMaker endpoints.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClarifyModelMonitor$new(
  role,
  instance_count = 1,
  instance_type = "ml.m5.xlarge",
  volume_size_in_gb = 30,
  volume_kms_key = NULL,
  output_kms_key = NULL,
  max_runtime_in_seconds = NULL,
  base_job_name = NULL,
  sagemaker_session = NULL,
  env = NULL,
  tags = NULL,
  network_config = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{role}}{(str): An AWS IAM role. The Amazon SageMaker jobs use this role.}

\item{\code{instance_count}}{(int): The number of instances to run
the jobs with.}

\item{\code{instance_type}}{(str): Type of EC2 instance to use for
the job, for example, 'ml.m5.xlarge'.}

\item{\code{volume_size_in_gb}}{(int): Size in GB of the EBS volume
to use for storing data during processing (default: 30).}

\item{\code{volume_kms_key}}{(str): A KMS key for the job's volume.}

\item{\code{output_kms_key}}{(str): The KMS key id for the job's outputs.}

\item{\code{max_runtime_in_seconds}}{(int): Timeout in seconds. After this amount of
time, Amazon SageMaker terminates the job regardless of its current status.
Default: 3600}

\item{\code{base_job_name}}{(str): Prefix for the job name. If not specified,
a default name is generated based on the training image name and
current timestamp.}

\item{\code{sagemaker_session}}{(sagemaker.session.Session): Session object which
manages interactions with Amazon SageMaker APIs and any other
AWS services needed. If not specified, one is created using
the default AWS configuration chain.}

\item{\code{env}}{(dict): Environment variables to be passed to the job.}

\item{\code{tags}}{([dict]): List of tags to be passed to the job.}

\item{\code{network_config}}{(sagemaker.network.NetworkConfig): A NetworkConfig
object that configures network isolation, encryption of
inter-container traffic, security group IDs, and subnets.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-run_baseline"></a>}}
\if{latex}{\out{\hypertarget{method-run_baseline}{}}}
\subsection{Method \code{run_baseline()}}{
Not implemented.
             .run_baseline()' is only allowed for ModelMonitor objects.
             Please use `suggest_baseline` instead.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClarifyModelMonitor$run_baseline(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{: Unused argument}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-latest_monitorying_statistics"></a>}}
\if{latex}{\out{\hypertarget{method-latest_monitorying_statistics}{}}}
\subsection{Method \code{latest_monitorying_statistics()}}{
Not implemented.
             The class doesn't support statistics.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClarifyModelMonitor$latest_monitorying_statistics(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{: Unused argument}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-list_executions"></a>}}
\if{latex}{\out{\hypertarget{method-list_executions}{}}}
\subsection{Method \code{list_executions()}}{
Get the list of the latest monitoring executions in descending order of "ScheduledTime".
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClarifyModelMonitor$list_executions()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
[sagemaker.model_monitor.ClarifyMonitoringExecution]: List of
             ClarifyMonitoringExecution in descending order of "ScheduledTime".
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClarifyModelMonitor$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
