% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_monitor_clarify_model_monitoring.R
\name{ClarifyMonitoringExecution}
\alias{ClarifyMonitoringExecution}
\title{ClarifyMonitoringExecution class}
\description{
Provides functionality to retrieve monitoring-specific files output from executions.
}
\section{Super classes}{
\code{\link[R6sagemaker.common:.Job]{R6sagemaker.common::.Job}} -> \code{\link[R6sagemaker.common:ProcessingJob]{R6sagemaker.common::ProcessingJob}} -> \code{\link[R6sagemaker:MonitoringExecution]{R6sagemaker::MonitoringExecution}} -> \code{ClarifyMonitoringExecution}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ClarifyMonitoringExecution$new()}}
\item \href{#method-statistics}{\code{ClarifyMonitoringExecution$statistics()}}
\item \href{#method-clone}{\code{ClarifyMonitoringExecution$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="R6sagemaker.common" data-topic=".Job" data-id="format">}\href{../../R6sagemaker.common/html/.Job.html#method-format}{\code{R6sagemaker.common::.Job$format()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="R6sagemaker.common" data-topic="ProcessingJob" data-id="describe">}\href{../../R6sagemaker.common/html/ProcessingJob.html#method-describe}{\code{R6sagemaker.common::ProcessingJob$describe()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="R6sagemaker.common" data-topic="ProcessingJob" data-id="from_processing_name">}\href{../../R6sagemaker.common/html/ProcessingJob.html#method-from_processing_name}{\code{R6sagemaker.common::ProcessingJob$from_processing_name()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="R6sagemaker.common" data-topic="ProcessingJob" data-id="start_new">}\href{../../R6sagemaker.common/html/ProcessingJob.html#method-start_new}{\code{R6sagemaker.common::ProcessingJob$start_new()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="R6sagemaker.common" data-topic="ProcessingJob" data-id="stop">}\href{../../R6sagemaker.common/html/ProcessingJob.html#method-stop}{\code{R6sagemaker.common::ProcessingJob$stop()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="R6sagemaker.common" data-topic="ProcessingJob" data-id="wait">}\href{../../R6sagemaker.common/html/ProcessingJob.html#method-wait}{\code{R6sagemaker.common::ProcessingJob$wait()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="R6sagemaker" data-topic="MonitoringExecution" data-id="constraint_violations">}\href{../../R6sagemaker/html/MonitoringExecution.html#method-constraint_violations}{\code{R6sagemaker::MonitoringExecution$constraint_violations()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="R6sagemaker" data-topic="MonitoringExecution" data-id="from_processing_arn">}\href{../../R6sagemaker/html/MonitoringExecution.html#method-from_processing_arn}{\code{R6sagemaker::MonitoringExecution$from_processing_arn()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initializes an object that tracks a monitoring execution by a Clarify model monitor
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClarifyMonitoringExecution$new(
  sagemaker_session,
  job_name,
  inputs,
  output,
  output_kms_key = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sagemaker_session}}{(sagemaker.session.Session): Session object which
manages interactions with Amazon SageMaker APIs and any other
AWS services needed. If not specified, one is created using
the default AWS configuration chain.}

\item{\code{job_name}}{(str): The name of the monitoring execution job.}

\item{\code{inputs}}{(list[:class:`~sagemaker.processing.ProcessingInput`]): A list of
:class:`~sagemaker.processing.ProcessingInput` objects.}

\item{\code{output}}{(sagemaker.Processing.ProcessingOutput): The output associated with the
monitoring execution.}

\item{\code{output_kms_key}}{(str): The output kms key associated with the job. Defaults to None
if not provided.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-statistics"></a>}}
\if{latex}{\out{\hypertarget{method-statistics}{}}}
\subsection{Method \code{statistics()}}{
Not implemented.
             The class doesn't support statistics.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClarifyMonitoringExecution$statistics()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClarifyMonitoringExecution$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
