% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debugger.R
\name{CollectionConfig}
\alias{CollectionConfig}
\title{CollectionConfig Class}
\description{
Creates tensor collections for SageMaker Debugger
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{The name of the collection configuration.}

\item{\code{parameters}}{The parameters for the collection configuration.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{CollectionConfig$new()}}
\item \href{#method-to_request_list}{\code{CollectionConfig$to_request_list()}}
\item \href{#method-format}{\code{CollectionConfig$format()}}
\item \href{#method-clone}{\code{CollectionConfig$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Constructor for collection configuration.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CollectionConfig$new(name, parameters = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{(str): Required. The name of the collection configuration.}

\item{\code{parameters}}{(dict): Optional. The parameters for the collection
configuration.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-to_request_list"></a>}}
\if{latex}{\out{\hypertarget{method-to_request_list}{}}}
\subsection{Method \code{to_request_list()}}{
Generate a request dictionary using the parameters initializing the object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CollectionConfig$to_request_list()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
dict: A portion of an API request as a dictionary.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-format"></a>}}
\if{latex}{\out{\hypertarget{method-format}{}}}
\subsection{Method \code{format()}}{
format class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CollectionConfig$format()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CollectionConfig$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
