% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session.R
\name{container_def}
\alias{container_def}
\title{Create a definition for executing a container as part of a SageMaker model.}
\usage{
container_def(image, model_data_url = NULL, env = NULL, container_mode = NULL)
}
\arguments{
\item{image}{(str): Docker image to run for this container.}

\item{model_data_url}{(str): S3 URI of data required by this container,
e.g. SageMaker training job model artifacts (default: None).}

\item{env}{(dict[str, str]): Environment variables to set inside the container (default: None).}

\item{container_mode}{(str): The model container mode. Valid modes:
\itemize{
  \item{\strong{MultiModel:} Indicates that model container can support hosting multiple models}
  \item{\strong{SingleModel:} Indicates that model container can support hosting a single model
                This is the default model container mode when container_mode = None}}}
}
\value{
dict[str, str]: A complete container definition object usable with the CreateModel API if
             passed via `PrimaryContainers` field.
}
\description{
Create a definition for executing a container as part of a SageMaker model.
}
