% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debugger_utils.R
\docType{data}
\name{cProfileTimer}
\alias{cProfileTimer}
\title{cProfileTimer enum environment list}
\format{
An object of class \code{cProfileTimer} (inherits from \code{Enum}, \code{environment}) of length 4.
}
\usage{
cProfileTimer
}
\value{
environment containing [TOTAL_TIME, CPU_TIME, OFF_CPU_TIME, DEFAULT]
}
\description{
List the possible cProfile timers for Python profiling.
}
\keyword{datasets}
