% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_monitor_cron_expression_generator.R
\name{CronExpressionGenerator}
\alias{CronExpressionGenerator}
\title{CronExpressionGenerator class}
\description{
Generates cron expression strings for the SageMaker Model Monitoring Schedule API.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-hourly}{\code{CronExpressionGenerator$hourly()}}
\item \href{#method-daily}{\code{CronExpressionGenerator$daily()}}
\item \href{#method-daily_every_x_hours}{\code{CronExpressionGenerator$daily_every_x_hours()}}
\item \href{#method-print}{\code{CronExpressionGenerator$print()}}
\item \href{#method-clone}{\code{CronExpressionGenerator$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-hourly"></a>}}
\if{latex}{\out{\hypertarget{method-hourly}{}}}
\subsection{Method \code{hourly()}}{
Generates hourly cron expression that denotes that a job runs at the top of every hour.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CronExpressionGenerator$hourly()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
str: The cron expression format accepted by the Amazon SageMaker Model Monitoring
             Schedule API.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-daily"></a>}}
\if{latex}{\out{\hypertarget{method-daily}{}}}
\subsection{Method \code{daily()}}{
Generates daily cron expression that denotes that a job runs at the top of every hour.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CronExpressionGenerator$daily(hour = 0L)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{hour}}{(int): The hour in HH24 format (UTC) to run the job at, on a daily schedule.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
str: The cron expression format accepted by the Amazon SageMaker Model Monitoring
             Schedule API.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-daily_every_x_hours"></a>}}
\if{latex}{\out{\hypertarget{method-daily_every_x_hours}{}}}
\subsection{Method \code{daily_every_x_hours()}}{
Generates "daily every x hours" cron expression.
             That denotes that a job runs every day at the specified hour, and then every x hours,
             as specified in hour_interval.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CronExpressionGenerator$daily_every_x_hours(hour_interval, starting_hour = 0L)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{hour_interval}}{(int): The hour interval to run the job at.}

\item{\code{starting_hour}}{(int): The hour at which to begin in HH24 format (UTC).}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
str: The cron expression format accepted by the Amazon SageMaker Model Monitoring
             Schedule API.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Printer.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CronExpressionGenerator$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{(ignored).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CronExpressionGenerator$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
