% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deserializers.R
\docType{data}
\name{csv_deserializer}
\alias{csv_deserializer}
\title{S3 method to call CsvBaseDeserializer class}
\format{
An object of class \code{CsvBaseDeserializer} (inherits from \code{BaseDeserializer}, \code{R6}) of length 6.
}
\usage{
csv_deserializer
}
\description{
S3 method to call CsvBaseDeserializer class
}
\keyword{datasets}
