% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serializers.R
\docType{data}
\name{csv_serializer}
\alias{csv_serializer}
\title{S3 method to call CsvSerializer class}
\format{
An object of class \code{CsvSerializer} (inherits from \code{BaseSerializer}, \code{R6}) of length 6.
}
\usage{
csv_serializer
}
\description{
S3 method to call CsvSerializer class
}
\keyword{datasets}
