% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_capture_config.R
\name{DataCaptureConfig}
\alias{DataCaptureConfig}
\title{DataCaptureConfig Class}
\description{
Configuration object passed in when deploying models to Amazon SageMaker Endpoints.
             This object specifies configuration related to endpoint data capture for use with
             Amazon SageMaker Model Monitoring.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{enable_capture}}{Whether data capture should be enabled or not.}

\item{\code{sampling_percentage}}{The percentage of data to sample.}

\item{\code{destination_s3_uri}}{Defaults to "s3://<default-session-bucket>/model-monitor/data-capture"}

\item{\code{kms_key_id}}{The kms key to use when writing to S3.}

\item{\code{capture_options}}{Denotes which data to capture between request and response.}

\item{\code{csv_content_types}}{Default=["text/csv"].}

\item{\code{json_content_types}}{Default=["application/json"].}

\item{\code{sagemaker_session}}{A SageMaker Session object}

\item{\code{API_MAPPING}}{Convert to API values or pass value directly through if unable to convert}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{to_request_list}}{Generates a request named list using the parameters provided to the class.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{DataCaptureConfig$new()}}
\item \href{#method-print}{\code{DataCaptureConfig$print()}}
\item \href{#method-clone}{\code{DataCaptureConfig$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialize a DataCaptureConfig object for capturing data from Amazon SageMaker Endpoints.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataCaptureConfig$new(
  enable_capture = TRUE,
  sampling_percentage = 20L,
  destination_s3_uri = NULL,
  kms_key_id = NULL,
  capture_options = NULL,
  csv_content_types = NULL,
  json_content_types = NULL,
  sagemaker_session = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{enable_capture}}{(bool): Required. Whether data capture should be enabled or not.}

\item{\code{sampling_percentage}}{(int): Optional. Default=20. The percentage of data to sample.
Must be between 0 and 100.}

\item{\code{destination_s3_uri}}{(str): Optional. Defaults to "s3://<default-session-bucket>/
model-monitor/data-capture".}

\item{\code{kms_key_id}}{(str): Optional. Default=None. The kms key to use when writing to S3.}

\item{\code{capture_options}}{([str]): Optional. Must be a list containing any combination of the
following values: "REQUEST", "RESPONSE". Default=["REQUEST", "RESPONSE"]. Denotes
which data to capture between request and response.}

\item{\code{csv_content_types}}{([str]): Optional. Default=["text/csv"].}

\item{\code{json_content_types}}{([str]): Optional. Default=["application/json"].}

\item{\code{sagemaker_session}}{(sagemaker.session.Session): A SageMaker Session
object, used for SageMaker interactions (default: None). If not
specified, one is created using the default AWS configuration
chain.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Printer.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataCaptureConfig$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{(ignored).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataCaptureConfig$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
