% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clarify.R
\name{DataConfig}
\alias{DataConfig}
\title{DataConfig Class}
\description{
Config object related to configurations of the input and output dataset.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{s3_data_input_path}}{Dataset S3 prefix/object URI.}

\item{\code{s3_output_path}}{S3 prefix to store the output.}

\item{\code{s3_data_distribution_type}}{Valid options are "FullyReplicated" or "ShardedByS3Key".}

\item{\code{s3_compression_type}}{Valid options are "None" or "Gzip".}

\item{\code{label}}{Target attribute of the model required by bias metrics}

\item{\code{features}}{JSONPath for locating the feature columns}

\item{\code{analysis_config}}{Analysis config dictionary}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{DataConfig$new()}}
\item \href{#method-get_config}{\code{DataConfig$get_config()}}
\item \href{#method-print}{\code{DataConfig$print()}}
\item \href{#method-clone}{\code{DataConfig$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initializes a configuration of both input and output datasets.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataConfig$new(
  s3_data_input_path,
  s3_output_path,
  label = NULL,
  headers = NULL,
  features = NULL,
  dataset_type = "text/csv",
  s3_data_distribution_type = "FullyReplicated",
  s3_compression_type = c("None", "Gzip")
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{s3_data_input_path}}{(str): Dataset S3 prefix/object URI.}

\item{\code{s3_output_path}}{(str): S3 prefix to store the output.}

\item{\code{label}}{(str): Target attribute of the model required by bias metrics (optional for SHAP)
Specified as column name or index for CSV dataset, or as JSONPath for JSONLines.}

\item{\code{headers}}{(list[str]): A list of column names in the input dataset.}

\item{\code{features}}{(str): JSONPath for locating the feature columns for bias metrics if the
dataset format is JSONLines.}

\item{\code{dataset_type}}{(str): Format of the dataset. Valid values are "text/csv" for CSV
and "application/jsonlines" for JSONLines.}

\item{\code{s3_data_distribution_type}}{(str): Valid options are "FullyReplicated" or
"ShardedByS3Key".}

\item{\code{s3_compression_type}}{(str): Valid options are "None" or "Gzip".}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_config"></a>}}
\if{latex}{\out{\hypertarget{method-get_config}{}}}
\subsection{Method \code{get_config()}}{
Returns part of an analysis config dictionary.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataConfig$get_config()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Printer.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataConfig$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{(ignored).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataConfig$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
