% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debugger_metrics_config.R
\name{DataloaderProfilingConfig}
\alias{DataloaderProfilingConfig}
\title{DataloaderProfilingConfig Class}
\description{
The configuration for framework metrics to be collected for data loader profiling.
}
\section{Super class}{
\code{\link[R6sagemaker:MetricsConfigBase]{R6sagemaker::MetricsConfigBase}} -> \code{DataloaderProfilingConfig}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{DataloaderProfilingConfig$new()}}
\item \href{#method-clone}{\code{DataloaderProfilingConfig$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="R6sagemaker" data-topic="MetricsConfigBase" data-id="print">}\href{../../R6sagemaker/html/MetricsConfigBase.html#method-print}{\code{R6sagemaker::MetricsConfigBase$print()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="R6sagemaker" data-topic="MetricsConfigBase" data-id="to_json_string">}\href{../../R6sagemaker/html/MetricsConfigBase.html#method-to_json_string}{\code{R6sagemaker::MetricsConfigBase$to_json_string()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Specify target steps or a target duration to profile.
             By default, it profiles step 7 of
             training. If \code{profile_default_steps} is set to `True` and none of the other
             range parameters is specified,
             the class uses the default config for dataloader profiling.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataloaderProfilingConfig$new(
  start_step = NULL,
  num_steps = NULL,
  start_unix_time = NULL,
  duration = NULL,
  profile_default_steps = FALSE,
  metrics_regex = ".*"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{start_step}}{(int): The step to start profiling. The default is step 7.}

\item{\code{num_steps}}{(int): The number of steps to profile. The default is for 1 step.}

\item{\code{start_unix_time}}{(int): The Unix time to start profiling. The default is for 1 step.}

\item{\code{duration}}{(float): The duration in seconds to profile.}

\item{\code{profile_default_steps}}{(bool): Indicates whether the default config should be used.}

\item{\code{metrics_regex}}{(str): Regex pattern}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataloaderProfilingConfig$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
