% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_monitor_dataset_format.R
\name{DatasetFormat}
\alias{DatasetFormat}
\title{DatasetFormat Class}
\description{
Represents a Dataset Format that is used when calling a DefaultModelMonitor.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-csv}{\code{DatasetFormat$csv()}}
\item \href{#method-josn}{\code{DatasetFormat$josn()}}
\item \href{#method-sagemaker_capture_json}{\code{DatasetFormat$sagemaker_capture_json()}}
\item \href{#method-format}{\code{DatasetFormat$format()}}
\item \href{#method-clone}{\code{DatasetFormat$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-csv"></a>}}
\if{latex}{\out{\hypertarget{method-csv}{}}}
\subsection{Method \code{csv()}}{
Returns a DatasetFormat JSON string for use with a DefaultModelMonitor.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatasetFormat$csv(header = TRUE, output_columns_position = c("START", "END"))}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{header}}{(bool): Whether the csv dataset to baseline and monitor has a header.
Default: True.}

\item{\code{output_columns_position}}{(str): The position of the output columns.
Must be one of ("START", "END"). Default: "START".}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
dict: JSON string containing DatasetFormat to be used by DefaultModelMonitor.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-josn"></a>}}
\if{latex}{\out{\hypertarget{method-josn}{}}}
\subsection{Method \code{josn()}}{
Returns a DatasetFormat JSON string for use with a DefaultModelMonitor.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatasetFormat$josn(lines = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{lines}}{(bool): Whether the file should be read as a json object per line. Default: True.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
dict: JSON string containing DatasetFormat to be used by DefaultModelMonitor.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sagemaker_capture_json"></a>}}
\if{latex}{\out{\hypertarget{method-sagemaker_capture_json}{}}}
\subsection{Method \code{sagemaker_capture_json()}}{
Returns a DatasetFormat SageMaker Capture Json string for use with a DefaultModelMonitor.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatasetFormat$sagemaker_capture_json()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
dict: JSON string containing DatasetFormat to be used by DefaultModelMonitor.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-format"></a>}}
\if{latex}{\out{\hypertarget{method-format}{}}}
\subsection{Method \code{format()}}{
format class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatasetFormat$format()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatasetFormat$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
