% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debugger.R
\name{DebuggerHookConfig}
\alias{DebuggerHookConfig}
\title{Create a Debugger hook configuration object to save the tensor for debugging.}
\description{
DebuggerHookConfig provides options to customize how debugging
             information is emitted and saved. This high-level DebuggerHookConfig class
             runs based on the `smdebug.SaveConfig
             <https://github.com/awslabs/sagemaker-debugger/blob/master/docs/api.md#saveconfig>`
             class.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{s3_output_path}}{The location in Amazon S3 to store the output tensors}

\item{\code{container_local_output_path}}{The local path in the container}

\item{\code{hook_parameters}}{A dictionary of parameters}

\item{\code{collection_configs}}{A list of :class:`~sagemaker.debugger.CollectionConfig}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{DebuggerHookConfig$new()}}
\item \href{#method-to_request_list}{\code{DebuggerHookConfig$to_request_list()}}
\item \href{#method-format}{\code{DebuggerHookConfig$format()}}
\item \href{#method-clone}{\code{DebuggerHookConfig$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialize the DebuggerHookConfig instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DebuggerHookConfig$new(
  s3_output_path = NULL,
  container_local_output_path = NULL,
  hook_parameters = NULL,
  collection_configs = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{s3_output_path}}{(str): Optional. The location in Amazon S3 to store the output tensors.
The default Debugger output path is created under the
default output path of the :class:`~sagemaker.estimator.Estimator` class.
For example,
s3://sagemaker-<region>-<12digit_account_id>/<training-job-name>/debug-output/.}

\item{\code{container_local_output_path}}{(str): Optional. The local path in the container.}

\item{\code{hook_parameters}}{(dict): Optional. A dictionary of parameters.}

\item{\code{collection_configs}}{([sagemaker.debugger.CollectionConfig]): Required. A list
of :class:`~sagemaker.debugger.CollectionConfig` objects to be saved
at the \code{s3_output_path}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-to_request_list"></a>}}
\if{latex}{\out{\hypertarget{method-to_request_list}{}}}
\subsection{Method \code{to_request_list()}}{
Generate a request dictionary using the parameters when initializing the object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DebuggerHookConfig$to_request_list()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
dict: An portion of an API request as a dictionary.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-format"></a>}}
\if{latex}{\out{\hypertarget{method-format}{}}}
\subsection{Method \code{format()}}{
format class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DebuggerHookConfig$format()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DebuggerHookConfig$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
