% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_monitor_model_monitoring.R
\name{DefaultModelMonitor}
\alias{DefaultModelMonitor}
\title{DefaultModelMonitor Class}
\description{
Sets up Amazon SageMaker Monitoring Schedules and baseline suggestions. Use this class when
             you want to utilize Amazon SageMaker Monitoring's plug-and-play solution that only requires
             your dataset and optional pre/postprocessing scripts.
             For a more customized experience, consider using the ModelMonitor class instead.
}
\section{Super class}{
\code{\link[R6sagemaker:ModelMonitor]{R6sagemaker::ModelMonitor}} -> \code{DefaultModelMonitor}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{JOB_DEFINITION_BASE_NAME}}{Model definition base name}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{DefaultModelMonitor$new()}}
\item \href{#method-monitoring_type}{\code{DefaultModelMonitor$monitoring_type()}}
\item \href{#method-suggest_baseline}{\code{DefaultModelMonitor$suggest_baseline()}}
\item \href{#method-create_monitoring_schedule}{\code{DefaultModelMonitor$create_monitoring_schedule()}}
\item \href{#method-update_monitoring_schedule}{\code{DefaultModelMonitor$update_monitoring_schedule()}}
\item \href{#method-delete_monitoring_schedule}{\code{DefaultModelMonitor$delete_monitoring_schedule()}}
\item \href{#method-run_baseline}{\code{DefaultModelMonitor$run_baseline()}}
\item \href{#method-attach}{\code{DefaultModelMonitor$attach()}}
\item \href{#method-latest_monitoring_statistics}{\code{DefaultModelMonitor$latest_monitoring_statistics()}}
\item \href{#method-latest_monitoring_constraint_violations}{\code{DefaultModelMonitor$latest_monitoring_constraint_violations()}}
\item \href{#method-clone}{\code{DefaultModelMonitor$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="R6sagemaker" data-topic="ModelMonitor" data-id="baseline_statistics">}\href{../../R6sagemaker/html/ModelMonitor.html#method-baseline_statistics}{\code{R6sagemaker::ModelMonitor$baseline_statistics()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="R6sagemaker" data-topic="ModelMonitor" data-id="describe_latest_baselining_job">}\href{../../R6sagemaker/html/ModelMonitor.html#method-describe_latest_baselining_job}{\code{R6sagemaker::ModelMonitor$describe_latest_baselining_job()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="R6sagemaker" data-topic="ModelMonitor" data-id="describe_schedule">}\href{../../R6sagemaker/html/ModelMonitor.html#method-describe_schedule}{\code{R6sagemaker::ModelMonitor$describe_schedule()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="R6sagemaker" data-topic="ModelMonitor" data-id="list_executions">}\href{../../R6sagemaker/html/ModelMonitor.html#method-list_executions}{\code{R6sagemaker::ModelMonitor$list_executions()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="R6sagemaker" data-topic="ModelMonitor" data-id="print">}\href{../../R6sagemaker/html/ModelMonitor.html#method-print}{\code{R6sagemaker::ModelMonitor$print()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="R6sagemaker" data-topic="ModelMonitor" data-id="start_monitoring_schedule">}\href{../../R6sagemaker/html/ModelMonitor.html#method-start_monitoring_schedule}{\code{R6sagemaker::ModelMonitor$start_monitoring_schedule()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="R6sagemaker" data-topic="ModelMonitor" data-id="stop_monitoring_schedule">}\href{../../R6sagemaker/html/ModelMonitor.html#method-stop_monitoring_schedule}{\code{R6sagemaker::ModelMonitor$stop_monitoring_schedule()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="R6sagemaker" data-topic="ModelMonitor" data-id="suggested_constraints">}\href{../../R6sagemaker/html/ModelMonitor.html#method-suggested_constraints}{\code{R6sagemaker::ModelMonitor$suggested_constraints()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initializes a ``Monitor`` instance. The Monitor handles baselining datasets and
             creating Amazon SageMaker Monitoring Schedules to monitor SageMaker endpoints.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DefaultModelMonitor$new(
  role,
  instance_count = 1,
  instance_type = "ml.m5.xlarge",
  volume_size_in_gb = 30,
  volume_kms_key = NULL,
  output_kms_key = NULL,
  max_runtime_in_seconds = NULL,
  base_job_name = NULL,
  sagemaker_session = NULL,
  env = NULL,
  tags = NULL,
  network_config = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{role}}{(str): An AWS IAM role name or ARN. The Amazon SageMaker jobs use this role.}

\item{\code{instance_count}}{(int): The number of instances to run the jobs with.}

\item{\code{instance_type}}{(str): Type of EC2 instance to use for the job, for example,
'ml.m5.xlarge'.}

\item{\code{volume_size_in_gb}}{(int): Size in GB of the EBS volume
to use for storing data during processing (default: 30).}

\item{\code{volume_kms_key}}{(str): A KMS key for the processing volume.}

\item{\code{output_kms_key}}{(str): The KMS key id for the job's outputs.}

\item{\code{max_runtime_in_seconds}}{(int): Timeout in seconds. After this amount of
time, Amazon SageMaker terminates the job regardless of its current status.
Default: 3600}

\item{\code{base_job_name}}{(str): Prefix for the job name. If not specified,
a default name is generated based on the training image name and
current timestamp.}

\item{\code{sagemaker_session}}{(sagemaker.session.Session): Session object which
manages interactions with Amazon SageMaker APIs and any other
AWS services needed. If not specified, one is created using
the default AWS configuration chain.}

\item{\code{env}}{(dict): Environment variables to be passed to the job.}

\item{\code{tags}}{([dict]): List of tags to be passed to the job.}

\item{\code{network_config}}{(sagemaker.network.NetworkConfig): A NetworkConfig
object that configures network isolation, encryption of
inter-container traffic, security group IDs, and subnets.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-monitoring_type"></a>}}
\if{latex}{\out{\hypertarget{method-monitoring_type}{}}}
\subsection{Method \code{monitoring_type()}}{
Type of the monitoring job.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DefaultModelMonitor$monitoring_type()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-suggest_baseline"></a>}}
\if{latex}{\out{\hypertarget{method-suggest_baseline}{}}}
\subsection{Method \code{suggest_baseline()}}{
Suggest baselines for use with Amazon SageMaker Model Monitoring Schedules.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DefaultModelMonitor$suggest_baseline(
  baseline_dataset,
  dataset_format,
  record_preprocessor_script = NULL,
  post_analytics_processor_script = NULL,
  output_s3_uri = NULL,
  wait = TRUE,
  logs = TRUE,
  job_name = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{baseline_dataset}}{(str): The path to the baseline_dataset file. This can be a local path
or an S3 uri.}

\item{\code{dataset_format}}{(dict): The format of the baseline_dataset.}

\item{\code{record_preprocessor_script}}{(str): The path to the record preprocessor script. This can
be a local path or an S3 uri.}

\item{\code{post_analytics_processor_script}}{(str): The path to the record post-analytics processor
script. This can be a local path or an S3 uri.}

\item{\code{output_s3_uri}}{(str): Desired S3 destination Destination of the constraint_violations
and statistics json files.
Default: "s3://<default_session_bucket>/<job_name>/output"}

\item{\code{wait}}{(bool): Whether the call should wait until the job completes (default: True).}

\item{\code{logs}}{(bool): Whether to show the logs produced by the job.
Only meaningful when wait is True (default: True).}

\item{\code{job_name}}{(str): Processing job name. If not specified, the processor generates
a default job name, based on the image name and current timestamp.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
sagemaker.processing.ProcessingJob: The ProcessingJob object representing the
             baselining job.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-create_monitoring_schedule"></a>}}
\if{latex}{\out{\hypertarget{method-create_monitoring_schedule}{}}}
\subsection{Method \code{create_monitoring_schedule()}}{
Creates a monitoring schedule to monitor an Amazon SageMaker Endpoint.
             If constraints and statistics are provided, or if they are able to be retrieved from a
             previous baselining job associated with this monitor, those will be used.
             If constraints and statistics cannot be automatically retrieved, baseline_inputs will be
             required in order to kick off a baselining job.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DefaultModelMonitor$create_monitoring_schedule(
  endpoint_input,
  record_preprocessor_script = NULL,
  post_analytics_processor_script = NULL,
  output_s3_uri = NULL,
  constraints = NULL,
  statistics = NULL,
  monitor_schedule_name = NULL,
  schedule_cron_expression = NULL,
  enable_cloudwatch_metrics = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{endpoint_input}}{(str or sagemaker.model_monitor.EndpointInput): The endpoint to monitor.
This can either be the endpoint name or an EndpointInput.}

\item{\code{record_preprocessor_script}}{(str): The path to the record preprocessor script. This can
be a local path or an S3 uri.}

\item{\code{post_analytics_processor_script}}{(str): The path to the record post-analytics processor
script. This can be a local path or an S3 uri.}

\item{\code{output_s3_uri}}{(str): Desired S3 destination of the constraint_violations and
statistics json files.
Default: "s3://<default_session_bucket>/<job_name>/output"}

\item{\code{constraints}}{(sagemaker.model_monitor.Constraints or str): If provided alongside
statistics, these will be used for monitoring the endpoint. This can be a
sagemaker.model_monitor.Constraints object or an s3_uri pointing to a constraints
JSON file.}

\item{\code{statistics}}{(sagemaker.model_monitor.Statistic or str): If provided alongside
constraints, these will be used for monitoring the endpoint. This can be a
sagemaker.model_monitor.Constraints object or an s3_uri pointing to a constraints
JSON file.}

\item{\code{monitor_schedule_name}}{(str): Schedule name. If not specified, the processor generates
a default job name, based on the image name and current timestamp.}

\item{\code{schedule_cron_expression}}{(str): The cron expression that dictates the frequency that
this job run. See sagemaker.model_monitor.CronExpressionGenerator for valid
expressions. Default: Daily.}

\item{\code{enable_cloudwatch_metrics}}{(bool): Whether to publish cloudwatch metrics as part of
the baselining or monitoring jobs.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update_monitoring_schedule"></a>}}
\if{latex}{\out{\hypertarget{method-update_monitoring_schedule}{}}}
\subsection{Method \code{update_monitoring_schedule()}}{
Updates the existing monitoring schedule.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DefaultModelMonitor$update_monitoring_schedule(
  endpoint_input = NULL,
  record_preprocessor_script = NULL,
  post_analytics_processor_script = NULL,
  output_s3_uri = NULL,
  statistics = NULL,
  constraints = NULL,
  schedule_cron_expression = NULL,
  instance_count = NULL,
  instance_type = NULL,
  volume_size_in_gb = NULL,
  volume_kms_key = NULL,
  output_kms_key = NULL,
  max_runtime_in_seconds = NULL,
  env = NULL,
  network_config = NULL,
  enable_cloudwatch_metrics = NULL,
  role = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{endpoint_input}}{(str or sagemaker.model_monitor.EndpointInput): The endpoint to monitor.
This can either be the endpoint name or an EndpointInput.}

\item{\code{record_preprocessor_script}}{(str): The path to the record preprocessor script. This can
be a local path or an S3 uri.}

\item{\code{post_analytics_processor_script}}{(str): The path to the record post-analytics processor
script. This can be a local path or an S3 uri.}

\item{\code{output_s3_uri}}{(str): Desired S3 destination of the constraint_violations and
statistics json files.}

\item{\code{statistics}}{(sagemaker.model_monitor.Statistic or str): If provided alongside
constraints, these will be used for monitoring the endpoint. This can be a
sagemaker.model_monitor.Constraints object or an S3 uri pointing to a constraints
JSON file.}

\item{\code{constraints}}{(sagemaker.model_monitor.Constraints or str): If provided alongside
statistics, these will be used for monitoring the endpoint. This can be a
sagemaker.model_monitor.Constraints object or an S3 uri pointing to a constraints
JSON file.}

\item{\code{schedule_cron_expression}}{(str): The cron expression that dictates the frequency that
this job runs at. See sagemaker.model_monitor.CronExpressionGenerator for valid
expressions.}

\item{\code{instance_count}}{(int): The number of instances to run
the jobs with.}

\item{\code{instance_type}}{(str): Type of EC2 instance to use for
the job, for example, 'ml.m5.xlarge'.}

\item{\code{volume_size_in_gb}}{(int): Size in GB of the EBS volume
to use for storing data during processing (default: 30).}

\item{\code{volume_kms_key}}{(str): A KMS key for the job's volume.}

\item{\code{output_kms_key}}{(str): The KMS key id for the job's outputs.}

\item{\code{max_runtime_in_seconds}}{(int): Timeout in seconds. After this amount of
time, Amazon SageMaker terminates the job regardless of its current status.
Default: 3600}

\item{\code{env}}{(dict): Environment variables to be passed to the job.}

\item{\code{network_config}}{(sagemaker.network.NetworkConfig): A NetworkConfig
object that configures network isolation, encryption of
inter-container traffic, security group IDs, and subnets.}

\item{\code{enable_cloudwatch_metrics}}{(bool): Whether to publish cloudwatch metrics as part of
the baselining or monitoring jobs.}

\item{\code{role}}{(str): An AWS IAM role name or ARN. The Amazon SageMaker jobs use this role.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-delete_monitoring_schedule"></a>}}
\if{latex}{\out{\hypertarget{method-delete_monitoring_schedule}{}}}
\subsection{Method \code{delete_monitoring_schedule()}}{
Deletes the monitoring schedule and its job definition.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DefaultModelMonitor$delete_monitoring_schedule()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-run_baseline"></a>}}
\if{latex}{\out{\hypertarget{method-run_baseline}{}}}
\subsection{Method \code{run_baseline()}}{
`run_baseline()` is only allowed for ModelMonitor objects. Please use suggest_baseline
              for DefaultModelMonitor objects, instead.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DefaultModelMonitor$run_baseline()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-attach"></a>}}
\if{latex}{\out{\hypertarget{method-attach}{}}}
\subsection{Method \code{attach()}}{
Sets this object's schedule name to point to the Amazon Sagemaker Monitoring Schedule
             name provided. This allows subsequent describe_schedule or list_executions calls to point
             to the given schedule.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DefaultModelMonitor$attach(monitor_schedule_name, sagemaker_session = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{monitor_schedule_name}}{(str): The name of the schedule to attach to.}

\item{\code{sagemaker_session}}{(sagemaker.session.Session): Session object which
manages interactions with Amazon SageMaker APIs and any other
AWS services needed. If not specified, one is created using
the default AWS configuration chain.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-latest_monitoring_statistics"></a>}}
\if{latex}{\out{\hypertarget{method-latest_monitoring_statistics}{}}}
\subsection{Method \code{latest_monitoring_statistics()}}{
Returns the sagemaker.model_monitor.Statistics generated by the latest monitoring
             execution.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DefaultModelMonitor$latest_monitoring_statistics()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
sagemaker.model_monitoring.Statistics: The Statistics object representing the file
             generated by the latest monitoring execution.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-latest_monitoring_constraint_violations"></a>}}
\if{latex}{\out{\hypertarget{method-latest_monitoring_constraint_violations}{}}}
\subsection{Method \code{latest_monitoring_constraint_violations()}}{
Returns the sagemaker.model_monitor.ConstraintViolations generated by the latest
             monitoring execution.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DefaultModelMonitor$latest_monitoring_constraint_violations()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
sagemaker.model_monitoring.ConstraintViolations: The ConstraintViolations object
             representing the file generated by the latest monitoring execution.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DefaultModelMonitor$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
