% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job.R
\name{.Job}
\alias{.Job}
\title{.Job Class}
\description{
Handle creating, starting and waiting for Amazon SageMaker jobs to
             finish.
}
\note{
This class shouldn't be directly instantiated.
             Subclasses must define a way to create, start and wait for an Amazon
             SageMaker job.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{sagemaker_session}}{Sagemaker Session Class}

\item{\code{job_name}}{name of job}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{name}}{Returns job name}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{.Job$new()}}
\item \href{#method-start_new}{\code{.Job$start_new()}}
\item \href{#method-wait}{\code{.Job$wait()}}
\item \href{#method-describe}{\code{.Job$describe()}}
\item \href{#method-stop}{\code{.Job$stop()}}
\item \href{#method-print}{\code{.Job$print()}}
\item \href{#method-clone}{\code{.Job$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Base class initializer. Subclasses which override ``__init__`` should
             invoke ``super()``
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{.Job$new(sagemaker_session = NULL, job_name = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sagemaker_session}}{(sagemaker.session.Session): Session object which
manages interactions with Amazon SageMaker APIs and any other
AWS services needed. If not specified, the estimator creates one
using the default AWS configuration chain.}

\item{\code{job_name}}{(str): Prefix for training job name}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-start_new"></a>}}
\if{latex}{\out{\hypertarget{method-start_new}{}}}
\subsection{Method \code{start_new()}}{
Create a new Amazon SageMaker job from the estimator.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{.Job$start_new(estimator, inputs)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{estimator}}{(sagemaker.estimator.EstimatorBase): Estimator object
created by the user.}

\item{\code{inputs}}{(str): Parameters used when called
:meth:`~sagemaker.estimator.EstimatorBase.fit`.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
sagemaker.job: Constructed object that captures all information
             about the started job.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-wait"></a>}}
\if{latex}{\out{\hypertarget{method-wait}{}}}
\subsection{Method \code{wait()}}{
Wait for the Amazon SageMaker job to finish.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{.Job$wait()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-describe"></a>}}
\if{latex}{\out{\hypertarget{method-describe}{}}}
\subsection{Method \code{describe()}}{
Describe the job.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{.Job$describe()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-stop"></a>}}
\if{latex}{\out{\hypertarget{method-stop}{}}}
\subsection{Method \code{stop()}}{
Stop the job.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{.Job$stop()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Printer.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{.Job$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{(ignored).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{.Job$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
