% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_monitoring.R
\name{EndpointInput}
\alias{EndpointInput}
\title{EndpointInput}
\description{
Accepts parameters that specify an endpoint input for a monitoring execution and provides
             a method to turn those parameters into a dictionary.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{EndpointInput$new()}}
\item \href{#method-to_request_list}{\code{EndpointInput$to_request_list()}}
\item \href{#method-print}{\code{EndpointInput$print()}}
\item \href{#method-clone}{\code{EndpointInput$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialize an ``EndpointInput`` instance. EndpointInput accepts parameters
             that specify an endpoint input for a monitoring job and provides a method
             to turn those parameters into a dictionary.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EndpointInput$new(
  endpoint_name,
  destination,
  s3_input_mode = c("File", "Pipe"),
  s3_data_distribution_type = c("FullyReplicated", "ShardedByS3Key")
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{endpoint_name}}{(str): The name of the endpoint.}

\item{\code{destination}}{(str): The destination of the input.}

\item{\code{s3_input_mode}}{(str): The S3 input mode. Can be one of: "File", "Pipe. Default: "File".}

\item{\code{s3_data_distribution_type}}{(str): The S3 Data Distribution Type. Can be one of:
"FullyReplicated", "ShardedByS3Key"}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-to_request_list"></a>}}
\if{latex}{\out{\hypertarget{method-to_request_list}{}}}
\subsection{Method \code{to_request_list()}}{
Generates a request dictionary using the parameters provided to the class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EndpointInput$to_request_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Printer.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EndpointInput$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{(ignored).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EndpointInput$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
