% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_monitor_model_monitoring.R
\name{EndpointOutput}
\alias{EndpointOutput}
\title{Accepts parameters that specify an S3 output for a monitoring job.}
\description{
It also provides a method to turn those parameters into a dictionary.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{EndpointOutput$new()}}
\item \href{#method-to_request_list}{\code{EndpointOutput$to_request_list()}}
\item \href{#method-print}{\code{EndpointOutput$print()}}
\item \href{#method-clone}{\code{EndpointOutput$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialize a ``MonitoringOutput`` instance. MonitoringOutput accepts parameters that
             specify an S3 output for a monitoring job and provides a method to turn
             those parameters into a dictionary.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EndpointOutput$new(source, destination = NULL, s3_upload_mode = "Continuous")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{source}}{(str): The source for the output.}

\item{\code{destination}}{(str): The destination of the output. Optional.
Default: s3://<default-session-bucket/schedule_name/output}

\item{\code{s3_upload_mode}}{(str): The S3 upload mode.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-to_request_list"></a>}}
\if{latex}{\out{\hypertarget{method-to_request_list}{}}}
\subsection{Method \code{to_request_list()}}{
Generates a request dictionary using the parameters provided to the class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EndpointOutput$to_request_list()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
dict: The request dictionary.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Printer.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EndpointOutput$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{(ignored).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EndpointOutput$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
