% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_entries.R
\name{Entity}
\alias{Entity}
\title{Base object for workflow entities.}
\description{
Entities must implement the to_request method.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-to_request}{\code{Entity$to_request()}}
\item \href{#method-clone}{\code{Entity$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-to_request"></a>}}
\if{latex}{\out{\hypertarget{method-to_request}{}}}
\subsection{Method \code{to_request()}}{
Get the request structure for workflow service calls.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Entity$to_request()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Entity$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
