% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics.R
\name{ExperimentAnalytics}
\alias{ExperimentAnalytics}
\title{ExperimentAnalytics class}
\description{
Fetch trial component data and make them accessible for analytics.
}
\section{Super class}{
\code{\link[R6sagemaker:AnalyticsMetricsBase]{R6sagemaker::AnalyticsMetricsBase}} -> \code{ExperimentAnalytics}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{MAX_TRIAL_COMPONENTS}}{class metadata}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{name}}{Name of the Experiment being analyzed}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ExperimentAnalytics$new()}}
\item \href{#method-clear_cache}{\code{ExperimentAnalytics$clear_cache()}}
\item \href{#method-print}{\code{ExperimentAnalytics$print()}}
\item \href{#method-clone}{\code{ExperimentAnalytics$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="R6sagemaker" data-topic="AnalyticsMetricsBase" data-id="dataframe">}\href{../../R6sagemaker/html/AnalyticsMetricsBase.html#method-dataframe}{\code{R6sagemaker::AnalyticsMetricsBase$dataframe()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="R6sagemaker" data-topic="AnalyticsMetricsBase" data-id="export_csv">}\href{../../R6sagemaker/html/AnalyticsMetricsBase.html#method-export_csv}{\code{R6sagemaker::AnalyticsMetricsBase$export_csv()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialize a ``ExperimentAnalytics`` instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExperimentAnalytics$new(
  experiment_name = NULL,
  search_expression = NULL,
  sort_by = NULL,
  sort_order = NULL,
  metric_names = NULL,
  parameter_names = NULL,
  sagemaker_session = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{experiment_name}}{(str, optional): Name of the experiment if you want to constrain the
search to only trial components belonging to an experiment.}

\item{\code{search_expression}}{(dict, optional): The search query to find the set of trial components
to use to populate the data frame.}

\item{\code{sort_by}}{(str, optional): The name of the resource property used to sort
the set of trial components.}

\item{\code{sort_order}}{(str optional): How trial components are ordered, valid values are Ascending
and Descending. The default is Descending.}

\item{\code{metric_names}}{(list, optional): string names of all the metrics to be shown in the
data frame. If not specified, all metrics will be shown of all trials.}

\item{\code{parameter_names}}{(list, optional): string names of the parameters to be shown in the
data frame. If not specified, all parameters will be shown of all trials.}

\item{\code{sagemaker_session}}{(sagemaker.session.Session): Session object which manages interactions
with Amazon SageMaker APIs and any other AWS services needed. If not specified,
one is created using the default AWS configuration chain.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clear_cache"></a>}}
\if{latex}{\out{\hypertarget{method-clear_cache}{}}}
\subsection{Method \code{clear_cache()}}{
Clear the object of all local caches of API methods.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExperimentAnalytics$clear_cache()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Printer.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExperimentAnalytics$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{(ignored).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExperimentAnalytics$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
