% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clarify.R
\name{ExplainabilityConfig}
\alias{ExplainabilityConfig}
\title{ExplainabilityConfig Class}
\description{
Abstract config class to configure an explainability method.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-get_explainability_config}{\code{ExplainabilityConfig$get_explainability_config()}}
\item \href{#method-print}{\code{ExplainabilityConfig$print()}}
\item \href{#method-clone}{\code{ExplainabilityConfig$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_explainability_config"></a>}}
\if{latex}{\out{\hypertarget{method-get_explainability_config}{}}}
\subsection{Method \code{get_explainability_config()}}{
Returns config.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExplainabilityConfig$get_explainability_config()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Printer.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExplainabilityConfig$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{(ignored).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExplainabilityConfig$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
