% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.R
\name{FileSystemInput}
\alias{FileSystemInput}
\title{Amazon SageMaker channel configurations for file system data sources.}
\description{
Amazon SageMaker channel configurations for file system data sources.

Amazon SageMaker channel configurations for file system data sources.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{config}}{(dict[str, dict])\cr
A Sagemaker File System ``DataSource``.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{FileSystemInput$new()}}
\item \href{#method-format}{\code{FileSystemInput$format()}}
\item \href{#method-clone}{\code{FileSystemInput$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new file system input used by an SageMaker training job.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FileSystemInput$new(
  file_system_id,
  file_system_type = c("FSxLustre", "EFS"),
  directory_path,
  file_system_access_mode = c("ro", "rw"),
  content_type = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{file_system_id}}{(str): An Amazon file system ID starting with 'fs-'.}

\item{\code{file_system_type}}{(str): The type of file system used for the input.
Valid values: 'EFS', 'FSxLustre'.}

\item{\code{directory_path}}{(str): Absolute or normalized path to the root directory (mount point) in
the file system.
Reference: https://docs.aws.amazon.com/efs/latest/ug/mounting-fs.html and
https://docs.aws.amazon.com/fsx/latest/LustreGuide/mount-fs-auto-mount-onreboot.html}

\item{\code{file_system_access_mode}}{(str): Permissions for read and write.
Valid values: 'ro' or 'rw'. Defaults to 'ro'.}

\item{\code{content_type}}{:}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-format"></a>}}
\if{latex}{\out{\hypertarget{method-format}{}}}
\subsection{Method \code{format()}}{
format class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FileSystemInput$format()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FileSystemInput$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
