% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amazon_estimator.R
\name{FileSystemRecordSet}
\alias{FileSystemRecordSet}
\title{FileSystemRecordSet Class}
\description{
Amazon SageMaker channel configuration for a file system data source
             for Amazon algorithms.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{feature_dim}}{The dimensionality of "values" arrays in the Record features}

\item{\code{num_records}}{The number of records in the set}

\item{\code{channel}}{The SageMaker Training Job channel this RecordSet should be bound to}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{FileSystemRecordSet$new()}}
\item \href{#method-print}{\code{FileSystemRecordSet$print()}}
\item \href{#method-data_channel}{\code{FileSystemRecordSet$data_channel()}}
\item \href{#method-clone}{\code{FileSystemRecordSet$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialize a ``FileSystemRecordSet`` object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FileSystemRecordSet$new(
  file_system_id,
  file_system_type,
  directory_path,
  num_records,
  feature_dim,
  file_system_access_mode = "ro",
  channel = "train"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{file_system_id}}{(str): An Amazon file system ID starting with 'fs-'.}

\item{\code{file_system_type}}{(str): The type of file system used for the input.
Valid values: 'EFS', 'FSxLustre'.}

\item{\code{directory_path}}{(str): Absolute or normalized path to the root directory (mount point) in
the file system. Reference:
https://docs.aws.amazon.com/efs/latest/ug/mounting-fs.html and
https://docs.aws.amazon.com/efs/latest/ug/wt1-test.html}

\item{\code{num_records}}{(int): The number of records in the set.}

\item{\code{feature_dim}}{(int): The dimensionality of "values" arrays in the Record features,
and label (if each Record is labeled).}

\item{\code{file_system_access_mode}}{(str): Permissions for read and write.
Valid values: 'ro' or 'rw'. Defaults to 'ro'.}

\item{\code{channel}}{(str): The SageMaker Training Job channel this RecordSet should be bound to}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Return an unambiguous representation of this RecordSet


Return an unambiguous representation of this RecordSet
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FileSystemRecordSet$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-data_channel"></a>}}
\if{latex}{\out{\hypertarget{method-data_channel}{}}}
\subsection{Method \code{data_channel()}}{
Return a dictionary to represent the training data in a channel for use with ``fit()``
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FileSystemRecordSet$data_channel()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FileSystemRecordSet$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
