% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debugger_framework_profile.R
\name{FrameworkProfile}
\alias{FrameworkProfile}
\title{Sets up the profiling configuration for framework metrics.}
\description{
Validates user inputs and fills in default values if no input is provided.
             There are three main profiling options to choose from:
             :class:`~sagemaker.debugger.metrics_config.DetailedProfilingConfig`,
             :class:`~sagemaker.debugger.metrics_config.DataloaderProfilingConfig`, and
             :class:`~sagemaker.debugger.metrics_config.PythonProfilingConfig`.
             The following list shows available scenarios of configuring the profiling options.
             \enumerate{
             \item None of the profiling configuration, step range, or time range is specified.
             SageMaker Debugger activates framework profiling based on the default settings
             of each profiling option.
             \item Target step or time range is specified to
             this :class:`~sagemaker.debugger.metrics_config.FrameworkProfile` class.
             The requested target step or time range setting propagates to all of
             the framework profiling options.
             For example, if you configure this class as following, all of the profiling options
             profiles the 6th step.
             \item Individual profiling configurations are specified through
             the ``*_profiling_config`` parameters.
             SageMaker Debugger profiles framework metrics only for the specified profiling configurations.
             For example, if the :class:`~sagemaker.debugger.metrics_config.DetailedProfilingConfig` class
             is configured but not the other profiling options, Debugger only profiles based on the settings
             specified to the
             :class:`~sagemaker.debugger.metrics_config.DetailedProfilingConfig` class.
             For example, the following example shows a profiling configuration to perform
             detailed profiling at step 10, data loader profiling at step 9 and 10,
             and Python profiling at step 12.
             If the individual profiling configurations are specified in addition to
             the step or time range,
             SageMaker Debugger prioritizes the individual profiling configurations and ignores
             the step or time range. For example, in the following code,
             the ``start_step=1`` and ``num_steps=10`` will be ignored.
             }
}
\examples{
library(R6sagemaker)
profiler_config=ProfilerConfig$new(
       framework_profile_params=FrameworkProfile$new(
       start_step=1,
       num_steps=10,
       detailed_profiling_config=DetailedProfilingConfig$new(start_step=10, num_steps=1),
       dataloader_profiling_config=DataloaderProfilingConfig$new(start_step=9, num_steps=2),
       python_profiling_config=PythonProfilingConfig$new(start_step=12, num_steps=1)
     )
   )
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-intialize}{\code{FrameworkProfile$intialize()}}
\item \href{#method-format}{\code{FrameworkProfile$format()}}
\item \href{#method-clone}{\code{FrameworkProfile$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-intialize"></a>}}
\if{latex}{\out{\hypertarget{method-intialize}{}}}
\subsection{Method \code{intialize()}}{
Initialize the FrameworkProfile class object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FrameworkProfile$intialize(
  local_path = BASE_FOLDER_DEFAULT,
  file_max_size = MAX_FILE_SIZE_DEFAULT,
  file_close_interval = CLOSE_FILE_INTERVAL_DEFAULT,
  file_open_fail_threshold = FILE_OPEN_FAIL_THRESHOLD_DEFAULT,
  detailed_profiling_config = NULL,
  dataloader_profiling_config = NULL,
  python_profiling_config = NULL,
  horovod_profiling_config = NULL,
  smdataparallel_profiling_config = NULL,
  start_step = NULL,
  num_steps = NULL,
  start_unix_time = NULL,
  duration = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{local_path}}{(str):}

\item{\code{file_max_size}}{(int):}

\item{\code{file_close_interval}}{(int):}

\item{\code{file_open_fail_threshold}}{(int):}

\item{\code{detailed_profiling_config}}{(DetailedProfilingConfig): The configuration for detailed
profiling. Configure it using the
:class:`~sagemaker.debugger.metrics_config.DetailedProfilingConfig` class.
Pass ``DetailedProfilingConfig()`` to use the default configuration.}

\item{\code{dataloader_profiling_config}}{(DataloaderProfilingConfig): The configuration for
dataloader metrics profiling. Configure it using the
:class:`~sagemaker.debugger.metrics_config.DataloaderProfilingConfig` class.
Pass ``DataloaderProfilingConfig()`` to use the default configuration.}

\item{\code{python_profiling_config}}{(PythonProfilingConfig): The configuration for stats
collected by the Python profiler (cProfile or Pyinstrument).
Configure it using the
:class:`~sagemaker.debugger.metrics_config.PythonProfilingConfig` class.
Pass ``PythonProfilingConfig()`` to use the default configuration.}

\item{\code{horovod_profiling_config}}{:}

\item{\code{smdataparallel_profiling_config}}{:}

\item{\code{start_step}}{(int): The step at which to start profiling.}

\item{\code{num_steps}}{(int): The number of steps to profile.}

\item{\code{start_unix_time}}{(int): The Unix time at which to start profiling.}

\item{\code{duration}}{(float): The duration in seconds to profile.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-format"></a>}}
\if{latex}{\out{\hypertarget{method-format}{}}}
\subsection{Method \code{format()}}{
format class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FrameworkProfile$format(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{(ignored).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FrameworkProfile$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
