% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amazon_estimator.R
\name{get_ecr_image_uri}
\alias{get_ecr_image_uri}
\title{Return algorithm image URI for the given ecr repository}
\usage{
get_ecr_image_uri(repo_name, repo_version = NULL, sagemaker_session = NULL)
}
\arguments{
\item{repo_version}{(str): Version fo repo to call}

\item{sagemaker_session}{(sagemaker.session.Session): Session object which
manages interactions with Amazon SageMaker APIs and any other
AWS services needed. If not specified, the estimator creates one
using the default AWS configuration chain.}

\item{region_name}{(str): The region name for the account.}
}
\description{
Decided to help R users integrate "bring your own R models" in sagemaker
}
