% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session.R
\name{get_execution_role}
\alias{get_execution_role}
\title{Return the role ARN whose credentials are used to call the API.}
\usage{
get_execution_role(sagemaker_session = NULL)
}
\arguments{
\item{sagemaker_session}{(Session): Current sagemaker session}
}
\value{
(str): The role ARN
}
\description{
Return the role ARN whose credentials are used to call the API.
}
