% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amazon_estimator.R
\name{get_image_uri}
\alias{get_image_uri}
\title{Return algorithm image URI for the given AWS region, repository name, and
repository version}
\usage{
get_image_uri(region_name, repo_name, repo_version = "1.0-1")
}
\arguments{
\item{region_name}{(str): The region name for the account.}

\item{repo_name}{(str): The repo name for the account}

\item{repo_version}{(str): Version fo repo to call}
}
\description{
Return algorithm image URI for the given AWS region, repository name, and
repository version
}
