% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xgboost_estimator.R
\name{get_xgboost_image_uri}
\alias{get_xgboost_image_uri}
\title{Get XGBoost framework image URI}
\usage{
get_xgboost_image_uri(region, framework_version, py_version = "py3")
}
\arguments{
\item{region}{(str): The aws region in which docker image is stored.}

\item{py_version}{(str): python version}

\item{frame_version}{(str):}
}
\description{
Get XGBoost framework image URI
}
