% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amazon_hyperparameter.R
\name{Hyperparameter}
\alias{Hyperparameter}
\title{Hyperparameter Class}
\description{
An algorithm hyperparameter with optional validation. Implemented as a
             python descriptor object.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{validation}}{validation function}

\item{\code{validation_message}}{validation message}

\item{\code{name}}{name of hyperparameter validate}

\item{\code{data_type}}{function to convert data type}

\item{\code{obj}}{parent class to alter}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{descriptor}}{active binding that mimic's python descriptor class}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Hyperparameter$new()}}
\item \href{#method-validate}{\code{Hyperparameter$validate()}}
\item \href{#method-serialize_all}{\code{Hyperparameter$serialize_all()}}
\item \href{#method-clone}{\code{Hyperparameter$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Hyperparameter$new(
  name,
  validate = function(x) TRUE,
  validation_message = "",
  data_type = as.character,
  obj
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{(str): The name of this hyperparameter validate}

\item{\code{validate}}{(callable[object]->[bool]): A validation function or list of validation
functions.
Each function validates an object and returns False if the object
value is invalid for this hyperparameter.}

\item{\code{validation_message}}{(str): A usage guide to display on validation
failure.}

\item{\code{data_type}}{: function to convert data types}

\item{\code{obj}}{(R6Class): R6Class for descriptor class to modify}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
Validate value
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Hyperparameter$validate(value = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{: values to be validated}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-serialize_all"></a>}}
\if{latex}{\out{\hypertarget{method-serialize_all}{}}}
\subsection{Method \code{serialize_all()}}{
Return all non-None ``hyperparameter`` values on ``obj`` as a
             ``dict[str,str].``
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Hyperparameter$serialize_all(obj)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{obj}}{: R object to be serialized}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Hyperparameter$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
