% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics.R
\name{HyperparameterTuningJobAnalytics}
\alias{HyperparameterTuningJobAnalytics}
\title{HyperparameterTuningJobAnalytics Class}
\description{
Fetch results about a hyperparameter tuning job and make them accessible
             for analytics.
}
\section{Super class}{
\code{\link[R6sagemaker:AnalyticsMetricsBase]{R6sagemaker::AnalyticsMetricsBase}} -> \code{HyperparameterTuningJobAnalytics}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{name}}{Name of the HyperparameterTuningJob being analyzed}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{HyperparameterTuningJobAnalytics$new()}}
\item \href{#method-clear_cache}{\code{HyperparameterTuningJobAnalytics$clear_cache()}}
\item \href{#method-tuning_ranges}{\code{HyperparameterTuningJobAnalytics$tuning_ranges()}}
\item \href{#method-description}{\code{HyperparameterTuningJobAnalytics$description()}}
\item \href{#method-training_job_summaries}{\code{HyperparameterTuningJobAnalytics$training_job_summaries()}}
\item \href{#method-clone}{\code{HyperparameterTuningJobAnalytics$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="R6sagemaker" data-topic="AnalyticsMetricsBase" data-id="dataframe">}\href{../../R6sagemaker/html/AnalyticsMetricsBase.html#method-dataframe}{\code{R6sagemaker::AnalyticsMetricsBase$dataframe()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="R6sagemaker" data-topic="AnalyticsMetricsBase" data-id="export_csv">}\href{../../R6sagemaker/html/AnalyticsMetricsBase.html#method-export_csv}{\code{R6sagemaker::AnalyticsMetricsBase$export_csv()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="R6sagemaker" data-topic="AnalyticsMetricsBase" data-id="print">}\href{../../R6sagemaker/html/AnalyticsMetricsBase.html#method-print}{\code{R6sagemaker::AnalyticsMetricsBase$print()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialize a ``HyperparameterTuningJobAnalytics`` instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HyperparameterTuningJobAnalytics$new(
  hyperparameter_tuning_job_name,
  sagemaker_session = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{hyperparameter_tuning_job_name}}{(str): name of the
HyperparameterTuningJob to analyze.}

\item{\code{sagemaker_session}}{(sagemaker.session.Session): Session object which
manages interactions with Amazon SageMaker APIs and any other
AWS services needed. If not specified, one is created using the
default AWS configuration chain.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clear_cache"></a>}}
\if{latex}{\out{\hypertarget{method-clear_cache}{}}}
\subsection{Method \code{clear_cache()}}{
Clear the object of all local caches of API methods.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HyperparameterTuningJobAnalytics$clear_cache()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-tuning_ranges"></a>}}
\if{latex}{\out{\hypertarget{method-tuning_ranges}{}}}
\subsection{Method \code{tuning_ranges()}}{
A dictionary describing the ranges of all tuned hyperparameters. The
             keys are the names of the hyperparameter, and the values are the ranges.
             The output can take one of two forms:
             * If the 'TrainingJobDefinition' field is present in the job description, the output
             is a dictionary constructed from 'ParameterRanges' in
             'HyperParameterTuningJobConfig' of the job description. The keys are the
             parameter names, while the values are the parameter ranges.
             Example:
             >>> {
             >>>     "eta": {"MaxValue": "1", "MinValue": "0", "Name": "eta"},
             >>>     "gamma": {"MaxValue": "10", "MinValue": "0", "Name": "gamma"},
             >>>     "iterations": {"MaxValue": "100", "MinValue": "50", "Name": "iterations"},
             >>>     "num_layers": {"MaxValue": "30", "MinValue": "5", "Name": "num_layers"},
             >>> }
             * If the 'TrainingJobDefinitions' field (list) is present in the job description,
             the output is a dictionary with keys as the 'DefinitionName' values from
             all items in 'TrainingJobDefinitions', and each value would be a dictionary
             constructed from 'HyperParameterRanges' in each item in 'TrainingJobDefinitions'
             in the same format as above
             Example:
             >>> {
             >>>     "estimator_1": {
             >>>         "eta": {"MaxValue": "1", "MinValue": "0", "Name": "eta"},
             >>>         "gamma": {"MaxValue": "10", "MinValue": "0", "Name": "gamma"},
             >>>     },
             >>>     "estimator_2": {
             >>>         "framework": {"Values": ["TF", "MXNet"], "Name": "framework"},
             >>>         "gamma": {"MaxValue": "1.0", "MinValue": "0.2", "Name": "gamma"}
             >>>     }
             >>> }
             For more details about the 'TrainingJobDefinition' and 'TrainingJobDefinitions' fields
             in job description, see
             https://botocore.readthedocs.io/en/latest/reference/services/sagemaker.html#SageMaker.Client.create_hyper_parameter_tuning_job
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HyperparameterTuningJobAnalytics$tuning_ranges()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-description"></a>}}
\if{latex}{\out{\hypertarget{method-description}{}}}
\subsection{Method \code{description()}}{
Call ``DescribeHyperParameterTuningJob`` for the hyperparameter
             tuning job.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HyperparameterTuningJobAnalytics$description(force_refresh = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{force_refresh}}{(bool): Set to True to fetch the latest data from
SageMaker API.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
dict: The Amazon SageMaker response for
             ``DescribeHyperParameterTuningJob``.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-training_job_summaries"></a>}}
\if{latex}{\out{\hypertarget{method-training_job_summaries}{}}}
\subsection{Method \code{training_job_summaries()}}{
A (paginated) list of everything from
             ``ListTrainingJobsForTuningJob``.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HyperparameterTuningJobAnalytics$training_job_summaries(force_refresh = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{force_refresh}}{(bool): Set to True to fetch the latest data from
SageMaker API.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
dict: The Amazon SageMaker response for
             ``ListTrainingJobsForTuningJob``.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HyperparameterTuningJobAnalytics$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
