% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter.R
\name{IntegerParameter}
\alias{IntegerParameter}
\title{IntegerParameter Class}
\description{
A class for representing hyperparameters that have an integer range of possible values.
}
\section{Super class}{
\code{\link[R6sagemaker:ParameterRange]{R6sagemaker::ParameterRange}} -> \code{IntegerParameter}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{.name}}{Helps to categorise Class}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{IntegerParameter$new()}}
\item \href{#method-cast_to_type}{\code{IntegerParameter$cast_to_type()}}
\item \href{#method-print}{\code{IntegerParameter$print()}}
\item \href{#method-clone}{\code{IntegerParameter$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="R6sagemaker" data-topic="ParameterRange" data-id="as_tuning_range">}\href{../../R6sagemaker/html/ParameterRange.html#method-as_tuning_range}{\code{R6sagemaker::ParameterRange$as_tuning_range()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="R6sagemaker" data-topic="ParameterRange" data-id="is_valid">}\href{../../R6sagemaker/html/ParameterRange.html#method-is_valid}{\code{R6sagemaker::ParameterRange$is_valid()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialize a IntegerParameter
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IntegerParameter$new(
  min_value,
  max_value,
  scaling_type = c("Auto", "Linear", "Logarithmic", "ReverseLogarithmic")
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{min_value}}{(int): The minimum value for the range.}

\item{\code{max_value}}{(int): The maximum value for the range.}

\item{\code{scaling_type}}{(str): The scale used for searching the range during
tuning (default: 'Auto'). Valid values: 'Auto', 'Linear',
Logarithmic' and 'ReverseLogarithmic'.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-cast_to_type"></a>}}
\if{latex}{\out{\hypertarget{method-cast_to_type}{}}}
\subsection{Method \code{cast_to_type()}}{
cast value to integer
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IntegerParameter$cast_to_type(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{The value to be verified.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Printer.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IntegerParameter$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{(ignored).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IntegerParameter$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
