% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debugger_utils.R
\name{is_valid_regex}
\alias{is_valid_regex}
\title{Helper function to determine whether the provided regex is valid.}
\usage{
is_valid_regex(regex)
}
\arguments{
\item{regex}{(str): The user provided regex.}
}
\value{
bool: Indicates whether the provided regex was valid or not.


}
\description{
Helper function to determine whether the provided regex is valid.
}
