% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictor.R
\docType{data}
\name{json_serializer}
\alias{json_serializer}
\title{S3 method to call JsonSerializer class}
\format{
An object of class \code{JsonSerializer} (inherits from \code{Serializer}, \code{R6}) of length 6.
}
\usage{
json_serializer
}
\description{
S3 method to call JsonSerializer class
}
\keyword{datasets}
