% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serializers.R
\name{JSONLinesSerializer}
\alias{JSONLinesSerializer}
\title{JSONLinesSerializer Class}
\description{
Serialize data to a JSON Lines formatted string.
}
\section{Super classes}{
\code{\link[R6sagemaker:BaseSerializer]{R6sagemaker::BaseSerializer}} -> \code{\link[R6sagemaker:SimpleBaseSerializer]{R6sagemaker::SimpleBaseSerializer}} -> \code{IdentitySerializer}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{JSONLinesSerializer$new()}}
\item \href{#method-serialize}{\code{JSONLinesSerializer$serialize()}}
\item \href{#method-clone}{\code{JSONLinesSerializer$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="R6sagemaker" data-topic="BaseSerializer" data-id="format">}\href{../../R6sagemaker/html/BaseSerializer.html#method-format}{\code{R6sagemaker::BaseSerializer$format()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialize a ``JSONLinesSerializer`` instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JSONLinesSerializer$new(content_type = "application/jsonlines")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{content_type}}{(str): The MIME type to signal to the inference endpoint when sending
request data (default: "application/jsonlines").}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-serialize"></a>}}
\if{latex}{\out{\hypertarget{method-serialize}{}}}
\subsection{Method \code{serialize()}}{
Serialize data of various formats to a JSON Lines formatted string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JSONLinesSerializer$serialize(data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{(object): Data to be serialized. The data can be a string,
iterable of JSON serializable objects, or a file-like object.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
str: The data serialized as a string containing newline-separated
             JSON values.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JSONLinesSerializer$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
