% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amazon_knn.R
\name{KNNPredictor}
\alias{KNNPredictor}
\title{Performs classification or regression prediction from input vectors.}
\description{
The implementation of
             :meth:`~sagemaker.predictor.Predictor.predict` in this
             `Predictor` requires a numpy ``ndarray`` as input. The array should
             contain the same number of columns as the feature-dimension of the data used
             to fit the model this Predictor performs inference on.
             :func:`predict` returns a list of
             :class:`~sagemaker.amazon.record_pb2.Record` objects, one for each row in
             the input ``ndarray``. The prediction is stored in the ``"predicted_label"``
             key of the ``Record.label`` field.
}
\section{Super class}{
\code{\link[R6sagemaker:Predictor]{R6sagemaker::Predictor}} -> \code{KNNPredictor}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{KNNPredictor$new()}}
\item \href{#method-clone}{\code{KNNPredictor$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="R6sagemaker" data-topic="Predictor" data-id="delete_endpoint">}\href{../../R6sagemaker/html/Predictor.html#method-delete_endpoint}{\code{R6sagemaker::Predictor$delete_endpoint()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="R6sagemaker" data-topic="Predictor" data-id="delete_model">}\href{../../R6sagemaker/html/Predictor.html#method-delete_model}{\code{R6sagemaker::Predictor$delete_model()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="R6sagemaker" data-topic="Predictor" data-id="disable_data_capture">}\href{../../R6sagemaker/html/Predictor.html#method-disable_data_capture}{\code{R6sagemaker::Predictor$disable_data_capture()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="R6sagemaker" data-topic="Predictor" data-id="enable_data_capture">}\href{../../R6sagemaker/html/Predictor.html#method-enable_data_capture}{\code{R6sagemaker::Predictor$enable_data_capture()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="R6sagemaker" data-topic="Predictor" data-id="list_monitor">}\href{../../R6sagemaker/html/Predictor.html#method-list_monitor}{\code{R6sagemaker::Predictor$list_monitor()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="R6sagemaker" data-topic="Predictor" data-id="predict">}\href{../../R6sagemaker/html/Predictor.html#method-predict}{\code{R6sagemaker::Predictor$predict()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="R6sagemaker" data-topic="Predictor" data-id="print">}\href{../../R6sagemaker/html/Predictor.html#method-print}{\code{R6sagemaker::Predictor$print()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="R6sagemaker" data-topic="Predictor" data-id="update_data_capture_config">}\href{../../R6sagemaker/html/Predictor.html#method-update_data_capture_config}{\code{R6sagemaker::Predictor$update_data_capture_config()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialize KNNPredictor class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KNNPredictor$new(endpoint_name, sagemaker_session = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{endpoint_name}}{(str): Name of the Amazon SageMaker endpoint to which
requests are sent.}

\item{\code{sagemaker_session}}{(sagemaker.session.Session): A SageMaker Session
object, used for SageMaker interactions (default: None). If not
specified, one is created using the default AWS configuration
chain.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KNNPredictor$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
