% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_metrics.R
\name{MetricsSource}
\alias{MetricsSource}
\title{MetricsSource Class}
\description{
Accepts metrics source parameters for conversion to request dict.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{MetricsSource$new()}}
\item \href{#method-to_request_list}{\code{MetricsSource$to_request_list()}}
\item \href{#method-format}{\code{MetricsSource$format()}}
\item \href{#method-clone}{\code{MetricsSource$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialize a ``MetricsSource`` instance and turn parameters into dict.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MetricsSource$new(content_type, s3_uri, content_digest = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{content_type}}{(str):}

\item{\code{s3_uri}}{(str):}

\item{\code{content_digest}}{(str):}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-to_request_list"></a>}}
\if{latex}{\out{\hypertarget{method-to_request_list}{}}}
\subsection{Method \code{to_request_list()}}{
Generates a request dictionary using the parameters provided to the class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MetricsSource$to_request_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-format"></a>}}
\if{latex}{\out{\hypertarget{method-format}{}}}
\subsection{Method \code{format()}}{
format class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MetricsSource$format()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MetricsSource$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
