% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_monitor_clarify_model_monitoring.R
\name{ModelExplainabilityMonitor}
\alias{ModelExplainabilityMonitor}
\title{Amazon SageMaker model monitor to monitor feature attribution of an endpoint.}
\description{
Please see the `initiliaze` method of its base class for how to instantiate it.
}
\section{Super classes}{
\code{\link[R6sagemaker:ModelMonitor]{R6sagemaker::ModelMonitor}} -> \code{\link[R6sagemaker:ClarifyModelMonitor]{R6sagemaker::ClarifyModelMonitor}} -> \code{ModelExplainabilityMonitor}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{JOB_DEFINITION_BASE_NAME}}{Model definition base name}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-monitoring_type}{\code{ModelExplainabilityMonitor$monitoring_type()}}
\item \href{#method-suggest_baseline}{\code{ModelExplainabilityMonitor$suggest_baseline()}}
\item \href{#method-create_monitoring_schedule}{\code{ModelExplainabilityMonitor$create_monitoring_schedule()}}
\item \href{#method-update_monitoring_schedule}{\code{ModelExplainabilityMonitor$update_monitoring_schedule()}}
\item \href{#method-delete_monitoring_schedule}{\code{ModelExplainabilityMonitor$delete_monitoring_schedule()}}
\item \href{#method-attach}{\code{ModelExplainabilityMonitor$attach()}}
\item \href{#method-clone}{\code{ModelExplainabilityMonitor$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="R6sagemaker" data-topic="ModelMonitor" data-id="baseline_statistics">}\href{../../R6sagemaker/html/ModelMonitor.html#method-baseline_statistics}{\code{R6sagemaker::ModelMonitor$baseline_statistics()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="R6sagemaker" data-topic="ModelMonitor" data-id="describe_latest_baselining_job">}\href{../../R6sagemaker/html/ModelMonitor.html#method-describe_latest_baselining_job}{\code{R6sagemaker::ModelMonitor$describe_latest_baselining_job()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="R6sagemaker" data-topic="ModelMonitor" data-id="describe_schedule">}\href{../../R6sagemaker/html/ModelMonitor.html#method-describe_schedule}{\code{R6sagemaker::ModelMonitor$describe_schedule()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="R6sagemaker" data-topic="ModelMonitor" data-id="latest_monitoring_constraint_violations">}\href{../../R6sagemaker/html/ModelMonitor.html#method-latest_monitoring_constraint_violations}{\code{R6sagemaker::ModelMonitor$latest_monitoring_constraint_violations()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="R6sagemaker" data-topic="ModelMonitor" data-id="latest_monitoring_statistics">}\href{../../R6sagemaker/html/ModelMonitor.html#method-latest_monitoring_statistics}{\code{R6sagemaker::ModelMonitor$latest_monitoring_statistics()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="R6sagemaker" data-topic="ModelMonitor" data-id="print">}\href{../../R6sagemaker/html/ModelMonitor.html#method-print}{\code{R6sagemaker::ModelMonitor$print()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="R6sagemaker" data-topic="ModelMonitor" data-id="start_monitoring_schedule">}\href{../../R6sagemaker/html/ModelMonitor.html#method-start_monitoring_schedule}{\code{R6sagemaker::ModelMonitor$start_monitoring_schedule()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="R6sagemaker" data-topic="ModelMonitor" data-id="stop_monitoring_schedule">}\href{../../R6sagemaker/html/ModelMonitor.html#method-stop_monitoring_schedule}{\code{R6sagemaker::ModelMonitor$stop_monitoring_schedule()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="R6sagemaker" data-topic="ModelMonitor" data-id="suggested_constraints">}\href{../../R6sagemaker/html/ModelMonitor.html#method-suggested_constraints}{\code{R6sagemaker::ModelMonitor$suggested_constraints()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="R6sagemaker" data-topic="ClarifyModelMonitor" data-id="initialize">}\href{../../R6sagemaker/html/ClarifyModelMonitor.html#method-initialize}{\code{R6sagemaker::ClarifyModelMonitor$initialize()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="R6sagemaker" data-topic="ClarifyModelMonitor" data-id="latest_monitorying_statistics">}\href{../../R6sagemaker/html/ClarifyModelMonitor.html#method-latest_monitorying_statistics}{\code{R6sagemaker::ClarifyModelMonitor$latest_monitorying_statistics()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="R6sagemaker" data-topic="ClarifyModelMonitor" data-id="list_executions">}\href{../../R6sagemaker/html/ClarifyModelMonitor.html#method-list_executions}{\code{R6sagemaker::ClarifyModelMonitor$list_executions()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="R6sagemaker" data-topic="ClarifyModelMonitor" data-id="run_baseline">}\href{../../R6sagemaker/html/ClarifyModelMonitor.html#method-run_baseline}{\code{R6sagemaker::ClarifyModelMonitor$run_baseline()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-monitoring_type"></a>}}
\if{latex}{\out{\hypertarget{method-monitoring_type}{}}}
\subsection{Method \code{monitoring_type()}}{
Type of the monitoring job.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelExplainabilityMonitor$monitoring_type()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-suggest_baseline"></a>}}
\if{latex}{\out{\hypertarget{method-suggest_baseline}{}}}
\subsection{Method \code{suggest_baseline()}}{
Suggest baselines for use with Amazon SageMaker Model Monitoring Schedules.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelExplainabilityMonitor$suggest_baseline(
  data_config,
  explainability_config,
  model_config,
  model_scores = NULL,
  wait = FALSE,
  logs = FALSE,
  job_name = NULL,
  kms_key = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data_config}}{(:class:`~sagemaker.clarify.DataConfig`): Config of the input/output data.}

\item{\code{explainability_config}}{(:class:`~sagemaker.clarify.ExplainabilityConfig`): Config of the
specific explainability method. Currently, only SHAP is supported.}

\item{\code{model_config}}{(:class:`~sagemaker.clarify.ModelConfig`): Config of the model and its
endpoint to be created.}

\item{\code{model_scores}}{:  Index or JSONPath location in the model output for the predicted scores
to be explained. This is not required if the model output is a single score.}

\item{\code{wait}}{(bool): Whether the call should wait until the job completes (default: False).}

\item{\code{logs}}{(bool): Whether to show the logs produced by the job.
Only meaningful when wait is True (default: False).}

\item{\code{job_name}}{(str): Processing job name. If not specified, the processor generates
a default job name, based on the image name and current timestamp.}

\item{\code{kms_key}}{(str): The ARN of the KMS key that is used to encrypt the
user code file (default: None).}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
sagemaker.processing.ProcessingJob: The ProcessingJob object representing the
             baselining job.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-create_monitoring_schedule"></a>}}
\if{latex}{\out{\hypertarget{method-create_monitoring_schedule}{}}}
\subsection{Method \code{create_monitoring_schedule()}}{
Creates a monitoring schedule.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelExplainabilityMonitor$create_monitoring_schedule(
  endpoint_input,
  analysis_config = NULL,
  output_s3_uri = NULL,
  constraints = NULL,
  monitor_schedule_name = NULL,
  schedule_cron_expression = NULL,
  enable_cloudwatch_metrics = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{endpoint_input}}{(str or sagemaker.model_monitor.EndpointInput): The endpoint to monitor.
This can either be the endpoint name or an EndpointInput.}

\item{\code{analysis_config}}{(str or ExplainabilityAnalysisConfig): URI to the analysis_config for
the explainability job. If it is None then configuration of the latest baselining
job will be reused, but if no baselining job then fail the call.}

\item{\code{output_s3_uri}}{(str): S3 destination of the constraint_violations and analysis result.
Default: "s3://<default_session_bucket>/<job_name>/output"}

\item{\code{constraints}}{(sagemaker.model_monitor.Constraints or str): If provided it will be used
for monitoring the endpoint. It can be a Constraints object or an S3 uri pointing
to a constraints JSON file.}

\item{\code{monitor_schedule_name}}{(str): Schedule name. If not specified, the processor generates
a default job name, based on the image name and current timestamp.}

\item{\code{schedule_cron_expression}}{(str): The cron expression that dictates the frequency that
this job run. See sagemaker.model_monitor.CronExpressionGenerator for valid
expressions. Default: Daily.}

\item{\code{enable_cloudwatch_metrics}}{(bool): Whether to publish cloudwatch metrics as part of
the baselining or monitoring jobs.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update_monitoring_schedule"></a>}}
\if{latex}{\out{\hypertarget{method-update_monitoring_schedule}{}}}
\subsection{Method \code{update_monitoring_schedule()}}{
Updates the existing monitoring schedule.
             If more options than schedule_cron_expression are to be updated, a new job definition will
             be created to hold them. The old job definition will not be deleted.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelExplainabilityMonitor$update_monitoring_schedule(
  endpoint_input = NULL,
  analysis_config = NULL,
  output_s3_uri = NULL,
  constraints = NULL,
  schedule_cron_expression = NULL,
  enable_cloudwatch_metrics = NULL,
  role = NULL,
  instance_count = NULL,
  instance_type = NULL,
  volume_size_in_gb = NULL,
  volume_kms_key = NULL,
  output_kms_key = NULL,
  max_runtime_in_seconds = NULL,
  env = NULL,
  network_config = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{endpoint_input}}{(str or sagemaker.model_monitor.EndpointInput): The endpoint to monitor.
This can either be the endpoint name or an EndpointInput.}

\item{\code{analysis_config}}{(str or BiasAnalysisConfig): URI to analysis_config for the bias job.
If it is None then configuration of the latest baselining job will be reused, but
if no baselining job then fail the call.}

\item{\code{output_s3_uri}}{(str): S3 destination of the constraint_violations and analysis result.
Default: "s3://<default_session_bucket>/<job_name>/output"}

\item{\code{constraints}}{(sagemaker.model_monitor.Constraints or str): If provided it will be used
for monitoring the endpoint. It can be a Constraints object or an S3 uri pointing
to a constraints JSON file.}

\item{\code{schedule_cron_expression}}{(str): The cron expression that dictates the frequency that
this job run. See sagemaker.model_monitor.CronExpressionGenerator for valid
expressions. Default: Daily.}

\item{\code{enable_cloudwatch_metrics}}{(bool): Whether to publish cloudwatch metrics as part of
the baselining or monitoring jobs.}

\item{\code{role}}{(str): An AWS IAM role. The Amazon SageMaker jobs use this role.}

\item{\code{instance_count}}{(int): The number of instances to run
the jobs with.}

\item{\code{instance_type}}{(str): Type of EC2 instance to use for
the job, for example, 'ml.m5.xlarge'.}

\item{\code{volume_size_in_gb}}{(int): Size in GB of the EBS volume
to use for storing data during processing (default: 30).}

\item{\code{volume_kms_key}}{(str): A KMS key for the job's volume.}

\item{\code{output_kms_key}}{(str): The KMS key id for the job's outputs.}

\item{\code{max_runtime_in_seconds}}{(int): Timeout in seconds. After this amount of
time, Amazon SageMaker terminates the job regardless of its current status.
Default: 3600}

\item{\code{env}}{(dict): Environment variables to be passed to the job.}

\item{\code{network_config}}{(sagemaker.network.NetworkConfig): A NetworkConfig
object that configures network isolation, encryption of
inter-container traffic, security group IDs, and subnets.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-delete_monitoring_schedule"></a>}}
\if{latex}{\out{\hypertarget{method-delete_monitoring_schedule}{}}}
\subsection{Method \code{delete_monitoring_schedule()}}{
Deletes the monitoring schedule and its job definition.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelExplainabilityMonitor$delete_monitoring_schedule()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-attach"></a>}}
\if{latex}{\out{\hypertarget{method-attach}{}}}
\subsection{Method \code{attach()}}{
Sets this object's schedule name to the name provided.
             This allows subsequent describe_schedule or list_executions calls to point
             to the given schedule.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelExplainabilityMonitor$attach(
  monitor_schedule_name,
  sagemaker_session = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{monitor_schedule_name}}{(str): The name of the schedule to attach to.}

\item{\code{sagemaker_session}}{(sagemaker.session.Session): Session object which
manages interactions with Amazon SageMaker APIs and any other
AWS services needed. If not specified, one is created using
the default AWS configuration chain.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelExplainabilityMonitor$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
