% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_monitoring.R
\name{ModelMonitor}
\alias{ModelMonitor}
\title{ModelMonitor Class}
\description{
Sets up Amazon SageMaker Monitoring Schedules and baseline suggestions. Use this class when
             you want to provide your own container image containing the code you'd like to run, in order
             to produce your own statistics and constraint validation files.
             For a more guided experience, consider using the DefaultModelMonitor class instead.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ModelMonitor$new()}}
\item \href{#method-run_baseline}{\code{ModelMonitor$run_baseline()}}
\item \href{#method-clone}{\code{ModelMonitor$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initializes a ``Monitor`` instance. The Monitor handles baselining datasets and
             creating Amazon SageMaker Monitoring Schedules to monitor SageMaker endpoints.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelMonitor$new(
  role = NULL,
  image_uri = NULL,
  instance_count = 1,
  instance_type = "ml.m5.xlarge",
  entrypoint = NULL,
  volume_size_in_gb = 30,
  volume_kms_key = NULL,
  output_kms_key = NULL,
  max_runtime_in_seconds = NULL,
  base_job_name = NULL,
  sagemaker_session = NULL,
  env = NULL,
  tags = NULL,
  network_config = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{role}}{(str): An AWS IAM role. The Amazon SageMaker jobs use this role.}

\item{\code{image_uri}}{(str): The uri of the image to use for the jobs started by
the Monitor.}

\item{\code{instance_count}}{(int): The number of instances to run
the jobs with.}

\item{\code{instance_type}}{(str): Type of EC2 instance to use for
the job, for example, 'ml.m5.xlarge'.}

\item{\code{entrypoint}}{([str]): The entrypoint for the job.}

\item{\code{volume_size_in_gb}}{(int): Size in GB of the EBS volume
to use for storing data during processing (default: 30).}

\item{\code{volume_kms_key}}{(str): A KMS key for the job's volume.}

\item{\code{output_kms_key}}{(str): The KMS key id for the job's outputs.}

\item{\code{max_runtime_in_seconds}}{(int): Timeout in seconds. After this amount of
time, Amazon SageMaker terminates the job regardless of its current status.
Default: 3600}

\item{\code{base_job_name}}{(str): Prefix for the job name. If not specified,
a default name is generated based on the training image name and
current timestamp.}

\item{\code{sagemaker_session}}{(sagemaker.session.Session): Session object which
manages interactions with Amazon SageMaker APIs and any other
AWS services needed. If not specified, one is created using
the default AWS configuration chain.}

\item{\code{env}}{(dict): Environment variables to be passed to the job.}

\item{\code{tags}}{([dict]): List of tags to be passed to the job.}

\item{\code{network_config}}{(sagemaker.network.NetworkConfig): A NetworkConfig
object that configures network isolation, encryption of
inter-container traffic, security group IDs, and subnets.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-run_baseline"></a>}}
\if{latex}{\out{\hypertarget{method-run_baseline}{}}}
\subsection{Method \code{run_baseline()}}{
Run a processing job meant to baseline your dataset.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelMonitor$run_baseline(
  baseline_inputs,
  output,
  arguments = NULL,
  wait = TRUE,
  logs = TRUE,
  job_name = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{baseline_inputs}}{([sagemaker.processing.ProcessingInput]): Input files for the processing
job. These must be provided as ProcessingInput objects.}

\item{\code{output}}{(sagemaker.processing.ProcessingOutput): Destination of the constraint_violations
and statistics json files.}

\item{\code{arguments}}{([str]): A list of string arguments to be passed to a processing job.}

\item{\code{wait}}{(bool): Whether the call should wait until the job completes (default: True).}

\item{\code{logs}}{(bool): Whether to show the logs produced by the job.
Only meaningful when wait is True (default: True).}

\item{\code{job_name}}{(str): Processing job name. If not specified, the processor generates
a default job name, based on the image name and current timestamp.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelMonitor$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
