% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_monitor_monitoring_files.R
\name{ModelMonitoringFile}
\alias{ModelMonitoringFile}
\title{ModelMonitoringFile Class}
\description{
Represents a file with a body and an S3 uri.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ModelMonitoringFile$new()}}
\item \href{#method-save}{\code{ModelMonitoringFile$save()}}
\item \href{#method-format}{\code{ModelMonitoringFile$format()}}
\item \href{#method-clone}{\code{ModelMonitoringFile$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initializes a file with a body and an S3 uri.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelMonitoringFile$new(body_dict, file_s3_uri, kms_key, sagemaker_session)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{body_dict}}{(str): The body of the JSON file.}

\item{\code{file_s3_uri}}{(str): The uri of the JSON file.}

\item{\code{kms_key}}{(str): The kms key to be used to decrypt the file in S3.}

\item{\code{sagemaker_session}}{(sagemaker.session.Session): A SageMaker Session
object, used for SageMaker interactions (default: None). If not
specified, one is created using the default AWS configuration
chain.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-save"></a>}}
\if{latex}{\out{\hypertarget{method-save}{}}}
\subsection{Method \code{save()}}{
Save the current instance's body to s3 using the instance's s3 path.
             The S3 path can be overridden by providing one. This also overrides the
             default save location for this object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelMonitoringFile$save(new_save_location_s3_uri = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{new_save_location_s3_uri}}{(str): Optional. The S3 path to save the file to. If not
provided, the file is saved in place in S3. If provided, the file's S3 path is
permanently updated.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
str: The s3 location to which the file was saved.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-format"></a>}}
\if{latex}{\out{\hypertarget{method-format}{}}}
\subsection{Method \code{format()}}{
format class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelMonitoringFile$format()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelMonitoringFile$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
