% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_monitor_model_monitoring.R
\name{ModelQualityMonitor}
\alias{ModelQualityMonitor}
\title{Amazon SageMaker model monitor to monitor quality metrics for an endpoint.}
\description{
Please see the `initialize` method of its base class for how to instantiate it.
}
\section{Super class}{
\code{\link[R6sagemaker:ModelMonitor]{R6sagemaker::ModelMonitor}} -> \code{ModelQualityMonitor}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{JOB_DEFINITION_BASE_NAME}}{Model definition base name}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ModelQualityMonitor$new()}}
\item \href{#method-monitoring_type}{\code{ModelQualityMonitor$monitoring_type()}}
\item \href{#method-suggest_baseline}{\code{ModelQualityMonitor$suggest_baseline()}}
\item \href{#method-create_monitorying_schedule}{\code{ModelQualityMonitor$create_monitorying_schedule()}}
\item \href{#method-update_monitoring_schedule}{\code{ModelQualityMonitor$update_monitoring_schedule()}}
\item \href{#method-delete_monitoring_schedule}{\code{ModelQualityMonitor$delete_monitoring_schedule()}}
\item \href{#method-attach}{\code{ModelQualityMonitor$attach()}}
\item \href{#method-clone}{\code{ModelQualityMonitor$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="R6sagemaker" data-topic="ModelMonitor" data-id="baseline_statistics">}\href{../../R6sagemaker/html/ModelMonitor.html#method-baseline_statistics}{\code{R6sagemaker::ModelMonitor$baseline_statistics()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="R6sagemaker" data-topic="ModelMonitor" data-id="create_monitoring_schedule">}\href{../../R6sagemaker/html/ModelMonitor.html#method-create_monitoring_schedule}{\code{R6sagemaker::ModelMonitor$create_monitoring_schedule()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="R6sagemaker" data-topic="ModelMonitor" data-id="describe_latest_baselining_job">}\href{../../R6sagemaker/html/ModelMonitor.html#method-describe_latest_baselining_job}{\code{R6sagemaker::ModelMonitor$describe_latest_baselining_job()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="R6sagemaker" data-topic="ModelMonitor" data-id="describe_schedule">}\href{../../R6sagemaker/html/ModelMonitor.html#method-describe_schedule}{\code{R6sagemaker::ModelMonitor$describe_schedule()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="R6sagemaker" data-topic="ModelMonitor" data-id="format">}\href{../../R6sagemaker/html/ModelMonitor.html#method-format}{\code{R6sagemaker::ModelMonitor$format()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="R6sagemaker" data-topic="ModelMonitor" data-id="latest_monitoring_constraint_violations">}\href{../../R6sagemaker/html/ModelMonitor.html#method-latest_monitoring_constraint_violations}{\code{R6sagemaker::ModelMonitor$latest_monitoring_constraint_violations()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="R6sagemaker" data-topic="ModelMonitor" data-id="latest_monitoring_statistics">}\href{../../R6sagemaker/html/ModelMonitor.html#method-latest_monitoring_statistics}{\code{R6sagemaker::ModelMonitor$latest_monitoring_statistics()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="R6sagemaker" data-topic="ModelMonitor" data-id="list_executions">}\href{../../R6sagemaker/html/ModelMonitor.html#method-list_executions}{\code{R6sagemaker::ModelMonitor$list_executions()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="R6sagemaker" data-topic="ModelMonitor" data-id="run_baseline">}\href{../../R6sagemaker/html/ModelMonitor.html#method-run_baseline}{\code{R6sagemaker::ModelMonitor$run_baseline()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="R6sagemaker" data-topic="ModelMonitor" data-id="start_monitoring_schedule">}\href{../../R6sagemaker/html/ModelMonitor.html#method-start_monitoring_schedule}{\code{R6sagemaker::ModelMonitor$start_monitoring_schedule()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="R6sagemaker" data-topic="ModelMonitor" data-id="stop_monitoring_schedule">}\href{../../R6sagemaker/html/ModelMonitor.html#method-stop_monitoring_schedule}{\code{R6sagemaker::ModelMonitor$stop_monitoring_schedule()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="R6sagemaker" data-topic="ModelMonitor" data-id="suggested_constraints">}\href{../../R6sagemaker/html/ModelMonitor.html#method-suggested_constraints}{\code{R6sagemaker::ModelMonitor$suggested_constraints()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initializes a monitor instance.
             The monitor handles baselining datasets and creating Amazon SageMaker
             Monitoring Schedules to monitor SageMaker endpoints.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelQualityMonitor$new(
  role,
  instance_count = 1,
  instance_type = "ml.m5.xlarge",
  volume_size_in_gb = 30,
  volume_kms_key = NULL,
  output_kms_key = NULL,
  max_runtime_in_seconds = NULL,
  base_job_name = NULL,
  sagemaker_session = NULL,
  env = NULL,
  tags = NULL,
  network_config = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{role}}{(str): An AWS IAM role. The Amazon SageMaker jobs use this role.}

\item{\code{instance_count}}{(int): The number of instances to run
the jobs with.}

\item{\code{instance_type}}{(str): Type of EC2 instance to use for
the job, for example, 'ml.m5.xlarge'.}

\item{\code{volume_size_in_gb}}{(int): Size in GB of the EBS volume
to use for storing data during processing (default: 30).}

\item{\code{volume_kms_key}}{(str): A KMS key for the job's volume.}

\item{\code{output_kms_key}}{(str): The KMS key id for the job's outputs.}

\item{\code{max_runtime_in_seconds}}{(int): Timeout in seconds. After this amount of
time, Amazon SageMaker terminates the job regardless of its current status.
Default: 3600}

\item{\code{base_job_name}}{(str): Prefix for the job name. If not specified,
a default name is generated based on the training image name and
current timestamp.}

\item{\code{sagemaker_session}}{(sagemaker.session.Session): Session object which
manages interactions with Amazon SageMaker APIs and any other
AWS services needed. If not specified, one is created using
the default AWS configuration chain.}

\item{\code{env}}{(dict): Environment variables to be passed to the job.}

\item{\code{tags}}{([dict]): List of tags to be passed to the job.}

\item{\code{network_config}}{(sagemaker.network.NetworkConfig): A NetworkConfig
object that configures network isolation, encryption of
inter-container traffic, security group IDs, and subnets.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-monitoring_type"></a>}}
\if{latex}{\out{\hypertarget{method-monitoring_type}{}}}
\subsection{Method \code{monitoring_type()}}{
Type of the monitoring job.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelQualityMonitor$monitoring_type()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-suggest_baseline"></a>}}
\if{latex}{\out{\hypertarget{method-suggest_baseline}{}}}
\subsection{Method \code{suggest_baseline()}}{
Suggest baselines for use with Amazon SageMaker Model Monitoring Schedules.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelQualityMonitor$suggest_baseline(
  baseline_dataset,
  dataset_format,
  problem_type,
  inference_attribute = NULL,
  probability_attribute = NULL,
  ground_truth_attribute = NULL,
  probability_threshold_attribute = NULL,
  post_analytics_processor_script = NULL,
  output_s3_uri = NULL,
  wait = FALSE,
  logs = FALSE,
  job_name = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{baseline_dataset}}{(str): The path to the baseline_dataset file. This can be a local
path or an S3 uri.}

\item{\code{dataset_format}}{(dict): The format of the baseline_dataset.}

\item{\code{problem_type}}{(str): The type of problem of this model quality monitoring. Valid
values are "Regression", "BinaryClassification", "MulticlassClassification".}

\item{\code{inference_attribute}}{(str): Index or JSONpath to locate predicted label(s).}

\item{\code{probability_attribute}}{(str or int): Index or JSONpath to locate probabilities.}

\item{\code{ground_truth_attribute}}{(str): Index or JSONpath to locate actual label(s).}

\item{\code{probability_threshold_attribute}}{(float): threshold to convert probabilities to binaries
Only used for ModelQualityMonitor, ModelBiasMonitor and ModelExplainabilityMonitor}

\item{\code{post_analytics_processor_script}}{(str): The path to the record post-analytics processor
script. This can be a local path or an S3 uri.}

\item{\code{output_s3_uri}}{(str): Desired S3 destination Destination of the constraint_violations
and statistics json files.
Default: "s3://<default_session_bucket>/<job_name>/output"}

\item{\code{wait}}{(bool): Whether the call should wait until the job completes (default: False).}

\item{\code{logs}}{(bool): Whether to show the logs produced by the job.
Only meaningful when wait is True (default: False).}

\item{\code{job_name}}{(str): Processing job name. If not specified, the processor generates
a default job name, based on the image name and current timestamp.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
sagemaker.processing.ProcessingJob: The ProcessingJob object representing the
             baselining job.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-create_monitorying_schedule"></a>}}
\if{latex}{\out{\hypertarget{method-create_monitorying_schedule}{}}}
\subsection{Method \code{create_monitorying_schedule()}}{
Creates a monitoring schedule.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelQualityMonitor$create_monitorying_schedule(
  endpoint_input,
  ground_truth_input,
  problem_type,
  record_preprocessor_script = NULL,
  post_analytics_processor_script = NULL,
  output_s3_uri = NULL,
  constraints = NULL,
  monitor_schedule_name = NULL,
  schedule_cron_expression = NULL,
  enable_cloudwatch_metrics = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{endpoint_input}}{(str or sagemaker.model_monitor.EndpointInput): The endpoint to
monitor. This can either be the endpoint name or an EndpointInput.}

\item{\code{ground_truth_input}}{(str): S3 URI to ground truth dataset.}

\item{\code{problem_type}}{(str): The type of problem of this model quality monitoring. Valid
values are "Regression", "BinaryClassification", "MulticlassClassification".}

\item{\code{record_preprocessor_script}}{(str): The path to the record preprocessor script. This can
be a local path or an S3 uri.}

\item{\code{post_analytics_processor_script}}{(str): The path to the record post-analytics processor
script. This can be a local path or an S3 uri.}

\item{\code{output_s3_uri}}{(str): S3 destination of the constraint_violations and analysis result.
Default: "s3://<default_session_bucket>/<job_name>/output"}

\item{\code{constraints}}{(sagemaker.model_monitor.Constraints or str): If provided it will be used
for monitoring the endpoint. It can be a Constraints object or an S3 uri pointing
to a constraints JSON file.}

\item{\code{monitor_schedule_name}}{(str): Schedule name. If not specified, the processor generates
a default job name, based on the image name and current timestamp.}

\item{\code{schedule_cron_expression}}{(str): The cron expression that dictates the frequency that
this job run. See sagemaker.model_monitor.CronExpressionGenerator for valid
expressions. Default: Daily.}

\item{\code{enable_cloudwatch_metrics}}{(bool): Whether to publish cloudwatch metrics as part of
the baselining or monitoring jobs.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update_monitoring_schedule"></a>}}
\if{latex}{\out{\hypertarget{method-update_monitoring_schedule}{}}}
\subsection{Method \code{update_monitoring_schedule()}}{
Updates the existing monitoring schedule.
             If more options than schedule_cron_expression are to be updated, a new job definition will
             be created to hold them. The old job definition will not be deleted.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelQualityMonitor$update_monitoring_schedule(
  endpoint_input = NULL,
  ground_truth_input = NULL,
  problem_type = NULL,
  record_preprocessor_script = NULL,
  post_analytics_processor_script = NULL,
  output_s3_uri = NULL,
  constraints = NULL,
  schedule_cron_expression = NULL,
  enable_cloudwatch_metrics = NULL,
  role = NULL,
  instance_count = NULL,
  instance_type = NULL,
  volume_size_in_gb = NULL,
  volume_kms_key = NULL,
  output_kms_key = NULL,
  max_runtime_in_seconds = NULL,
  env = NULL,
  network_config = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{endpoint_input}}{(str or sagemaker.model_monitor.EndpointInput): The endpoint
to monitor. This can either be the endpoint name or an EndpointInput.}

\item{\code{ground_truth_input}}{(str): S3 URI to ground truth dataset.}

\item{\code{problem_type}}{(str): The type of problem of this model quality monitoring. Valid values
are "Regression", "BinaryClassification", "MulticlassClassification".}

\item{\code{record_preprocessor_script}}{(str): The path to the record preprocessor script. This can
be a local path or an S3 uri.}

\item{\code{post_analytics_processor_script}}{(str): The path to the record post-analytics processor
script. This can be a local path or an S3 uri.}

\item{\code{output_s3_uri}}{(str): S3 destination of the constraint_violations and analysis result.
Default: "s3://<default_session_bucket>/<job_name>/output"}

\item{\code{constraints}}{(sagemaker.model_monitor.Constraints or str): If provided it will be used
for monitoring the endpoint. It can be a Constraints object or an S3 uri pointing
to a constraints JSON file.}

\item{\code{schedule_cron_expression}}{(str): The cron expression that dictates the frequency that
this job run. See sagemaker.model_monitor.CronExpressionGenerator for valid
expressions. Default: Daily.}

\item{\code{enable_cloudwatch_metrics}}{(bool): Whether to publish cloudwatch metrics as part of
the baselining or monitoring jobs.}

\item{\code{role}}{(str): An AWS IAM role. The Amazon SageMaker jobs use this role.}

\item{\code{instance_count}}{(int): The number of instances to run
the jobs with.}

\item{\code{instance_type}}{(str): Type of EC2 instance to use for
the job, for example, 'ml.m5.xlarge'.}

\item{\code{volume_size_in_gb}}{(int): Size in GB of the EBS volume
to use for storing data during processing (default: 30).}

\item{\code{volume_kms_key}}{(str): A KMS key for the job's volume.}

\item{\code{output_kms_key}}{(str): The KMS key id for the job's outputs.}

\item{\code{max_runtime_in_seconds}}{(int): Timeout in seconds. After this amount of
time, Amazon SageMaker terminates the job regardless of its current status.
Default: 3600}

\item{\code{env}}{(dict): Environment variables to be passed to the job.}

\item{\code{network_config}}{(sagemaker.network.NetworkConfig): A NetworkConfig
object that configures network isolation, encryption of
inter-container traffic, security group IDs, and subnets.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-delete_monitoring_schedule"></a>}}
\if{latex}{\out{\hypertarget{method-delete_monitoring_schedule}{}}}
\subsection{Method \code{delete_monitoring_schedule()}}{
Deletes the monitoring schedule and its job definition."
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelQualityMonitor$delete_monitoring_schedule()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-attach"></a>}}
\if{latex}{\out{\hypertarget{method-attach}{}}}
\subsection{Method \code{attach()}}{
Sets this object's schedule name to the name provided.
             This allows subsequent describe_schedule or list_executions calls to point
             to the given schedule.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelQualityMonitor$attach(monitor_schedule_name, sagemaker_session = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{monitor_schedule_name}}{(str): The name of the schedule to attach to.}

\item{\code{sagemaker_session}}{(sagemaker.session.Session): Session object which
manages interactions with Amazon SageMaker APIs and any other
AWS services needed. If not specified, one is created using
the default AWS configuration chain.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelQualityMonitor$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
