% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serializers.R
\name{NumpySerializer}
\alias{NumpySerializer}
\title{NumpySerializer Class}
\description{
Serialize data of various formats to a numpy npy file format.
             This serializer class uses python numpy package to serialize,
             R objects through the use of the `reticualte` package.
}
\section{Super classes}{
\code{\link[R6sagemaker:BaseSerializer]{R6sagemaker::BaseSerializer}} -> \code{\link[R6sagemaker:SimpleBaseSerializer]{R6sagemaker::SimpleBaseSerializer}} -> \code{NumpySerializer}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{np}}{Python Numpy package}

\item{\code{dtype}}{The dtype of the data}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{NumpySerializer$new()}}
\item \href{#method-serialize}{\code{NumpySerializer$serialize()}}
\item \href{#method-clone}{\code{NumpySerializer$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="R6sagemaker" data-topic="BaseSerializer" data-id="format">}\href{../../R6sagemaker/html/BaseSerializer.html#method-format}{\code{R6sagemaker::BaseSerializer$format()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialize a ``NumpySerializer`` instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NumpySerializer$new(dtype = NULL, content_type = "application/x-npy")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dtype}}{(str): The dtype of the data.}

\item{\code{content_type}}{(str): The MIME type to signal to the inference endpoint when sending
request data (default: "application/x-npy").}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-serialize"></a>}}
\if{latex}{\out{\hypertarget{method-serialize}{}}}
\subsection{Method \code{serialize()}}{
Serialize data to a buffer using the .npy format.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NumpySerializer$serialize(data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{(object): Data to be serialized. Can be a NumPy array, list,
file, or buffer.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
(raw): A buffer containing data serialzied in the .npy format.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NumpySerializer$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
