% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter.R
\name{ParameterRange}
\alias{ParameterRange}
\title{ParameterRange Class}
\description{
Base class for representing parameter ranges. This is used to define what
             hyperparameters to tune for an Amazon SageMaker hyperparameter tuning job
             and to verify hyperparameters for Marketplace Algorithms.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{.all_types}}{All types of child class}

\item{\code{min_value}}{The minimum value for the range}

\item{\code{max_value}}{The maximum value for the rang}

\item{\code{scaling_type}}{The scale used for searching the range during tuning}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ParameterRange$new()}}
\item \href{#method-is_valid}{\code{ParameterRange$is_valid()}}
\item \href{#method-cast_to_type}{\code{ParameterRange$cast_to_type()}}
\item \href{#method-as_tuning_range}{\code{ParameterRange$as_tuning_range()}}
\item \href{#method-format}{\code{ParameterRange$format()}}
\item \href{#method-clone}{\code{ParameterRange$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialize a parameter range.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ParameterRange$new(
  min_value,
  max_value,
  scaling_type = c("Auto", "Linear", "Logarithmic", "ReverseLogarithmic")
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{min_value}}{(float or int): The minimum value for the range.}

\item{\code{max_value}}{(float or int): The maximum value for the range.}

\item{\code{scaling_type}}{(str): The scale used for searching the range during
tuning (default: 'Auto'). Valid values: 'Auto', 'Linear',
Logarithmic' and 'ReverseLogarithmic'.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-is_valid"></a>}}
\if{latex}{\out{\hypertarget{method-is_valid}{}}}
\subsection{Method \code{is_valid()}}{
Determine if a value is valid within this ParameterRange.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ParameterRange$is_valid(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{(float or int): The value to be verified.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
bool: True if valid, False otherwise.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-cast_to_type"></a>}}
\if{latex}{\out{\hypertarget{method-cast_to_type}{}}}
\subsection{Method \code{cast_to_type()}}{
cast value to numeric
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ParameterRange$cast_to_type(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{The value to be verified.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-as_tuning_range"></a>}}
\if{latex}{\out{\hypertarget{method-as_tuning_range}{}}}
\subsection{Method \code{as_tuning_range()}}{
Represent the parameter range as a dicionary suitable for a request
             to create an Amazon SageMaker hyperparameter tuning job.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ParameterRange$as_tuning_range(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{(str): The name of the hyperparameter.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
dict[str, str]: A dictionary that contains the name and values of
             the hyperparameter.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-format"></a>}}
\if{latex}{\out{\hypertarget{method-format}{}}}
\subsection{Method \code{format()}}{
format class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ParameterRange$format()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ParameterRange$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
