% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_credentials.R
\name{paws_cred}
\alias{paws_cred}
\title{S3 method to call PawsCredentials class}
\usage{
paws_cred(
  aws_access_key_id = NULL,
  aws_secret_access_key = NULL,
  aws_session_token = NULL,
  region_name = NULL,
  profile_name = NULL
)
}
\arguments{
\item{aws_access_key_id}{(str): AWS access key}

\item{aws_secret_access_key}{(str): AWS secret access key}

\item{aws_session_token}{(str): AWS token}

\item{region_name}{(str): AWS region}

\item{profile_name}{(str): AWS profile}
}
\description{
S3 method to call PawsCredentials class
}
