% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_credentials.R
\name{PawsCredentials}
\alias{PawsCredentials}
\title{Sagemaker Session Class}
\description{
Sets up AWS credentials to pass through to other sagemaker classes#'
}
\seealso{
[paws_cred]
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{aws_access_key_id}}{aws access key}

\item{\code{aws_secret_access_key}}{aws secret access key}

\item{\code{aws_session_token}}{aws session token}

\item{\code{region_name}}{aws region}

\item{\code{profile_name}}{aws cli profile name}

\item{\code{credentials}}{formatted aws credentials to pass to paws objects}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{PawsCredentials$new()}}
\item \href{#method-print}{\code{PawsCredentials$print()}}
\item \href{#method-clone}{\code{PawsCredentials$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this [R6][R6::R6Class] class.

Note: This object is typically constructed via a derived classes via the S3 method
[paws_cred()].
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PawsCredentials$new(
  aws_access_key_id = NULL,
  aws_secret_access_key = NULL,
  aws_session_token = NULL,
  region_name = NULL,
  profile_name = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{aws_access_key_id}}{(str): AWS access key}

\item{\code{aws_secret_access_key}}{(str): AWS secret access key}

\item{\code{aws_session_token}}{(str): AWS token}

\item{\code{region_name}}{(str): AWS region}

\item{\code{profile_name}}{(str): AWS profile}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Printer.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PawsCredentials$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{(ignored).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PawsCredentials$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
