% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session.R
\name{pipeline_container_def}
\alias{pipeline_container_def}
\title{Create a definition for executing a pipeline of containers as part of a SageMaker model.}
\usage{
pipeline_container_def(models, instance_type = NULL)
}
\arguments{
\item{models}{(list[sagemaker.Model]): this will be a list of ``sagemaker.Model`` objects in the
order the inference should be invoked.}

\item{instance_type}{(str): The EC2 instance type to deploy this Model to. For example,
'ml.p2.xlarge' (default: None).}
}
\value{
list[dict[str, str]]: list of container definition objects usable with with the
             CreateModel API for inference pipelines if passed via `Containers` field.
}
\description{
Create a definition for executing a pipeline of containers as part of a SageMaker model.
}
