% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictor.R
\name{predict.RealTimePredictor}
\alias{predict.RealTimePredictor}
\title{S3 method that wraps RealTimePredictor Class}
\usage{
\method{predict}{RealTimePredictor}(
  object,
  newdata,
  serializer = csv_serializer,
  deserializer = csv_deserializer,
  ...
)
}
\arguments{
\item{object}{a sagemaker model}

\item{newdata}{data for model to predict}

\item{serializer}{method class to serializer data to sagemaker model. Requires to be
a class inherited from \link{BaseSerializer}. (Default: \link{csv_serializer})}

\item{deserializer}{method class to deserializer return data streams from sagemaker model.
Requires to be a class inherited from \link{BaseDeserializer}.
(Default: \link{csv_deserializer})}

\item{...}{arguments passed to ``RealTimePredictor$predict``}
}
\description{
Predicted values returned from endpoint
}
