% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debugger_profiler_config.R
\name{ProfilerConfig}
\alias{ProfilerConfig}
\title{Configuration for collecting system and framework metrics of SageMaker training jobs.}
\description{
SageMaker Debugger collects system and framework profiling
             information of training jobs and identify performance bottlenecks.
}
\examples{

## ------------------------------------------------
## Method `ProfilerConfig$new`
## ------------------------------------------------

# The following example shows the basic ``profiler_config``
# parameter configuration, enabling system monitoring every 5000 milliseconds
# and framework profiling with default parameter values.
library(R6sagemaker)

profiler_config = ProfilerConfig$new(
      system_monitor_interval_millis = 5000,
      framework_profile_params = FrameworkProfile$new()
 )
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ProfilerConfig$new()}}
\item \href{#method-to_request_list}{\code{ProfilerConfig$to_request_list()}}
\item \href{#method-format}{\code{ProfilerConfig$format()}}
\item \href{#method-clone}{\code{ProfilerConfig$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialize a ``ProfilerConfig`` instance.
             Pass the output of this class
             to the ``profiler_config`` parameter of the generic :class:`~sagemaker.estimator.Estimator`
             class and SageMaker Framework estimators.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProfilerConfig$new(
  s3_output_path = NULL,
  system_monitor_interval_millis = NULL,
  framework_profile_params = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{s3_output_path}}{(str): The location in Amazon S3 to store the output.
The default Debugger output path for profiling data is created under the
default output path of the :class:`~sagemaker.estimator.Estimator` class.
For example,
s3://sagemaker-<region>-<12digit_account_id>/<training-job-name>/profiler-output/.}

\item{\code{system_monitor_interval_millis}}{(int): The time interval in milliseconds
to collect system metrics. Available values are 100, 200, 500, 1000 (1 second),
5000 (5 seconds), and 60000 (1 minute) milliseconds.
The default is 500 milliseconds.}

\item{\code{framework_profile_params}}{(:class:`~sagemaker.debugger.FrameworkProfile`):
A parameter object for framework metrics profiling. Configure it using
the :class:`~sagemaker.debugger.FrameworkProfile` class.
To use the default framework profile parameters, pass ``FrameworkProfile()``.
For more information about the default values,
see :class:`~sagemaker.debugger.FrameworkProfile`.}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{# The following example shows the basic ``profiler_config``
# parameter configuration, enabling system monitoring every 5000 milliseconds
# and framework profiling with default parameter values.
library(R6sagemaker)

profiler_config = ProfilerConfig$new(
      system_monitor_interval_millis = 5000,
      framework_profile_params = FrameworkProfile$new()
 )
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-to_request_list"></a>}}
\if{latex}{\out{\hypertarget{method-to_request_list}{}}}
\subsection{Method \code{to_request_list()}}{
Generate a request dictionary using the parameters provided when initializing the object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProfilerConfig$to_request_list()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
dict: An portion of an API request as a dictionary.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-format"></a>}}
\if{latex}{\out{\hypertarget{method-format}{}}}
\subsection{Method \code{format()}}{
format class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProfilerConfig$format()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProfilerConfig$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
