% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debugger_utils.R
\docType{data}
\name{PythonProfiler}
\alias{PythonProfiler}
\title{PythonProfiler enum environment list}
\format{
An object of class \code{PythonProfiler} (inherits from \code{Enum}, \code{environment}) of length 2.
}
\usage{
PythonProfiler
}
\value{
environment containing [CPROFILE, PYINSTRUMENT]
}
\description{
List the Python profiler options for Python profiling.
}
\keyword{datasets}
