% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pytorch_model.R
\name{PyTorchPredictor}
\alias{PyTorchPredictor}
\title{A Predictor for inference against PyTorch Endpoints.}
\description{
This is able to serialize Python lists, dictionaries, and numpy arrays to
             multidimensional tensors for PyTorch inference.
}
\section{Super class}{
\code{\link[R6sagemaker:Predictor]{R6sagemaker::Predictor}} -> \code{PyTorchPredictor}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{PyTorchPredictor$new()}}
\item \href{#method-clone}{\code{PyTorchPredictor$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="R6sagemaker" data-topic="Predictor" data-id="delete_endpoint">}\href{../../R6sagemaker/html/Predictor.html#method-delete_endpoint}{\code{R6sagemaker::Predictor$delete_endpoint()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="R6sagemaker" data-topic="Predictor" data-id="delete_model">}\href{../../R6sagemaker/html/Predictor.html#method-delete_model}{\code{R6sagemaker::Predictor$delete_model()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="R6sagemaker" data-topic="Predictor" data-id="disable_data_capture">}\href{../../R6sagemaker/html/Predictor.html#method-disable_data_capture}{\code{R6sagemaker::Predictor$disable_data_capture()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="R6sagemaker" data-topic="Predictor" data-id="enable_data_capture">}\href{../../R6sagemaker/html/Predictor.html#method-enable_data_capture}{\code{R6sagemaker::Predictor$enable_data_capture()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="R6sagemaker" data-topic="Predictor" data-id="endpoint_context">}\href{../../R6sagemaker/html/Predictor.html#method-endpoint_context}{\code{R6sagemaker::Predictor$endpoint_context()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="R6sagemaker" data-topic="Predictor" data-id="list_monitors">}\href{../../R6sagemaker/html/Predictor.html#method-list_monitors}{\code{R6sagemaker::Predictor$list_monitors()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="R6sagemaker" data-topic="Predictor" data-id="predict">}\href{../../R6sagemaker/html/Predictor.html#method-predict}{\code{R6sagemaker::Predictor$predict()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="R6sagemaker" data-topic="Predictor" data-id="print">}\href{../../R6sagemaker/html/Predictor.html#method-print}{\code{R6sagemaker::Predictor$print()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="R6sagemaker" data-topic="Predictor" data-id="update_data_capture_config">}\href{../../R6sagemaker/html/Predictor.html#method-update_data_capture_config}{\code{R6sagemaker::Predictor$update_data_capture_config()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="R6sagemaker" data-topic="Predictor" data-id="update_endpoint">}\href{../../R6sagemaker/html/Predictor.html#method-update_endpoint}{\code{R6sagemaker::Predictor$update_endpoint()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialize an ``PyTorchPredictor``.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PyTorchPredictor$new(endpoint_name, sagemaker_session = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{endpoint_name}}{(str): The name of the endpoint to perform inference
on.}

\item{\code{sagemaker_session}}{(sagemaker.session.Session): Session object which
manages interactions with Amazon SageMaker APIs and any other
AWS services needed. If not specified, the estimator creates one
using the default AWS configuration chain.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PyTorchPredictor$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
