% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r6sagemaker.R
\docType{package}
\name{R6sagemaker-package}
\alias{R6sagemaker}
\alias{R6sagemaker-package}
\title{r6 sagemaker: this is just a placeholder}
\description{
Interface to Amazon Web Services SageMaker <https://aws.amazon.com/sagemaker/>
    Use four spaces when indenting paragraphs within the Description.
}
\author{
\strong{Maintainer}: Dyfan Jones \email{dyfan.r.jones@gmail.com}

Other contributors:
\itemize{
  \item Amazon.com, Inc. [copyright holder]
}

}
