% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apiutils_base_types.R
\name{Record}
\alias{Record}
\title{Record class}
\description{
A boto based Active Record class based on convention of CRUD operations.
}
\section{Super class}{
\code{\link[R6sagemaker:ApiObject]{R6sagemaker::ApiObject}} -> \code{Record}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Record$new()}}
\item \href{#method-with_paws}{\code{Record$with_paws()}}
\item \href{#method-clone}{\code{Record$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="R6sagemaker" data-topic="ApiObject" data-id="from_paws">}\href{../../R6sagemaker/html/ApiObject.html#method-from_paws}{\code{R6sagemaker::ApiObject$from_paws()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="R6sagemaker" data-topic="ApiObject" data-id="print">}\href{../../R6sagemaker/html/ApiObject.html#method-print}{\code{R6sagemaker::ApiObject$print()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="R6sagemaker" data-topic="ApiObject" data-id="to_paws">}\href{../../R6sagemaker/html/ApiObject.html#method-to_paws}{\code{R6sagemaker::ApiObject$to_paws()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Init Record.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Record$new(sagemaker_session = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sagemaker_session}}{(sagemaker.session.Session): Session object which
manages interactions with Amazon SageMaker APIs and any other
AWS services needed. If not specified, the estimator creates one
using the default AWS configuration chain.}

\item{\code{...}}{parameters passed to `R6` class `ApiObject`}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-with_paws"></a>}}
\if{latex}{\out{\hypertarget{method-with_paws}{}}}
\subsection{Method \code{with_paws()}}{
Update this ApiObject with a paws response.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Record$with_paws(paws_list)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{paws_list}}{(dict): A dictionary of a paws response.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Record$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
