% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amazon_common.R
\name{RecordDeserializer}
\alias{RecordDeserializer}
\title{RecordDeserializer Class}
\description{
Deserialize RecordIO Protobuf data from an inference endpoint.
}
\section{Super classes}{
\code{\link[R6sagemaker:BaseDeserializer]{R6sagemaker::BaseDeserializer}} -> \code{\link[R6sagemaker:SimpleBaseDeserializer]{R6sagemaker::SimpleBaseDeserializer}} -> \code{RecordDeserializer}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{RecordDeserializer$new()}}
\item \href{#method-deserializer}{\code{RecordDeserializer$deserializer()}}
\item \href{#method-clone}{\code{RecordDeserializer$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="R6sagemaker" data-topic="BaseDeserializer" data-id="deserialize">}\href{../../R6sagemaker/html/BaseDeserializer.html#method-deserialize}{\code{R6sagemaker::BaseDeserializer$deserialize()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="R6sagemaker" data-topic="BaseDeserializer" data-id="print">}\href{../../R6sagemaker/html/BaseDeserializer.html#method-print}{\code{R6sagemaker::BaseDeserializer$print()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Intialize RecordDeserializer class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RecordDeserializer$new(accept = "application/x-recordio-protobuf")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{accept}}{(union[str, tuple[str]]): The MIME type (or tuple of allowable MIME types) that
is expected from the inference endpoint (default:
"application/x-recordio-protobuf").}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-deserializer"></a>}}
\if{latex}{\out{\hypertarget{method-deserializer}{}}}
\subsection{Method \code{deserializer()}}{
Deserialize RecordIO Protobuf data from an inference endpoint.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RecordDeserializer$deserializer(data, content_type)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{(object): The protobuf message to deserialize.}

\item{\code{content_type}}{(str): The MIME type of the data.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
list: A list of records.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RecordDeserializer$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
