% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amazon_estimator.R
\name{RecordSet}
\alias{RecordSet}
\title{RecordSet Class}
\description{
RecordSet Class

RecordSet Class
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{RecordSet$new()}}
\item \href{#method-data_channel}{\code{RecordSet$data_channel()}}
\item \href{#method-recods_s3_input}{\code{RecordSet$recods_s3_input()}}
\item \href{#method-print}{\code{RecordSet$print()}}
\item \href{#method-clone}{\code{RecordSet$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
A collection of Amazon :class:~`Record` objects serialized and stored
             in S3.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RecordSet$new(
  s3_data,
  num_records,
  feature_dim,
  s3_data_type = "ManifestFile",
  channel = "train"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{s3_data}}{(str): The S3 location of the training data}

\item{\code{num_records}}{(int): The number of records in the set.}

\item{\code{feature_dim}}{(int): The dimensionality of "values" arrays in the
Record features, and label (if each Record is labeled).}

\item{\code{s3_data_type}}{(str): Valid values: 'S3Prefix', 'ManifestFile'. If
'S3Prefix', ``s3_data`` defines a prefix of s3 objects to train
on. All objects with s3 keys beginning with ``s3_data`` will be
used to train. If 'ManifestFile', then ``s3_data`` defines a
single s3 manifest file, listing each s3 object to train on.}

\item{\code{channel}}{(str): The SageMaker Training Job channel this RecordSet
should be bound to}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-data_channel"></a>}}
\if{latex}{\out{\hypertarget{method-data_channel}{}}}
\subsection{Method \code{data_channel()}}{
Return a dictionary to represent the training data in a channel for
             use with ``fit()``
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RecordSet$data_channel()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-recods_s3_input"></a>}}
\if{latex}{\out{\hypertarget{method-recods_s3_input}{}}}
\subsection{Method \code{recods_s3_input()}}{
Return a TrainingInput to represent the training data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RecordSet$recods_s3_input()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Return an unambiguous representation of this RecordSet
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RecordSet$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RecordSet$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
