% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rl_estimator.R
\docType{data}
\name{RLFramework}
\alias{RLFramework}
\title{RLFramework enum environment list}
\format{
An object of class \code{Enum} (inherits from \code{environment}) of length 3.
}
\usage{
RLFramework
}
\value{
environment containing [TENSORFLOW, MXNET, PYTORCH]
}
\description{
Framework (MXNet, TensorFlow or PyTorch) you want to be used
             as a toolkit backed for
             reinforcement learning training.
}
\keyword{datasets}
