% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rl_estimator.R
\docType{data}
\name{RLToolkit}
\alias{RLToolkit}
\title{RLToolkit enum environment list}
\format{
An object of class \code{Enum} (inherits from \code{environment}) of length 2.
}
\usage{
RLToolkit
}
\value{
environment containing [COACH, RAY]
}
\description{
RL toolkit you want to use for
             executing your model training code.
}
\keyword{datasets}
