% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debugger.R
\name{Rule}
\alias{Rule}
\title{Debug Rule Class}
\description{
The SageMaker Debugger Rule class configures *debugging* rules to debug your training job.
             The debugging rules analyze tensor outputs from your training job
             and monitor conditions that are critical for the success of the training
             job.
             SageMaker Debugger comes pre-packaged with built-in *debugging* rules.
             For example, the debugging rules can detect whether gradients are getting too large or
             too small, or if a model is overfitting.
             For a full list of built-in rules for debugging, see
             `List of Debugger Built-in Rules
             <https://docs.aws.amazon.com/sagemaker/latest/dg/debugger-built-in-rules.html>`_.
             You can also write your own rules using the custom rule classmethod.
}
\section{Super class}{
\code{\link[R6sagemaker:RuleBase]{R6sagemaker::RuleBase}} -> \code{Rule}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Rule$new()}}
\item \href{#method-sagemaker}{\code{Rule$sagemaker()}}
\item \href{#method-custom}{\code{Rule$custom()}}
\item \href{#method-prepare_actions}{\code{Rule$prepare_actions()}}
\item \href{#method-to_debugger_rule_config_dict}{\code{Rule$to_debugger_rule_config_dict()}}
\item \href{#method-clone}{\code{Rule$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="R6sagemaker" data-topic="RuleBase" data-id="format">}\href{../../R6sagemaker/html/RuleBase.html#method-format}{\code{R6sagemaker::RuleBase$format()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Configure the debugging rules using the following classmethods.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Rule$new(
  name,
  image_uri,
  instance_type,
  container_local_output_path,
  s3_output_path,
  volume_size_in_gb,
  rule_parameters,
  collections_to_save,
  actions = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{(str): The name of the rule.}

\item{\code{image_uri}}{(str): The image URI to use the rule.}

\item{\code{instance_type}}{(str): Type of EC2 instance to use. For example, 'ml.c4.xlarge'.}

\item{\code{container_local_output_path}}{(str): The local path to store the Rule output.}

\item{\code{s3_output_path}}{(str): The location in S3 to store the output.}

\item{\code{volume_size_in_gb}}{(int): Size in GB of the EBS volume to use for storing data.}

\item{\code{rule_parameters}}{(dict): A dictionary of parameters for the rule.}

\item{\code{collections_to_save}}{([sagemaker.debugger.CollectionConfig]): Optional. A list
of :class:`~sagemaker.debugger.CollectionConfig` objects to be saved.}

\item{\code{actions}}{:}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sagemaker"></a>}}
\if{latex}{\out{\hypertarget{method-sagemaker}{}}}
\subsection{Method \code{sagemaker()}}{
Initialize a ``Rule`` object for a \code{built-in} debugging rule.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Rule$sagemaker(
  base_config,
  name = NULL,
  container_local_output_path = NULL,
  s3_output_path = NULL,
  other_trials_s3_input_paths = NULL,
  rule_parameters = NULL,
  collections_to_save = NULL,
  actions = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{base_config}}{(dict): Required. This is the base rule config dictionary returned from the
:class:`~sagemaker.debugger.rule_configs` method.
For example, ``rule_configs.dead_relu()``.
For a full list of built-in rules for debugging, see
`List of Debugger Built-in Rules
<https://docs.aws.amazon.com/sagemaker/latest/dg/debugger-built-in-rules.html>`_.}

\item{\code{name}}{(str): Optional. The name of the debugger rule. If one is not provided,
the name of the base_config will be used.}

\item{\code{container_local_output_path}}{(str): Optional. The local path in the rule processing
container.}

\item{\code{s3_output_path}}{(str): Optional. The location in Amazon S3 to store the output tensors.
The default Debugger output path for debugging data is created under the
default output path of the :class:`~sagemaker.estimator.Estimator` class.
For example,
s3://sagemaker-<region>-<12digit_account_id>/<training-job-name>/debug-output/.}

\item{\code{other_trials_s3_input_paths}}{([str]): Optional. The Amazon S3 input paths
of other trials to use the SimilarAcrossRuns rule.}

\item{\code{rule_parameters}}{(dict): Optional. A dictionary of parameters for the rule.}

\item{\code{collections_to_save}}{(:class:`~sagemaker.debugger.CollectionConfig`):
Optional. A list
of :class:`~sagemaker.debugger.CollectionConfig` objects to be saved.}

\item{\code{actions}}{:}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
:class:`~sagemaker.debugger.Rule`: An instance of the built-in rule.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-custom"></a>}}
\if{latex}{\out{\hypertarget{method-custom}{}}}
\subsection{Method \code{custom()}}{
Initialize a ``Rule`` object for a *custom* debugging rule.
             You can create a custom rule that analyzes tensors emitted
             during the training of a model
             and monitors conditions that are critical for the success of a training
             job. For more information, see `Create Debugger Custom Rules for Training Job
             Analysis
             <https://docs.aws.amazon.com/sagemaker/latest/dg/debugger-custom-rules.html>`_.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Rule$custom(
  name,
  image_uri,
  instance_type,
  volume_size_in_gb,
  source = NULL,
  rule_to_invoke = NULL,
  container_local_output_path = NULL,
  s3_output_path = NULL,
  other_trials_s3_input_paths = NULL,
  rule_parameters = NULL,
  collections_to_save = NULL,
  actions = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{(str): Required. The name of the debugger rule.}

\item{\code{image_uri}}{(str): Required. The URI of the image to be used by the debugger rule.}

\item{\code{instance_type}}{(str): Required. Type of EC2 instance to use, for example,
'ml.c4.xlarge'.}

\item{\code{volume_size_in_gb}}{(int): Required. Size in GB of the EBS volume
to use for storing data.}

\item{\code{source}}{(str): Optional. A source file containing a rule to invoke. If provided,
you must also provide rule_to_invoke. This can either be an S3 uri or
a local path.}

\item{\code{rule_to_invoke}}{(str): Optional. The name of the rule to invoke within the source.
If provided, you must also provide source.}

\item{\code{container_local_output_path}}{(str): Optional. The local path in the container.}

\item{\code{s3_output_path}}{(str): Optional. The location in Amazon S3 to store the output tensors.
The default Debugger output path for debugging data is created under the
default output path of the :class:`~sagemaker.estimator.Estimator` class.
For example,
s3://sagemaker-<region>-<12digit_account_id>/<training-job-name>/debug-output/.}

\item{\code{other_trials_s3_input_paths}}{([str]): Optional. The Amazon S3 input paths
of other trials to use the SimilarAcrossRuns rule.}

\item{\code{rule_parameters}}{(dict): Optional. A dictionary of parameters for the rule.}

\item{\code{collections_to_save}}{([sagemaker.debugger.CollectionConfig]): Optional. A list
of :class:`~sagemaker.debugger.CollectionConfig` objects to be saved.}

\item{\code{actions}}{:}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
:class:`~sagemaker.debugger.Rule`: The instance of the custom rule.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-prepare_actions"></a>}}
\if{latex}{\out{\hypertarget{method-prepare_actions}{}}}
\subsection{Method \code{prepare_actions()}}{
Prepare actions for Debugger Rule.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Rule$prepare_actions(training_job_name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{training_job_name}}{(str): The training job name. To be set as the default training job
prefix for the StopTraining action if it is specified.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-to_debugger_rule_config_dict"></a>}}
\if{latex}{\out{\hypertarget{method-to_debugger_rule_config_dict}{}}}
\subsection{Method \code{to_debugger_rule_config_dict()}}{
Generates a request dictionary using the parameters provided when initializing object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Rule$to_debugger_rule_config_dict()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
dict: An portion of an API request as a dictionary.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Rule$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
