% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3.R
\name{s3_path_join}
\alias{s3_path_join}
\title{Creates S3 uri paths}
\usage{
s3_path_join(...)
}
\arguments{
\item{...}{: The strings to join with a slash.}
}
\value{
charater: The joined string.


}
\description{
Returns the arguments joined by a slash ("/"), similarly to ``file.path()`` (on Unix).
             If the first argument is "s3://", then that is preserved.
}
