% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3.R
\name{S3Downloader}
\alias{S3Downloader}
\title{S3Downloader}
\description{
Contains static methods for downloading directories or files from S3.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-download}{\code{S3Downloader$download()}}
\item \href{#method-read_file}{\code{S3Downloader$read_file()}}
\item \href{#method-list_s3_uri}{\code{S3Downloader$list_s3_uri()}}
\item \href{#method-clone}{\code{S3Downloader$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-download"></a>}}
\if{latex}{\out{\hypertarget{method-download}{}}}
\subsection{Method \code{download()}}{
Static method that downloads a given S3 uri to the local machine.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{S3Downloader$download(
  s3_uri,
  local_path,
  kms_key = NULL,
  sagemaker_session = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{s3_uri}}{(str): An S3 uri to download from.}

\item{\code{local_path}}{(str): A local path to download the file(s) to.}

\item{\code{kms_key}}{(str): The KMS key to use to decrypt the files.}

\item{\code{sagemaker_session}}{(sagemaker.session.Session): Session object which
manages interactions with Amazon SageMaker APIs and any other
AWS services needed. If not specified, the estimator creates one
using the default AWS configuration chain.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-read_file"></a>}}
\if{latex}{\out{\hypertarget{method-read_file}{}}}
\subsection{Method \code{read_file()}}{
Static method that returns the contents of an s3 uri file body as a string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{S3Downloader$read_file(s3_uri, sagemaker_session = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{s3_uri}}{(str): An S3 uri that refers to a single file.}

\item{\code{sagemaker_session}}{(sagemaker.session.Session): AWS session to use. Automatically
generates one if not provided.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
str: The body of the file.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-list_s3_uri"></a>}}
\if{latex}{\out{\hypertarget{method-list_s3_uri}{}}}
\subsection{Method \code{list_s3_uri()}}{
Static method that lists the contents of an S3 uri.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{S3Downloader$list_s3_uri(s3_uri, sagemaker_session = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{s3_uri}}{(str): The S3 base uri to list objects in.}

\item{\code{sagemaker_session}}{(sagemaker.session.Session): AWS session to use. Automatically
generates one if not provided.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
[str]: The list of S3 URIs in the given S3 base uri.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{S3Downloader$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
