% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3.R
\name{S3Uploader}
\alias{S3Uploader}
\title{S3Uploader Class}
\description{
Contains static methods for uploading directories or files to S3
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-upload}{\code{S3Uploader$upload()}}
\item \href{#method-upload_string_as_file_body}{\code{S3Uploader$upload_string_as_file_body()}}
\item \href{#method-format}{\code{S3Uploader$format()}}
\item \href{#method-clone}{\code{S3Uploader$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-upload"></a>}}
\if{latex}{\out{\hypertarget{method-upload}{}}}
\subsection{Method \code{upload()}}{
Static method that uploads a given file or directory to S3.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{S3Uploader$upload(
  local_path = NULL,
  desired_s3_uri = NULL,
  kms_key = NULL,
  sagemaker_session = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{local_path}}{(str): Path (absolute or relative) of local file or directory to upload.}

\item{\code{desired_s3_uri}}{(str): The desired S3 location to upload to. It is the prefix to
which the local filename will be added.}

\item{\code{kms_key}}{(str): The KMS key to use to encrypt the files.}

\item{\code{sagemaker_session}}{(sagemaker.session.Session): Session object which
manages interactions with Amazon SageMaker APIs and any other
AWS services needed. If not specified, the estimator creates one
using the default AWS configuration chain.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The S3 uri of the uploaded file(s).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-upload_string_as_file_body"></a>}}
\if{latex}{\out{\hypertarget{method-upload_string_as_file_body}{}}}
\subsection{Method \code{upload_string_as_file_body()}}{
Static method that uploads a given file or directory to S3.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{S3Uploader$upload_string_as_file_body(
  body,
  desired_s3_uri = NULL,
  kms_key = NULL,
  sagemaker_session = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{body}}{(str): String representing the body of the file.}

\item{\code{desired_s3_uri}}{(str): The desired S3 uri to upload to.}

\item{\code{kms_key}}{(str): The KMS key to use to encrypt the files.}

\item{\code{sagemaker_session}}{(sagemaker.session.Session): AWS session to use. Automatically
generates one if not provided.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
str: The S3 uri of the uploaded file(s).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-format"></a>}}
\if{latex}{\out{\hypertarget{method-format}{}}}
\subsection{Method \code{format()}}{
format class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{S3Uploader$format()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{S3Uploader$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
