% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clarify.R
\name{SageMakerClarifyProcessor}
\alias{SageMakerClarifyProcessor}
\title{SageMakerClarifyProcessor Class}
\description{
Handles SageMaker Processing task to compute bias metrics and explain a model.
}
\section{Super class}{
\code{\link[R6sagemaker:Processor]{R6sagemaker::Processor}} -> \code{SageMakerClarifyProcessor}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{SageMakerClarifyProcessor$new()}}
\item \href{#method-run}{\code{SageMakerClarifyProcessor$run()}}
\item \href{#method-run_pre_training_bias}{\code{SageMakerClarifyProcessor$run_pre_training_bias()}}
\item \href{#method-run_post_training_bias}{\code{SageMakerClarifyProcessor$run_post_training_bias()}}
\item \href{#method-run_bias}{\code{SageMakerClarifyProcessor$run_bias()}}
\item \href{#method-run_explainability}{\code{SageMakerClarifyProcessor$run_explainability()}}
\item \href{#method-clone}{\code{SageMakerClarifyProcessor$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="R6sagemaker" data-topic="Processor" data-id="format">}\href{../../R6sagemaker/html/Processor.html#method-format}{\code{R6sagemaker::Processor$format()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initializes a ``Processor`` instance, computing bias metrics and model explanations.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SageMakerClarifyProcessor$new(
  role,
  instance_count,
  instance_type,
  volume_size_in_gb = 30,
  volume_kms_key = NULL,
  output_kms_key = NULL,
  max_runtime_in_seconds = NULL,
  sagemaker_session = NULL,
  env = NULL,
  tags = NULL,
  network_config = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{role}}{(str): An AWS IAM role name or ARN. Amazon SageMaker Processing
uses this role to access AWS resources, such as
data stored in Amazon S3.}

\item{\code{instance_count}}{(int): The number of instances to run
a processing job with.}

\item{\code{instance_type}}{(str): The type of EC2 instance to use for
processing, for example, 'ml.c4.xlarge'.}

\item{\code{volume_size_in_gb}}{(int): Size in GB of the EBS volume
to use for storing data during processing (default: 30).}

\item{\code{volume_kms_key}}{(str): A KMS key for the processing
volume (default: None).}

\item{\code{output_kms_key}}{(str): The KMS key ID for processing job outputs (default: None).}

\item{\code{max_runtime_in_seconds}}{(int): Timeout in seconds (default: None).
After this amount of time, Amazon SageMaker terminates the job,
regardless of its current status. If `max_runtime_in_seconds` is not
specified, the default value is 24 hours.}

\item{\code{sagemaker_session}}{(:class:`~sagemaker.session.Session`):
Session object which manages interactions with Amazon SageMaker and
any other AWS services needed. If not specified, the processor creates
one using the default AWS configuration chain.}

\item{\code{env}}{(dict[str, str]): Environment variables to be passed to
the processing jobs (default: None).}

\item{\code{tags}}{(list[dict]): List of tags to be passed to the processing job
(default: None). For more, see
https://docs.aws.amazon.com/sagemaker/latest/dg/API_Tag.html.}

\item{\code{network_config}}{(:class:`~sagemaker.network.NetworkConfig`):
A :class:`~sagemaker.network.NetworkConfig`
object that configures network isolation, encryption of
inter-container traffic, security group IDs, and subnets.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-run"></a>}}
\if{latex}{\out{\hypertarget{method-run}{}}}
\subsection{Method \code{run()}}{
Overriding the base class method but deferring to specific run_* methods.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SageMakerClarifyProcessor$run()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-run_pre_training_bias"></a>}}
\if{latex}{\out{\hypertarget{method-run_pre_training_bias}{}}}
\subsection{Method \code{run_pre_training_bias()}}{
Runs a ProcessingJob to compute the requested bias 'methods' of the input data.
             Computes the requested methods that compare 'methods' (e.g. fraction of examples) for the
             sensitive group vs the other examples.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SageMakerClarifyProcessor$run_pre_training_bias(
  data_config,
  data_bias_config,
  methods = "all",
  wait = TRUE,
  logs = TRUE,
  job_name = NULL,
  kms_key = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data_config}}{(:class:`~sagemaker.clarify.DataConfig`): Config of the input/output data.}

\item{\code{data_bias_config}}{(:class:`~sagemaker.clarify.BiasConfig`): Config of sensitive groups.}

\item{\code{methods}}{(str or list[str]): Selector of a subset of potential metrics:
["CI", "DPL", "KL", "JS", "LP", "TVD", "KS", "CDDL"]. Defaults to computing all.}

\item{\code{wait}}{(bool): Whether the call should wait until the job completes (default: True).}

\item{\code{logs}}{(bool): Whether to show the logs produced by the job.
Only meaningful when ``wait`` is True (default: True).}

\item{\code{job_name}}{(str): Processing job name. If not specified, a name is composed of
"Clarify-Pretraining-Bias" and current timestamp.}

\item{\code{kms_key}}{(str): The ARN of the KMS key that is used to encrypt the
user code file (default: None).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-run_post_training_bias"></a>}}
\if{latex}{\out{\hypertarget{method-run_post_training_bias}{}}}
\subsection{Method \code{run_post_training_bias()}}{
Runs a ProcessingJob to compute the requested bias 'methods' of the model predictions.
             Spins up a model endpoint, runs inference over the input example in the
             's3_data_input_path' to obtain predicted labels. Computes a the requested methods that
             compare 'methods' (e.g. accuracy, precision, recall) for the sensitive group vs the other
             examples.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SageMakerClarifyProcessor$run_post_training_bias(
  data_config,
  data_bias_config,
  model_config,
  model_predicted_label_config,
  methods = "all",
  wait = TRUE,
  logs = TRUE,
  job_name = NULL,
  kms_key = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data_config}}{(:class:`~sagemaker.clarify.DataConfig`): Config of the input/output data.}

\item{\code{data_bias_config}}{(:class:`~sagemaker.clarify.BiasConfig`): Config of sensitive groups.}

\item{\code{model_config}}{(:class:`~sagemaker.clarify.ModelConfig`): Config of the model and its
endpoint to be created.}

\item{\code{model_predicted_label_config}}{(:class:`~sagemaker.clarify.ModelPredictedLabelConfig`):
Config of how to extract the predicted label from the model output.}

\item{\code{methods}}{(str or list[str]): Selector of a subset of potential metrics:
["DPPL", "DI", "DCA", "DCR", "RD", "DAR", "DRR", "AD", "CDDPL", "TE", "FT"].
Defaults to computing all.}

\item{\code{wait}}{(bool): Whether the call should wait until the job completes (default: True).}

\item{\code{logs}}{(bool): Whether to show the logs produced by the job.
Only meaningful when ``wait`` is True (default: True).}

\item{\code{job_name}}{(str): Processing job name. If not specified, a name is composed of
"Clarify-Posttraining-Bias" and current timestamp.}

\item{\code{kms_key}}{(str): The ARN of the KMS key that is used to encrypt the
user code file (default: None).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-run_bias"></a>}}
\if{latex}{\out{\hypertarget{method-run_bias}{}}}
\subsection{Method \code{run_bias()}}{
Runs a ProcessingJob to compute the requested bias 'methods' of the model predictions.
             Spins up a model endpoint, runs inference over the input example in the
             's3_data_input_path' to obtain predicted labels. Computes a the requested methods that
             compare 'methods' (e.g. accuracy, precision, recall) for the sensitive group vs the other
             examples.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SageMakerClarifyProcessor$run_bias(
  data_config,
  bias_config,
  model_config,
  model_predicted_label_config = NULL,
  pre_training_methods = "all",
  post_training_methods = "all",
  wait = TRUE,
  logs = TRUE,
  job_name = NULL,
  kms_key = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data_config}}{(:class:`~sagemaker.clarify.DataConfig`): Config of the input/output data.}

\item{\code{bias_config}}{(:class:`~sagemaker.clarify.BiasConfig`): Config of sensitive groups.}

\item{\code{model_config}}{(:class:`~sagemaker.clarify.ModelConfig`): Config of the model and its
endpoint to be created.}

\item{\code{model_predicted_label_config}}{(:class:`~sagemaker.clarify.ModelPredictedLabelConfig`):
Config of how to extract the predicted label from the model output.}

\item{\code{pre_training_methods}}{(str or list[str]): Selector of a subset of potential metrics:
["DPPL", "DI", "DCA", "DCR", "RD", "DAR", "DRR", "AD", "CDDPL", "TE", "FT"].
Defaults to computing all.}

\item{\code{post_training_methods}}{(str or list[str]): Selector of a subset of potential metrics:
["DPPL", "DI", "DCA", "DCR", "RD", "DAR", "DRR", "AD", "CDDPL", "TE", "FT"].
Defaults to computing all.}

\item{\code{wait}}{(bool): Whether the call should wait until the job completes (default: True).}

\item{\code{logs}}{(bool): Whether to show the logs produced by the job.
Only meaningful when ``wait`` is True (default: True).}

\item{\code{job_name}}{(str): Processing job name. If not specified, a name is composed of
"Clarify-Bias" and current timestamp.}

\item{\code{kms_key}}{(str): The ARN of the KMS key that is used to encrypt the
user code file (default: None).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-run_explainability"></a>}}
\if{latex}{\out{\hypertarget{method-run_explainability}{}}}
\subsection{Method \code{run_explainability()}}{
Runs a ProcessingJob computing for each example in the input the feature importance.
             Currently, only SHAP is supported as explainability method.
             Spins up a model endpoint.
             For each input example in the 's3_data_input_path' the SHAP algorithm determines
             feature importance, by creating 'num_samples' copies of the example with a subset
             of features replaced with values from the 'baseline'.
             Model inference is run to see how the prediction changes with the replaced features.
             If the model output returns multiple scores importance is computed for each of them.
             Across examples, feature importance is aggregated using 'agg_method'.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SageMakerClarifyProcessor$run_explainability(
  data_config,
  model_config,
  explainability_config,
  model_scores = NULL,
  wait = TRUE,
  logs = TRUE,
  job_name = NULL,
  kms_key = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data_config}}{(:class:`~sagemaker.clarify.DataConfig`): Config of the input/output data.}

\item{\code{model_config}}{(:class:`~sagemaker.clarify.ModelConfig`): Config of the model and its
endpoint to be created.}

\item{\code{explainability_config}}{(:class:`~sagemaker.clarify.ExplainabilityConfig`): Config of the
specific explainability method. Currently, only SHAP is supported.}

\item{\code{model_scores}}{:  Index or JSONPath location in the model output for the predicted scores
to be explained. This is not required if the model output is a single score.}

\item{\code{wait}}{(bool): Whether the call should wait until the job completes (default: True).}

\item{\code{logs}}{(bool): Whether to show the logs produced by the job.
Only meaningful when ``wait`` is True (default: True).}

\item{\code{job_name}}{(str): Processing job name. If not specified, a name is composed of
"Clarify-Explainability" and current timestamp.}

\item{\code{kms_key}}{(str): The ARN of the KMS key that is used to encrypt the
user code file (default: None).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SageMakerClarifyProcessor$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
