% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session.R
\name{Session}
\alias{Session}
\title{Sagemaker Session Class}
\description{
Manage interactions with the Amazon SageMaker APIs and any other AWS services needed.
This class provides convenient methods for manipulating entities and resources that Amazon
SageMaker uses, such as training jobs, endpoints, and input datasets in S3.
AWS service calls are delegated to an underlying Boto3 session, which by default
is initialized using the AWS configuration chain. When you make an Amazon SageMaker API call
that accesses an S3 bucket location and one is not specified, the ``Session`` creates a default
bucket based on a naming convention which includes the current AWS account ID.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Session$new()}}
\item \href{#method-get_caller_identity_arn}{\code{Session$get_caller_identity_arn()}}
\item \href{#method-upload_data}{\code{Session$upload_data()}}
\item \href{#method-download_data}{\code{Session$download_data()}}
\item \href{#method-list_s3_files}{\code{Session$list_s3_files()}}
\item \href{#method-default_bucket}{\code{Session$default_bucket()}}
\item \href{#method-train}{\code{Session$train()}}
\item \href{#method-process}{\code{Session$process()}}
\item \href{#method-create_monitoring_schedule}{\code{Session$create_monitoring_schedule()}}
\item \href{#method-update_monitoring_schedule}{\code{Session$update_monitoring_schedule()}}
\item \href{#method-start_monitoring_schedule}{\code{Session$start_monitoring_schedule()}}
\item \href{#method-stop_monitoring_schedule}{\code{Session$stop_monitoring_schedule()}}
\item \href{#method-delete_monitoring_schedule}{\code{Session$delete_monitoring_schedule()}}
\item \href{#method-describe_monitoring_schedule}{\code{Session$describe_monitoring_schedule()}}
\item \href{#method-list_monitoring_executions}{\code{Session$list_monitoring_executions()}}
\item \href{#method-list_monitoring_schedules}{\code{Session$list_monitoring_schedules()}}
\item \href{#method-was_processing_job_successful}{\code{Session$was_processing_job_successful()}}
\item \href{#method-describe_processing_job}{\code{Session$describe_processing_job()}}
\item \href{#method-stop_processing_job}{\code{Session$stop_processing_job()}}
\item \href{#method-stop_training_job}{\code{Session$stop_training_job()}}
\item \href{#method-describe_training_job}{\code{Session$describe_training_job()}}
\item \href{#method-print}{\code{Session$print()}}
\item \href{#method-clone}{\code{Session$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Session$new(paws_credentials = NULL, bucket = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_caller_identity_arn"></a>}}
\if{latex}{\out{\hypertarget{method-get_caller_identity_arn}{}}}
\subsection{Method \code{get_caller_identity_arn()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Session$get_caller_identity_arn()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-upload_data"></a>}}
\if{latex}{\out{\hypertarget{method-upload_data}{}}}
\subsection{Method \code{upload_data()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Session$upload_data(path, bucket = NULL, key_prefix = "data", ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-download_data"></a>}}
\if{latex}{\out{\hypertarget{method-download_data}{}}}
\subsection{Method \code{download_data()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Session$download_data(path = "", bucket, key_prefix = NULL, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-list_s3_files"></a>}}
\if{latex}{\out{\hypertarget{method-list_s3_files}{}}}
\subsection{Method \code{list_s3_files()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Session$list_s3_files(bucket, key_prefix)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-default_bucket"></a>}}
\if{latex}{\out{\hypertarget{method-default_bucket}{}}}
\subsection{Method \code{default_bucket()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Session$default_bucket()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-train"></a>}}
\if{latex}{\out{\hypertarget{method-train}{}}}
\subsection{Method \code{train()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Session$train(
  input_mode,
  input_config,
  role,
  job_name,
  output_config = NULL,
  resource_config = NULL,
  vpc_config = NULL,
  hyperparameters = NULL,
  stop_condition = NULL,
  tags = NULL,
  metric_definitions = NULL,
  enable_network_isolation = FALSE,
  image = NULL,
  algorithm_arn = NULL,
  encrypt_inter_container_traffic = FALSE,
  train_use_spot_instances = FALSE,
  checkpoint_s3_uri = NULL,
  checkpoint_local_path = NULL,
  experiment_config = NULL,
  debugger_rule_configs = NULL,
  debugger_hook_config = NULL,
  tensorboard_output_config = NULL,
  enable_sagemaker_metrics = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-process"></a>}}
\if{latex}{\out{\hypertarget{method-process}{}}}
\subsection{Method \code{process()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Session$process(
  inputs = NULL,
  output_config = NULL,
  job_name = NULL,
  resources = NULL,
  stopping_condition = NULL,
  app_specification = NULL,
  environment = NULL,
  network_config = NULL,
  role_arn,
  tags = NULL,
  experiment_config = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-create_monitoring_schedule"></a>}}
\if{latex}{\out{\hypertarget{method-create_monitoring_schedule}{}}}
\subsection{Method \code{create_monitoring_schedule()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Session$create_monitoring_schedule(
  monitoring_schedule_name,
  schedule_expression = NULL,
  statistics_s3_uri = NULL,
  constraints_s3_uri = NULL,
  monitoring_inputs = NULL,
  monitoring_output_config = NULL,
  instance_count = 1,
  instance_type = c("ml.t3.medium", "ml.t3.large", "ml.t3.xlarge", "ml.t3.2xlarge",
    "ml.m4.xlarge", "ml.m4.2xlarge", "ml.m4.4xlarge", "ml.m4.10xlarge", "ml.m4.16xlarge",
    "ml.c4.xlarge", "ml.c4.2xlarge", "ml.c4.4xlarge", "ml.c4.8xlarge", "ml.p2.xlarge",
    "ml.p2.8xlarge", "ml.p2.16xlarge", "ml.p3.2xlarge", "ml.p3.8xlarge",
    "ml.p3.16xlarge", "ml.c5.xlarge", "ml.c5.2xlarge", "ml.c5.4xlarge", "ml.c5.9xlarge",
    "ml.c5.18xlarge", "ml.m5.large", "ml.m5.xlarge", "ml.m5.2xlarge", "ml.m5.4xlarge",
    "ml.m5.12xlarge", "ml.m5.24xlarge",      "ml.r5.large", "ml.r5.xlarge",
    "ml.r5.2xlarge", "ml.r5.4xlarge", "ml.r5.8xlarge", "ml.r5.12xlarge",
    "ml.r5.16xlarge", "ml.r5.24xlarge"),
  volume_size_in_gb = NULL,
  volume_kms_key = NULL,
  image_uri = NULL,
  entrypoint = NULL,
  arguments = NULL,
  record_preprocessor_source_uri = NULL,
  post_analytics_processor_source_uri = NULL,
  max_runtime_in_seconds = NULL,
  environment = NULL,
  network_config = NULL,
  role_arn = NULL,
  tags = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update_monitoring_schedule"></a>}}
\if{latex}{\out{\hypertarget{method-update_monitoring_schedule}{}}}
\subsection{Method \code{update_monitoring_schedule()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Session$update_monitoring_schedule()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-start_monitoring_schedule"></a>}}
\if{latex}{\out{\hypertarget{method-start_monitoring_schedule}{}}}
\subsection{Method \code{start_monitoring_schedule()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Session$start_monitoring_schedule(monitoring_schedule_name)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-stop_monitoring_schedule"></a>}}
\if{latex}{\out{\hypertarget{method-stop_monitoring_schedule}{}}}
\subsection{Method \code{stop_monitoring_schedule()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Session$stop_monitoring_schedule(monitoring_schedule_name)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-delete_monitoring_schedule"></a>}}
\if{latex}{\out{\hypertarget{method-delete_monitoring_schedule}{}}}
\subsection{Method \code{delete_monitoring_schedule()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Session$delete_monitoring_schedule(monitoring_schedule_name)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-describe_monitoring_schedule"></a>}}
\if{latex}{\out{\hypertarget{method-describe_monitoring_schedule}{}}}
\subsection{Method \code{describe_monitoring_schedule()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Session$describe_monitoring_schedule(monitoring_schedule_name)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-list_monitoring_executions"></a>}}
\if{latex}{\out{\hypertarget{method-list_monitoring_executions}{}}}
\subsection{Method \code{list_monitoring_executions()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Session$list_monitoring_executions(
  monitoring_schedule_name,
  sort_by = "ScheduledTime",
  sort_order = "Descending",
  max_results = 100
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-list_monitoring_schedules"></a>}}
\if{latex}{\out{\hypertarget{method-list_monitoring_schedules}{}}}
\subsection{Method \code{list_monitoring_schedules()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Session$list_monitoring_schedules(
  endpoint_name = NULL,
  sort_by = "CreationTime",
  sort_order = "Descending",
  max_results = 100
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-was_processing_job_successful"></a>}}
\if{latex}{\out{\hypertarget{method-was_processing_job_successful}{}}}
\subsection{Method \code{was_processing_job_successful()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Session$was_processing_job_successful(job_name)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-describe_processing_job"></a>}}
\if{latex}{\out{\hypertarget{method-describe_processing_job}{}}}
\subsection{Method \code{describe_processing_job()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Session$describe_processing_job(job_name)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-stop_processing_job"></a>}}
\if{latex}{\out{\hypertarget{method-stop_processing_job}{}}}
\subsection{Method \code{stop_processing_job()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Session$stop_processing_job(job_name)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-stop_training_job"></a>}}
\if{latex}{\out{\hypertarget{method-stop_training_job}{}}}
\subsection{Method \code{stop_training_job()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Session$stop_training_job(job_name)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-describe_training_job"></a>}}
\if{latex}{\out{\hypertarget{method-describe_training_job}{}}}
\subsection{Method \code{describe_training_job()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Session$describe_training_job(job_name)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Session$print(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Session$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
