% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clarify.R
\name{SHAPConfig}
\alias{SHAPConfig}
\title{SHAPConfig Class}
\description{
Config class of SHAP.
}
\section{Super class}{
\code{\link[R6sagemaker:ExplainabilityConfig]{R6sagemaker::ExplainabilityConfig}} -> \code{SHAPConfig}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{shap_config}}{Shap Config}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{SHAPConfig$new()}}
\item \href{#method-get_explainability_config}{\code{SHAPConfig$get_explainability_config()}}
\item \href{#method-clone}{\code{SHAPConfig$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="R6sagemaker" data-topic="ExplainabilityConfig" data-id="print">}\href{../../R6sagemaker/html/ExplainabilityConfig.html#method-print}{\code{R6sagemaker::ExplainabilityConfig$print()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initializes config for SHAP.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SHAPConfig$new(
  baseline,
  num_samples,
  agg_method = c("mean_abs", "median", "mean_sq"),
  use_logit = FALSE,
  save_local_shap_values = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{baseline}}{(str or list): A list of rows (at least one) or S3 object URI to be used as
the baseline dataset in the Kernel SHAP algorithm. The format should be the same
as the dataset format. Each row should contain only the feature columns/values
and omit the label column/values.}

\item{\code{num_samples}}{(int): Number of samples to be used in the Kernel SHAP algorithm.
This number determines the size of the generated synthetic dataset to compute the
SHAP values.}

\item{\code{agg_method}}{(str): Aggregation method for global SHAP values. Valid values are
"mean_abs" (mean of absolute SHAP values for all instances),
"median" (median of SHAP values for all instances) and
"mean_sq" (mean of squared SHAP values for all instances).}

\item{\code{use_logit}}{(bool): Indicator of whether the logit function is to be applied to the model
predictions. Default is False. If "use_logit" is true then the SHAP values will
have log-odds units.}

\item{\code{save_local_shap_values}}{(bool): Indicator of whether to save the local SHAP values
in the output location. Default is True.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_explainability_config"></a>}}
\if{latex}{\out{\hypertarget{method-get_explainability_config}{}}}
\subsection{Method \code{get_explainability_config()}}{
Returns config.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SHAPConfig$get_explainability_config()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SHAPConfig$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
