% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.R
\name{ShuffleConfig}
\alias{ShuffleConfig}
\title{ShuffleConfig Class}
\description{
For configuring channel shuffling using a seed
             For more detail, see the AWS documentation:
             https://docs.aws.amazon.com/sagemaker/latest/dg/API_ShuffleConfig.html
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{seed}}{value used to seed the shuffled sequence.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ShuffleConfig$new()}}
\item \href{#method-format}{\code{ShuffleConfig$format()}}
\item \href{#method-clone}{\code{ShuffleConfig$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a ShuffleConfig.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ShuffleConfig$new(seed)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{seed}}{(numeric): value used to seed the shuffled sequence.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-format"></a>}}
\if{latex}{\out{\hypertarget{method-format}{}}}
\subsection{Method \code{format()}}{
format class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ShuffleConfig$format()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ShuffleConfig$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
