% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sm.R
\docType{package}
\name{sm-package}
\alias{sm}
\alias{sm-package}
\title{sm: this is just a placeholder}
\description{
More about what it does (maybe more than one line)
    Use four spaces when indenting paragraphs within the Description.
}
