% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_monitor_monitoring_files.R
\name{Statistics}
\alias{Statistics}
\title{Statistics Class}
\description{
Represents the statistics JSON file used in Amazon SageMaker Model Monitoring.
}
\section{Super class}{
\code{\link[R6sagemaker:ModelMonitoringFile]{R6sagemaker::ModelMonitoringFile}} -> \code{Statistics}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Statistics$new()}}
\item \href{#method-from_s3_uri}{\code{Statistics$from_s3_uri()}}
\item \href{#method-from_string}{\code{Statistics$from_string()}}
\item \href{#method-from_file_path}{\code{Statistics$from_file_path()}}
\item \href{#method-clone}{\code{Statistics$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="R6sagemaker" data-topic="ModelMonitoringFile" data-id="print">}\href{../../R6sagemaker/html/ModelMonitoringFile.html#method-print}{\code{R6sagemaker::ModelMonitoringFile$print()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="R6sagemaker" data-topic="ModelMonitoringFile" data-id="save">}\href{../../R6sagemaker/html/ModelMonitoringFile.html#method-save}{\code{R6sagemaker::ModelMonitoringFile$save()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initializes the Statistics object used in Amazon SageMaker Model Monitoring.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Statistics$new(
  body_dict = NULL,
  statistics_file_s3_uri = NULL,
  kms_key = NULL,
  sagemaker_session = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{body_dict}}{(str): The body of the statistics JSON file.}

\item{\code{statistics_file_s3_uri}}{(str): The uri of the statistics JSON file.}

\item{\code{kms_key}}{(str): The kms key to be used to decrypt the file in S3.}

\item{\code{sagemaker_session}}{(sagemaker.session.Session): A SageMaker Session
object, used for SageMaker interactions (default: None). If not
specified, one is created using the default AWS configuration
chain.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-from_s3_uri"></a>}}
\if{latex}{\out{\hypertarget{method-from_s3_uri}{}}}
\subsection{Method \code{from_s3_uri()}}{
Generates a Statistics object from an s3 uri.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Statistics$from_s3_uri(
  statistics_file_s3_uri,
  kms_key = NULL,
  sagemaker_session = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{statistics_file_s3_uri}}{(str): The uri of the statistics JSON file.}

\item{\code{kms_key}}{(str): The kms key to be used to decrypt the file in S3.}

\item{\code{sagemaker_session}}{(sagemaker.session.Session): A SageMaker Session
object, used for SageMaker interactions (default: None). If not
specified, one is created using the default AWS configuration
chain.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
sagemaker.model_monitor.Statistics: The instance of Statistics generated from
             the s3 uri.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-from_string"></a>}}
\if{latex}{\out{\hypertarget{method-from_string}{}}}
\subsection{Method \code{from_string()}}{
Generates a Statistics object from an s3 uri.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Statistics$from_string(
  statistics_file_string,
  kms_key = NULL,
  file_name = NULL,
  sagemaker_session = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{statistics_file_string}}{(str): The uri of the statistics JSON file.}

\item{\code{kms_key}}{(str): The kms key to be used to encrypt the file in S3.}

\item{\code{file_name}}{(str): The file name to use when uploading to S3.}

\item{\code{sagemaker_session}}{(sagemaker.session.Session): A SageMaker Session
object, used for SageMaker interactions (default: None). If not
specified, one is created using the default AWS configuration
chain.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
sagemaker.model_monitor.Statistics: The instance of Statistics generated from
             the s3 uri.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-from_file_path"></a>}}
\if{latex}{\out{\hypertarget{method-from_file_path}{}}}
\subsection{Method \code{from_file_path()}}{
Initializes a Statistics object from a file path.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Statistics$from_file_path(
  statistics_file_path,
  kms_key = NULL,
  sagemaker_session = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{statistics_file_path}}{(str): The path to the statistics file.}

\item{\code{kms_key}}{(str): The kms_key to use when encrypting the file in S3.}

\item{\code{sagemaker_session}}{(sagemaker.session.Session): A SageMaker Session
object, used for SageMaker interactions (default: None). If not
specified, one is created using the default AWS configuration
chain.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
sagemaker.model_monitor.Statistics: The instance of Statistics generated from
             the local file path.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Statistics$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
