% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debugger_metrics_config.R
\name{StepRange}
\alias{StepRange}
\title{Configuration for the range of steps to profile.}
\description{
It returns the target steps in dictionary format that you can pass to the
             :class:`~sagemaker.debugger.FrameworkProfile` class.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{StepRange$new()}}
\item \href{#method-to_json}{\code{StepRange$to_json()}}
\item \href{#method-format}{\code{StepRange$format()}}
\item \href{#method-clone}{\code{StepRange$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Set the start step and num steps.
             If the start step is not specified,
             Debugger starts profiling
             at step 0. If num steps is not specified, profile for 1 step.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StepRange$new(start_step, num_steps)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{start_step}}{(int): The step to start profiling.}

\item{\code{num_steps}}{(int): The number of steps to profile.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-to_json"></a>}}
\if{latex}{\out{\hypertarget{method-to_json}{}}}
\subsection{Method \code{to_json()}}{
Convert the step range into a dictionary.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StepRange$to_json()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
list: The step range as a dictionary.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-format"></a>}}
\if{latex}{\out{\hypertarget{method-format}{}}}
\subsection{Method \code{format()}}{
format class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StepRange$format()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StepRange$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
