% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictor.R
\docType{data}
\name{string_deserializer}
\alias{string_deserializer}
\title{S3 method to call StringDeserializer class}
\format{
An object of class \code{StringDeserializer} (inherits from \code{Deserializer}, \code{R6}) of length 7.
}
\usage{
string_deserializer
}
\description{
S3 method to call StringDeserializer class
}
\keyword{datasets}
