% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debugger.R
\name{TensorBoardOutputConfig}
\alias{TensorBoardOutputConfig}
\title{TensorBoardOutputConfig Class}
\description{
Create a tensor ouput configuration object for debugging visualizations on TensorBoard
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{s3_output_path}}{The location in Amazon S3 to store the output.}

\item{\code{container_local_output_path}}{The local path in the container.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{TensorBoardOutputConfig$new()}}
\item \href{#method-to_request_list}{\code{TensorBoardOutputConfig$to_request_list()}}
\item \href{#method-format}{\code{TensorBoardOutputConfig$format()}}
\item \href{#method-clone}{\code{TensorBoardOutputConfig$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialize the TensorBoardOutputConfig instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TensorBoardOutputConfig$new(s3_output_path, container_local_output_path = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{s3_output_path}}{(str): Optional. The location in Amazon S3 to store the output.}

\item{\code{container_local_output_path}}{(str): Optional. The local path in the container.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-to_request_list"></a>}}
\if{latex}{\out{\hypertarget{method-to_request_list}{}}}
\subsection{Method \code{to_request_list()}}{
Generate a request dictionary using the instances attributes.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TensorBoardOutputConfig$to_request_list()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
dict: An portion of an API request as a dictionary.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-format"></a>}}
\if{latex}{\out{\hypertarget{method-format}{}}}
\subsection{Method \code{format()}}{
format class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TensorBoardOutputConfig$format()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TensorBoardOutputConfig$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
