% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debugger_metrics_config.R
\name{TimeRange}
\alias{TimeRange}
\title{Configuration for the range of Unix time to profile.}
\description{
It returns the target time duration in dictionary format that you can pass to the
             :class:`~sagemaker.debugger.FrameworkProfile` class.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{TimeRange$new()}}
\item \href{#method-to_json}{\code{TimeRange$to_json()}}
\item \href{#method-format}{\code{TimeRange$format()}}
\item \href{#method-clone}{\code{TimeRange$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Set the start Unix time and duration.
             If the start Unix time is not specified,
             profile starting at step 0. If the duration is not specified, profile for 1 step.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TimeRange$new(start_unix_time = NULL, duration = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{start_unix_time}}{(int): The Unix time to start profiling.}

\item{\code{duration}}{(float): The duration in seconds to profile.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-to_json"></a>}}
\if{latex}{\out{\hypertarget{method-to_json}{}}}
\subsection{Method \code{to_json()}}{
Convert the time range into a dictionary.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TimeRange$to_json()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
dict: The time range as a dictionary.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-format"></a>}}
\if{latex}{\out{\hypertarget{method-format}{}}}
\subsection{Method \code{format()}}{
format class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TimeRange$format()}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{(ignored).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TimeRange$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
