% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics.R
\name{TrainingJobAnalytics}
\alias{TrainingJobAnalytics}
\title{TrainingJobAnalytics Class}
\description{
Fetch training curve data from CloudWatch Metrics for a specific training
             job.
}
\section{Super class}{
\code{\link[R6sagemaker:AnalyticsMetricsBase]{R6sagemaker::AnalyticsMetricsBase}} -> \code{TrainingJobAnalytics}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{CLOUDWATCH_NAMESPACE}}{CloudWatch namespace to return Training Job Analytics data}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{name}}{Name of the TrainingJob being analyzed}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{TrainingJobAnalytics$new()}}
\item \href{#method-clear_cache}{\code{TrainingJobAnalytics$clear_cache()}}
\item \href{#method-clone}{\code{TrainingJobAnalytics$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="R6sagemaker" data-topic="AnalyticsMetricsBase" data-id="dataframe">}\href{../../R6sagemaker/html/AnalyticsMetricsBase.html#method-dataframe}{\code{R6sagemaker::AnalyticsMetricsBase$dataframe()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="R6sagemaker" data-topic="AnalyticsMetricsBase" data-id="export_csv">}\href{../../R6sagemaker/html/AnalyticsMetricsBase.html#method-export_csv}{\code{R6sagemaker::AnalyticsMetricsBase$export_csv()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="R6sagemaker" data-topic="AnalyticsMetricsBase" data-id="format">}\href{../../R6sagemaker/html/AnalyticsMetricsBase.html#method-format}{\code{R6sagemaker::AnalyticsMetricsBase$format()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialize a ``TrainingJobAnalytics`` instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TrainingJobAnalytics$new(
  training_job_name,
  metric_names = NULL,
  sagemaker_session = NULL,
  start_time = NULL,
  end_time = NULL,
  period = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{training_job_name}}{(str): name of the TrainingJob to analyze.}

\item{\code{metric_names}}{(list, optional): string names of all the metrics to
collect for this training job. If not specified, then it will
use all metric names configured for this job.}

\item{\code{sagemaker_session}}{(sagemaker.session.Session): Session object which
manages interactions with Amazon SageMaker APIs and any other
AWS services needed. If not specified, one is specified using
the default AWS configuration chain.}

\item{\code{start_time}}{:}

\item{\code{end_time}}{:}

\item{\code{period}}{:}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clear_cache"></a>}}
\if{latex}{\out{\hypertarget{method-clear_cache}{}}}
\subsection{Method \code{clear_cache()}}{
Clear the object of all local caches of API methods, so that the next
             time any properties are accessed they will be refreshed from the
             service.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TrainingJobAnalytics$clear_cache()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TrainingJobAnalytics$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
