% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuner.R
\name{WarmStartConfig}
\alias{WarmStartConfig}
\title{WarmStartConfig Class}
\description{
Warm Start Configuration which defines the nature of the warm start
             ``HyperparameterTuner``, with type and parents for warm start.
             Examples:
             >>> warm_start_config = WarmStartConfig(
             >>>           type=WarmStartTypes.TransferLearning, parents={"p1","p2"})
             >>> warm_start_config.type
             "TransferLearning"
             >>> warm_start_config.parents
             {"p1","p2"}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{type}}{Supported warm start types}

\item{\code{parents}}{Set of parent tuning jobs}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{WarmStartConfig$new()}}
\item \href{#method-from_job_desc}{\code{WarmStartConfig$from_job_desc()}}
\item \href{#method-to_input_req}{\code{WarmStartConfig$to_input_req()}}
\item \href{#method-format}{\code{WarmStartConfig$format()}}
\item \href{#method-clone}{\code{WarmStartConfig$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initializes the ``WarmStartConfig`` with the provided
             ``WarmStartTypes`` and parents.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WarmStartConfig$new(
  warm_start_type = c("IdenticalDataAndAlgorithm", "TransferLearning"),
  parents = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{warm_start_type}}{(str): This should be one
of the supported warm start types ```("IdenticalDataAndAlgorithm", "TransferLearning")```}

\item{\code{parents}}{(str/list): Set of parent tuning jobs which will be used to
warm start the new tuning job.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-from_job_desc"></a>}}
\if{latex}{\out{\hypertarget{method-from_job_desc}{}}}
\subsection{Method \code{from_job_desc()}}{
Creates an instance of ``WarmStartConfig`` class, from warm start
             configuration response from DescribeTrainingJob.
             Examples:
             >>> warm_start_config = WarmStartConfig$new()$from_job_desc(warm_start_config=list(
             >>>    "WarmStartType"="TransferLearning",
             >>>    "ParentHyperParameterTuningJobs"= list(
             >>>        list('HyperParameterTuningJobName'= "p1"),
             >>>        list('HyperParameterTuningJobName'= "p2")
             >>>    )
             >>>))
             >>> warm_start_config.type
             "TransferLearning"
             >>> warm_start_config.parents
             ["p1","p2"]
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WarmStartConfig$from_job_desc(warm_start_config)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{warm_start_config}}{(dict): The expected format of the
``warm_start_config`` contains two first-class}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
sagemaker.tuner.WarmStartConfig: De-serialized instance of
             WarmStartConfig containing the type and parents provided as part of
             ``warm_start_config``.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-to_input_req"></a>}}
\if{latex}{\out{\hypertarget{method-to_input_req}{}}}
\subsection{Method \code{to_input_req()}}{
Converts the ``self`` instance to the desired input request format.
             Examples:
             >>> warm_start_config = WarmStartConfig$new("TransferLearning",parents="p1,p2")
             >>> warm_start_config$to_input_req()
             list(
                "WarmStartType"="TransferLearning",
                "ParentHyperParameterTuningJobs"= list(
                     list('HyperParameterTuningJobName': "p1"),
                     list('HyperParameterTuningJobName': "p2")
                  )
             )
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WarmStartConfig$to_input_req()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
list: Containing the "WarmStartType" and
             "ParentHyperParameterTuningJobs" as the first class fields.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-format"></a>}}
\if{latex}{\out{\hypertarget{method-format}{}}}
\subsection{Method \code{format()}}{
format class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WarmStartConfig$format()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WarmStartConfig$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
