% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictor.R
\docType{data}
\name{csv_deserializer}
\alias{csv_deserializer}
\title{S3 method to call CsvDeserializer class}
\format{
An object of class \code{CsvDeserializer} (inherits from \code{Deserializer}, \code{R6}) of length 6.
}
\usage{
csv_deserializer
}
\description{
S3 method to call CsvDeserializer class
}
\keyword{datasets}
