% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serializers.R
\name{CSVSerializer}
\alias{CSVSerializer}
\title{CSVSerializer Class}
\description{
Make Raw data using text/csv format
}
\section{Super classes}{
\code{\link[R6sagemaker:BaseSerializer]{R6sagemaker::BaseSerializer}} -> \code{\link[R6sagemaker:SimpleBaseSerializer]{R6sagemaker::SimpleBaseSerializer}} -> \code{CSVSerializer}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{CSVSerializer$new()}}
\item \href{#method-serialize}{\code{CSVSerializer$serialize()}}
\item \href{#method-clone}{\code{CSVSerializer$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="R6sagemaker" data-topic="BaseSerializer" data-id="format">}\href{../../R6sagemaker/html/BaseSerializer.html#method-format}{\code{R6sagemaker::BaseSerializer$format()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialize a ``CSVSerializer`` instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CSVSerializer$new(content_type = "text/csv")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{content_type}}{(str): The MIME type to signal to the inference endpoint when sending
request data (default: "text/csv").}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-serialize"></a>}}
\if{latex}{\out{\hypertarget{method-serialize}{}}}
\subsection{Method \code{serialize()}}{
Take data of various data formats and serialize them into CSV.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CSVSerializer$serialize(data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{(object): Data to be serialized. Any list of same length vectors; e.g. data.frame and data.table.
If matrix, it gets internally coerced to data.table preserving col names but not row names}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CSVSerializer$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
