% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deserializers.R
\docType{data}
\name{json_deserializer}
\alias{json_deserializer}
\title{S3 method to call StringBaseDeserializer class}
\format{
An object of class \code{JsonBaseDeserializer} (inherits from \code{BaseDeserializer}, \code{R6}) of length 7.
}
\usage{
json_deserializer
}
\description{
S3 method to call StringBaseDeserializer class
}
\keyword{datasets}
