% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serializers.R
\docType{data}
\name{json_serializer}
\alias{json_serializer}
\title{S3 method to call JsonSerializer class}
\format{
An object of class \code{JSONSerializer} (inherits from \code{SimpleBaseSerializer}, \code{BaseSerializer}, \code{R6}) of length 7.
}
\usage{
json_serializer
}
\description{
S3 method to call JsonSerializer class
}
\keyword{datasets}
