% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r6sagemaker.R
\docType{package}
\name{R6sagemaker-package}
\alias{R6sagemaker}
\alias{R6sagemaker-package}
\title{r6 sagemaker: this is just a placeholder}
\description{
More about what it does (maybe more than one line)
    Use four spaces when indenting paragraphs within the Description.
}
