% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r6sagemaker.R
\docType{package}
\name{R6sagemaker-package}
\alias{R6sagemaker}
\alias{R6sagemaker-package}
\title{r6 sagemaker: this is just a placeholder}
\description{
A library for training and deploying machine learning models on 
    Amazon SageMaker <https://aws.amazon.com/sagemaker/> using R through `paws SDK`.
}
\author{
\strong{Maintainer}: Dyfan Jones \email{dyfan.r.jones@gmail.com}

Other contributors:
\itemize{
  \item Amazon.com, Inc. [copyright holder]
}

}
