% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deserializers.R
\docType{data}
\name{string_deserializer}
\alias{string_deserializer}
\title{S3 method to call StringBaseDeserializer class}
\format{
An object of class \code{StringBaseDeserializer} (inherits from \code{BaseDeserializer}, \code{R6}) of length 7.
}
\usage{
string_deserializer
}
\description{
S3 method to call StringBaseDeserializer class
}
\keyword{datasets}
